fix: repair UI styles in auth inputs
