--- server/public/assets/app-auth-BOk5gCNS.js
@@ -1,9 +1,0 @@
-var q=Object.defineProperty,K=Object.defineProperties;var Q=Object.getOwnPropertyDescriptors;var B=Object.getOwnPropertySymbols;var X=Object.prototype.hasOwnProperty,Y=Object.prototype.propertyIsEnumerable;var T=(s,t,a)=>t in s?q(s,t,{enumerable:!0,configurable:!0,writable:!0,value:a}):s[t]=a,w=(s,t)=>{for(var a in t||(t={}))X.call(t,a)&&T(s,a,t[a]);if(B)for(var a of B(t))Y.call(t,a)&&T(s,a,t[a]);return s},k=(s,t)=>K(s,Q(t));var p=(s,t,a)=>new Promise((o,r)=>{var i=c=>{try{d(a.next(c))}catch(m){r(m)}},l=c=>{try{d(a.throw(c))}catch(m){r(m)}},d=c=>c.done?o(c.value):Promise.resolve(c.value).then(i,l);d((a=a.apply(s,t)).next())});import{r as x,u as b,j as e,G as ee,a as A,F as C,b as se,c as te,d as re,L as S,k as U,e as ae,f as Z,H as oe,g as V,h as _,i as ne}from"./vendor-react-DZE_uCEP.js";import{F as E,a as N,b as W,c as I,P,A as ie,S as le,E as ce,B as de}from"./app-components-DCNmHqgN.js";import{m as f,A as me}from"./vendor-framer-motion-D6A4MpK3.js";import{z as j,t as L}from"./vendor-other-DT6kVT6s.js";import{h as z}from"./app-services-MwNJNg2u.js";const J="https://codelinesjs.pl/api/",ue=()=>{const[s,t]=x.useState(!1),[a,o]=x.useState(null),[r,i]=x.useState(!1),[l,d]=x.useState(!0),[c,m]=x.useState(null);return{loading:s,setLoading:t,error:a,setError:o,isAuthenticated:r,setIsAuthenticated:i,isLoading:l,setIsLoading:d,user:c,setUser:m}},xe=s=>{const t=b(),{setLoading:a,setError:o,setIsAuthenticated:r,setUser:i}=s;return(d,c,m=!1)=>p(void 0,null,function*(){try{a(!0),o(null);const n=yield z.post("auth/login",{email:d,password:c,rememberMe:m});if(n.error)throw new Error(n.error);if(!n.data)throw new Error("Nieznany błąd logowania. Spróbuj ponownie później.");const{user:u}=n.data;i(u),r(!0),t("/dashboard")}catch(n){const u=n instanceof Error?n.message:"Wystąpił błąd podczas logowania";o(u),r(!1)}finally{a(!1)}})},ge=s=>{const t=b(),{setLoading:a,setError:o,setIsAuthenticated:r,setUser:i}=s;return(d,c,m)=>p(void 0,null,function*(){try{a(!0),o(null);const n=yield z.post("auth/register",{email:d,password:c,username:m});if(n.error)throw n.error.includes("już istnieje")?(o("Użytkownik o podanym adresie email lub nazwie użytkownika już istnieje w systemie"),new Error("Użytkownik o podanym adresie email lub nazwie użytkownika już istnieje w systemie")):(o(n.error),new Error(n.error));if(!n.data){const g="Nieznany błąd rejestracji. Spróbuj ponownie później.";throw o(g),new Error(g)}const{user:u}=n.data;i&&u&&i(u),r(!0),t("/dashboard")}catch(n){const u=n instanceof Error?n.message:"Wystąpił błąd podczas rejestracji";o(u),r(!1)}finally{a(!1)}})},we=s=>{const{setLoading:t,setError:a}=s;return r=>p(void 0,null,function*(){var i;try{t(!0),a(null);const l=yield z.post("auth/forgot-password",{email:r});if(l.error)throw new Error(l.error);return((i=l.data)==null?void 0:i.message)||"Link do resetowania hasła został wysłany na Twój adres email."}catch(l){if(l instanceof Error)throw a(l.message),l;{const d="Wystąpił błąd podczas wysyłania linku resetującego hasło";throw a(d),new Error(d)}}finally{t(!1)}})},he=s=>{const t=b(),{setLoading:a,setIsAuthenticated:o,setUser:r}=s;return()=>p(void 0,null,function*(){try{a(!0),yield z.post("auth/logout",{}),r&&r(null),o(!1),t("/")}catch(l){console.error("Błąd podczas wylogowania:",l),r&&r(null),o(!1),t("/")}finally{a(!1)}})},je=s=>{const t=b(),{setLoading:a,setError:o,setIsAuthenticated:r,setUser:i}=s;return(d,c=!1)=>p(void 0,null,function*(){try{if(a(!0),o(null),!d.credential){o("Nie udało się uzyskać tokenu uwierzytelniającego z Google. Spróbuj ponownie."),r(!1);return}const m=yield z.post("auth/google-login",{credential:d.credential,rememberMe:c});if(m.error)throw new Error(m.error);if(!m.data)throw new Error("Nieznany błąd logowania przez Google. Spróbuj ponownie później.");const n=m.data,u=n.user||n.user&&n.user.user;i&&i(u),r(!0),t("/dashboard")}catch(m){const n=m instanceof Error?m.message:"Wystąpił błąd podczas logowania przez Google";o(n),r(!1)}finally{a(!1)}})},pe=s=>{const t=b(),{setLoading:a,setError:o}=s;return(i,l,d)=>p(void 0,null,function*(){var c;try{if(!i||i.length<10)throw new Error("Nieprawidłowy token resetowania hasła. Sprawdź, czy link jest poprawny.");if(l!==d)throw new Error("Hasła nie są identyczne. Upewnij się, że oba pola zawierają to samo hasło.");a(!0),o(null);const m=yield z.post("auth/reset-password",{token:i,password:l,confirmPassword:d});if(m.error)throw new Error(m.error);return setTimeout(()=>{t("/logowanie")},3e3),((c=m.data)==null?void 0:c.message)||"Hasło zostało pomyślnie zmienione. Za chwilę zostaniesz przekierowany do strony logowania."}catch(m){if(m instanceof Error)throw o(m.message),m;{const n="Wystąpił nieznany błąd podczas resetowania hasła";throw o(n),new Error(n)}}finally{a(!1)}})},ye=s=>{const t=xe(s),a=ge(s),o=we(s),r=he(s),i=je(s),l=pe(s);return{login:t,register:a,forgotPassword:o,logout:r,loginWithGoogle:i,resetPassword:l}},v=()=>{const s=ue(),t=ye(s),a=b(),[o,r]=x.useState(!0),i=()=>p(void 0,null,function*(){try{r(!0);const c=J.replace("www.",""),m=yield fetch(`${c}auth/verify`,{method:"GET",headers:{Accept:"application/json"},mode:"cors",credentials:"include"});if(m.status===429)throw new Error("Zbyt wiele prób weryfikacji. Spróbuj ponownie za chwilę.");let n,u="";try{if(u=yield m.text(),!u)throw new Error("Pusta odpowiedź serwera");n=JSON.parse(u)}catch(g){throw new Error("Nieprawidłowa odpowiedź serwera: "+g)}if((n==null?void 0:n.status)==="success"&&(n!=null&&n.data)){s.setUser(n.data.user||n.data),s.setIsAuthenticated(!0);return}if(!m.ok)throw new Error((n==null?void 0:n.error)||(n==null?void 0:n.message)||"Nieznany błąd weryfikacji tokenu");if(!(n!=null&&n.user))throw new Error("Odpowiedź API nie zawiera danych użytkownika");s.setUser(n.user),s.setIsAuthenticated(!0)}catch(c){s.setError(c instanceof Error?c.message:"Wystąpił błąd podczas weryfikacji tokenu"),s.setIsAuthenticated(!1)}finally{r(!1)}});x.useEffect(()=>{i()},[]);const l=()=>p(void 0,null,function*(){try{const c=J.replace("www.","");yield fetch(`${c}auth/logout`,{method:"POST",credentials:"include"})}catch(c){console.error("Błąd podczas wylogowania:",c)}finally{s.setIsAuthenticated(!1),s.setUser(null),a("/")}}),d=k(w({},s),{isAuthChecking:o,token:null});return k(w(w({},d),t),{logout:l})},M=({text:s,description:t,icon:a="⚡"})=>e.jsxs("div",{className:"flex items-start justify-center lg:justify-start gap-3 sm:gap-4",children:[e.jsx("div",{className:"w-8 h-8 sm:w-10 sm:h-10 bg-[#f7df1e] rounded-lg flex items-center justify-center flex-shrink-0 mt-1",children:e.jsx("span",{className:"text-black text-sm sm:text-base font-bold",children:a})}),e.jsxs("div",{className:"flex flex-col text-center lg:text-left",children:[e.jsx("span",{className:"text-white font-semibold text-base sm:text-lg",children:s}),t&&e.jsx("span",{className:"text-gray-300 text-sm sm:text-base mt-1",children:t})]})]}),fe=()=>e.jsxs("div",{className:"space-y-4 sm:space-y-6",children:[e.jsx(M,{text:"Szybki rozwój",description:"Naucz się nowoczesnych technologii JS",icon:"⚡"}),e.jsx(M,{text:"Społeczność",description:"Połącz się z innymi developerami",icon:"☕"}),e.jsx(M,{text:"Praktyka",description:"Rozwiązuj prawdziwe problemy",icon:">"})]}),be=()=>e.jsxs("div",{className:"w-full max-w-xl",children:[e.jsxs("div",{className:"flex items-center justify-center lg:justify-start gap-3 mb-6 sm:mb-8",children:[e.jsx("div",{className:"w-12 h-12 sm:w-16 sm:h-16 bg-[#f7df1e] rounded-xl flex items-center justify-center",children:e.jsx("span",{className:"text-black font-bold text-xl sm:text-2xl",children:"JS"})}),e.jsx("h2",{className:"text-2xl sm:text-3xl font-bold text-white",children:"CodeLines"})]}),e.jsxs("h1",{className:"text-2xl sm:text-3xl lg:text-4xl xl:text-5xl font-bold text-white mb-4 sm:mb-6 leading-tight",children:["Witaj w świecie ",e.jsx("span",{className:"text-[#f7df1e]",children:"JavaScript"})]}),e.jsx("p",{className:"text-base sm:text-lg text-gray-300 mb-6 sm:mb-8 leading-relaxed",children:"Dołącz do społeczności programistów i rozwijaj swoje umiejętności w najpopularniejszym języku programowania na świecie."}),e.jsx(fe,{}),e.jsx("div",{className:"mt-8 sm:mt-12 bg-gradient-to-br from-gray-800 to-gray-900 rounded-xl p-4 sm:p-6 border border-gray-700",children:e.jsx("pre",{className:"text-[#f7df1e] text-xs sm:text-sm font-mono leading-relaxed",children:e.jsx("code",{children:`const developer = {
-  name: "Twoje imię",
-  skills: ["JavaScript", "React", "Node.js"],
-  passion: "coding"
-};`})})})]}),ze=()=>e.jsx(f.div,{initial:{opacity:0,y:20},animate:{opacity:1,y:0},transition:{duration:.5,delay:.2},className:"w-full max-w-xl flex flex-col items-center lg:items-start text-center lg:text-left px-4 sm:px-0",children:e.jsx(be,{})}),ke=({activeTab:s,onTabChange:t})=>e.jsxs("div",{className:"flex items-center justify-center gap-1 mb-6 sm:mb-8 bg-gray-50 rounded-lg p-1",children:[e.jsx(F,{isActive:s==="login",onClick:()=>t("login"),label:"Logowanie"}),e.jsx(F,{isActive:s==="register",onClick:()=>t("register"),label:"Rejestracja"}),e.jsx(F,{isActive:s==="forgot",onClick:()=>t("forgot"),label:"Reset hasła"})]}),F=({isActive:s,onClick:t,label:a})=>e.jsx("button",{onClick:t,className:`flex-1 px-2 sm:px-4 py-2 rounded-md text-xs sm:text-sm font-medium transition-colors duration-150 ${s?"bg-white text-gray-900 border border-gray-900 border-2":"bg-white text-gray-900"}`,children:a}),Ne=j.object({email:j.string().min(1,"Email jest wymagany").email("Nieprawidłowy format email"),password:j.string().min(1,"Hasło jest wymagane"),rememberMe:j.boolean().optional()}),Se=j.object({username:j.string().min(3,"Nazwa użytkownika musi mieć minimum 3 znaki").max(20,"Nazwa użytkownika nie może przekraczać 20 znaków").regex(/^[a-zA-Z0-9_]+$/,"Nazwa użytkownika może zawierać tylko litery, cyfry i podkreślenia"),email:j.string().email("Nieprawidłowy format adresu email"),password:j.string().min(8,"Hasło musi mieć minimum 8 znaków").regex(/[A-Z]/,"Hasło musi zawierać przynajmniej jedną wielką literę").regex(/[a-z]/,"Hasło musi zawierać przynajmniej jedną małą literę").regex(/[0-9]/,"Hasło musi zawierać przynajmniej jedną cyfrę"),confirmPassword:j.string(),acceptPrivacy:j.boolean().refine(s=>s===!0,{message:"Musisz zaakceptować politykę prywatności"})}).refine(s=>s.password===s.confirmPassword,{message:"Hasła muszą być identyczne",path:["confirmPassword"]}),ve=j.object({email:j.string().min(1,"Email jest wymagany").email("Nieprawidłowy format email")}),Ee=j.object({password:j.string().min(8,"Hasło musi mieć minimum 8 znaków").regex(/[A-Z]/,"Hasło musi zawierać przynajmniej jedną wielką literę").regex(/[a-z]/,"Hasło musi zawierać przynajmniej jedną małą literę").regex(/[0-9]/,"Hasło musi zawierać przynajmniej jedną cyfrę"),confirmPassword:j.string()}).refine(s=>s.password===s.confirmPassword,{message:"Hasła muszą być identyczne",path:["confirmPassword"]}),O=x.memo(({rememberMe:s})=>{const{loginWithGoogle:t}=v(),[a,o]=x.useState(null),r=l=>{t(l,s)},i=()=>{o("Logowanie przez Google nie powiodło się. Spróbuj ponownie później.")};return e.jsxs(e.Fragment,{children:[a&&e.jsx("div",{className:"mb-4 p-3 rounded-lg bg-red-50 border border-red-200 text-red-600 text-sm",children:a}),e.jsx("div",{className:"w-full",children:e.jsx(ee,{onSuccess:r,onError:i,theme:"outline",shape:"rectangular",text:"continue_with",locale:"pl",useOneTap:!1,context:"signin",type:"standard",size:"large",width:"100%"})})]})});O.displayName="GoogleLoginButton";const D=x.memo(({register:s})=>e.jsx("div",{className:"flex items-center",children:e.jsxs("label",{className:"flex items-center gap-2 cursor-pointer group",children:[e.jsxs("div",{className:"relative",children:[e.jsx("input",w({type:"checkbox",className:"peer sr-only"},s("rememberMe"))),e.jsx("div",{className:"w-4 h-4 border border-gray-300 rounded bg-white peer-checked:bg-[#f7df1e] peer-checked:border-[#f7df1e] transition-colors"}),e.jsx("div",{className:"absolute inset-0 flex items-center justify-center text-black scale-0 peer-checked:scale-100 transition-transform",children:e.jsx("svg",{xmlns:"http://www.w3.org/2000/svg",className:"h-3 w-3",viewBox:"0 0 20 20",fill:"currentColor",children:e.jsx("path",{fillRule:"evenodd",d:"M16.707 5.293a1 1 0 010 1.414l-8 8a1 1 0 01-1.414 0l-4-4a1 1 0 011.414-1.414L8 12.586l7.293-7.293a1 1 0 011.414 0z",clipRule:"evenodd"})})})]}),e.jsx("span",{className:"text-sm text-gray-600 group-hover:text-gray-800 transition-colors",children:"Zapamiętaj mnie"})]})}));D.displayName="RememberMeCheckbox";const Pe=()=>{var u,g;const{login:s,loading:t,error:a}=v(),{register:o,handleSubmit:r,formState:{errors:i},watch:l}=A({resolver:L(Ne),mode:"onSubmit"}),d=l("rememberMe",!1)||!1,[c,m]=x.useState(!1),n=y=>p(void 0,null,function*(){yield s(y.email,y.password,y.rememberMe)});return e.jsxs("form",{onSubmit:r(n),className:"space-y-6",children:[a&&e.jsx(E,{message:a,withIcon:!0,variant:"standard"}),e.jsx(O,{rememberMe:d}),e.jsxs("div",{className:"relative flex items-center justify-center my-6",children:[e.jsx("div",{className:"border-t border-gray-200 w-full"}),e.jsx("span",{className:"bg-white px-4 text-sm text-gray-500 absolute",children:"LUB"})]}),e.jsx(N,w({type:"email",label:"Email",placeholder:"twoj@email.com",icon:e.jsx(C,{}),error:(u=i.email)==null?void 0:u.message},o("email"))),e.jsx(N,k(w({type:c?"text":"password",label:"Hasło",placeholder:"Wprowadź hasło",icon:e.jsx(se,{}),error:(g=i.password)==null?void 0:g.message},o("password")),{rightIcon:e.jsx("button",{type:"button",onClick:()=>m(!c),className:"bg-white focus:outline-none hover:text-gray-600 transition-colors",children:c?e.jsx(te,{}):e.jsx(re,{})})})),e.jsxs("div",{className:"flex items-center justify-between",children:[e.jsx(D,{register:o}),e.jsx("a",{href:"/logowanie?mode=forgot",className:"text-sm text-[#f7df1e] hover:underline font-medium",children:"Zapomniałeś hasła?"})]}),e.jsx("div",{className:"pt-2",children:e.jsx(W,{isLoading:t,loadingText:"Logowanie...",children:"Zaloguj się"})})]})},$e=()=>{const[s,t]=x.useState(!1);return{isVisible:s,toggleVisibility:()=>t(o=>!o),inputType:s?"text":"password"}},$=x.memo(({register:s,error:t})=>e.jsxs("div",{className:"space-y-1",children:[e.jsxs("label",{className:"flex items-start gap-2 cursor-pointer group",children:[e.jsx("input",w({type:"checkbox",className:`
-          mt-1 w-4 h-4 rounded bg-dark text-js 
-          transition-colors duration-200 ease-in-out
-          ${t?"border-red-500 focus:ring-red-500/20":"border-js/20 focus:ring-js/20 hover:border-js/30"}
-        `},s)),e.jsxs("span",{className:"text-sm text-gray-400 group-hover:text-gray-300 transition-colors",children:["Akceptuję"," ",e.jsx(S,{to:"/polityka-prywatnosci",className:"text-js hover:text-js/80 underline transition-colors",target:"_blank",rel:"noopener noreferrer",children:"politykę prywatności"})]})]}),t&&e.jsx(f.p,{initial:{opacity:0,y:-10},animate:{opacity:1,y:0},className:"text-sm text-red-400",children:t})]}));$.displayName="PrivacyPolicyCheckbox";const R=({initialError:s=null,resetOnSuccess:t=!0,showToasts:a=!0}={})=>{const[o,r]=x.useState(null),[i,l]=x.useState(s);x.useEffect(()=>{s!=null&&l(s)},[s]);const d=x.useCallback(u=>{r(u),l(null),a&&U.success(u),t&&setTimeout(()=>{r(null)},5e3)},[t,a]),c=x.useCallback(u=>{l(u),r(null),a&&U.error(u)},[a]),m=x.useCallback(u=>{var y;let g="";if(u instanceof Error)g=u.message;else if(typeof u=="string")g=u;else if(u&&typeof u=="object"){const h=u;g=h.message||h.error||h.errors&&((y=h.errors[0])==null?void 0:y.message)||"Wystąpił nieznany błąd. Spróbuj ponownie później."}else g="Wystąpił nieznany błąd. Spróbuj ponownie później.";return c(g),g},[c]),n=x.useCallback(()=>{r(null),l(null)},[]);return{successMessage:o,errorMessage:i,setSuccess:d,setError:c,handleError:m,resetStatus:n,hasSuccess:!!o,hasError:!!i}},Ae=({schema:s,defaultValues:t,mode:a="onSubmit",resetOnSuccess:o=!0,onSuccess:r,onError:i})=>{const[l,d]=x.useState(!1),c=R({resetOnSuccess:o}),m=A({resolver:L(s),defaultValues:t,mode:a}),n=x.useCallback(h=>p(void 0,null,function*(){try{d(!0),c.resetStatus(),r&&(yield r(h))}catch(H){c.handleError(H),i&&i(H)}finally{d(!1)}}),[r,i,c]),u=x.useCallback(h=>{n(h)},[n]),g=x.useCallback(h=>{console.error("Form validation errors:",h),c.setError("Formularz zawiera błędy. Sprawdź wprowadzone dane.")},[c]),y=m.handleSubmit(u,g);return k(w({},m),{isSubmitting:l,errorMessage:c.errorMessage,successMessage:c.successMessage,onFormSubmit:y,handleSubmitWithValidation:n})},Le=()=>{const{register:s,loading:t,error:a}=v(),[o,r]=x.useState(!1),[i,l]=x.useState(null),d=Ae({schema:Se,mode:"onChange",defaultValues:{acceptPrivacy:!1},onSuccess(m){return p(this,null,function*(){try{r(!0),l(null),yield s(m.email,m.password,m.username)}catch(n){const u=n instanceof Error?n.message:"Wystąpił błąd podczas rejestracji";l(u)}finally{r(!1)}})}});return x.useEffect(()=>{a&&l(a)},[a]),{register:d.register,errors:d.formState.errors,loading:t||o||d.isSubmitting,handleSubmit:d.onFormSubmit,errorMessage:i||a||d.errorMessage,successMessage:d.successMessage}},Me=()=>{var i,l,d,c,m;const{register:s,errors:t,loading:a,handleSubmit:o,errorMessage:r}=Le();return e.jsxs(I,{onSubmit:o,children:[r&&e.jsx(E,{message:r,variant:"alert",title:"Błąd rejestracji",className:"mb-6 text-base border-2"}),e.jsx(N,w({type:"text",label:"Nazwa użytkownika",placeholder:"jankowalski",icon:e.jsx(ae,{}),error:(i=t.username)==null?void 0:i.message},s("username"))),e.jsx(N,w({type:"email",label:"Email",placeholder:"twoj@email.com",icon:e.jsx(C,{}),error:(l=t.email)==null?void 0:l.message},s("email"))),e.jsx(P,w({label:"Hasło",error:(d=t.password)==null?void 0:d.message},s("password"))),e.jsx(P,w({label:"Potwierdź hasło",error:(c=t.confirmPassword)==null?void 0:c.message},s("confirmPassword"))),e.jsx($,{register:s("acceptPrivacy"),error:(m=t.acceptPrivacy)==null?void 0:m.message}),e.jsx(W,{isLoading:a,loadingText:"Rejestracja...",children:"Zarejestruj się"}),r&&e.jsx("div",{className:"mt-4",children:e.jsx(E,{message:r,withIcon:!0,className:"text-sm"})})]})},Fe=()=>{const{forgotPassword:s,loading:t,error:a}=v(),o=R({initialError:a}),{register:r,handleSubmit:i,reset:l,formState:{errors:d}}=A({resolver:L(ve)});return w({register:r,errors:d,loading:t,handleSubmit:i(m=>p(void 0,null,function*(){try{o.resetStatus();const n=yield s(m.email);o.setSuccess(n),l()}catch(n){o.handleError(n)}}))},o)},Ce=()=>{var l;const{register:s,errors:t,loading:a,handleSubmit:o,errorMessage:r,successMessage:i}=Fe();return e.jsxs(I,{onSubmit:o,children:[r&&e.jsx(E,{message:r,withIcon:!0,variant:"alert",title:"Błąd resetowania hasła"}),i&&e.jsxs("div",{className:"p-4 bg-green-50 border border-green-200 rounded-lg text-green-700 text-sm",children:[e.jsx("p",{className:"font-semibold mb-1",children:"Email wysłany!"}),e.jsx("p",{children:i})]}),!i&&e.jsxs(e.Fragment,{children:[e.jsx("p",{className:"text-gray-600 text-sm mb-6",children:"Podaj swój adres email, a wyślemy Ci link do zresetowania hasła."}),e.jsx(N,w({type:"email",label:"Email",placeholder:"twoj@email.com",icon:e.jsx(C,{}),error:(l=t.email)==null?void 0:l.message},s("email"))),e.jsx(W,{isLoading:a,loadingText:"Wysyłanie...",children:"Wyślij link resetujący"})]})]})},We=({mode:s}={})=>{const[t,a]=Z(),o=t.get("mode")||"login",[r,i]=x.useState(s||o);x.useEffect(()=>{i(s||o)},[s,o]);const l=d=>{i(d),a({mode:d})};return e.jsxs(f.div,{initial:{opacity:0,y:20},animate:{opacity:1,y:0},transition:{duration:.3},className:"w-full max-w-sm sm:max-w-md bg-white rounded-xl sm:rounded-2xl p-6 sm:p-8 border border-gray-100",role:"main","aria-label":"Formularz autoryzacji",children:[e.jsxs("div",{className:"flex flex-col justify-center items-center mb-6 sm:mb-8",children:[e.jsx("div",{className:"w-10 h-10 sm:w-12 sm:h-12 bg-[#f7df1e] rounded-lg flex items-center justify-center mb-3 sm:mb-4",children:e.jsx("span",{className:"text-black font-bold text-lg sm:text-xl",children:"JS"})}),e.jsx("h2",{className:"text-xl sm:text-2xl font-bold text-gray-900",children:"CodeLinesJS"})]}),e.jsxs("div",{className:"text-center mb-6 sm:mb-8",children:[e.jsxs("h1",{className:"text-xl sm:text-2xl font-medium text-gray-900 mb-2",children:[r==="login"&&"Zaloguj się",r==="register"&&"Dołącz do nas",r==="forgot"&&"Odzyskaj hasło"]}),e.jsxs("p",{className:"text-xs sm:text-sm text-gray-600",children:[r==="login"&&"Wprowadź swoje dane aby uzyskać dostęp do konta",r==="register"&&"Rozpocznij swoją przygodę z programowaniem",r==="forgot"&&"Wprowadź swój email aby zresetować hasło"]})]}),e.jsx(ke,{activeTab:r,onTabChange:l}),e.jsx(me,{mode:"wait",children:e.jsxs(f.div,{initial:{opacity:0,x:20},animate:{opacity:1,x:0},exit:{opacity:0,x:-20},transition:{duration:.2},children:[r==="login"&&e.jsx(Pe,{}),r==="register"&&e.jsx(Me,{}),r==="forgot"&&e.jsx(Ce,{})]},r)}),e.jsx("div",{className:"mt-4 sm:mt-6 text-center text-xs text-gray-500",children:e.jsxs("p",{children:["Logując się, akceptujesz"," ",e.jsx(S,{to:"/polityka-prywatnosci",className:"text-[#f7df1e] hover:underline font-medium",children:"Politykę Prywatności"}),"."]})})]})},Ie=({children:s,title:t,subtitle:a,mode:o})=>e.jsx("div",{className:"w-full max-w-sm sm:max-w-md",children:s?e.jsxs("div",{className:"w-full bg-white rounded-xl sm:rounded-2xl p-6 sm:p-8 border border-gray-100",children:[e.jsxs("div",{className:"flex flex-col justify-center items-center mb-6 sm:mb-8",children:[e.jsx("div",{className:"w-10 h-10 sm:w-12 sm:h-12 bg-[#f7df1e] rounded-lg flex items-center justify-center mb-3 sm:mb-4",children:e.jsx("span",{className:"text-black font-bold text-lg sm:text-xl",children:"JS"})}),e.jsx("h2",{className:"text-xl sm:text-2xl font-bold text-gray-900",children:"CodeLinesJS"})]}),e.jsxs("div",{className:"text-center mb-6 sm:mb-8",children:[e.jsx("h1",{className:"text-xl sm:text-2xl font-medium text-gray-900 mb-2",children:t||"Zaloguj się"}),a&&e.jsx("p",{className:"text-xs sm:text-sm text-gray-600",children:a})]}),s]}):e.jsx(We,{mode:o})}),G=({children:s,title:t,subtitle:a}={})=>{const[o]=Z(),r=o.get("mode")||"login",i=()=>{switch(r){case"register":return"Rejestracja | CodeLinesJS";case"forgot":return"Reset hasła | CodeLinesJS";default:return"Logowanie | CodeLinesJS"}},l=()=>{switch(r){case"register":return"Dołącz do platformy CodeLinesJS - rozpocznij swoją przygodę z programowaniem w JavaScript.";case"forgot":return"Zresetuj hasło do swojego konta CodeLinesJS i wróć do nauki programowania.";default:return"Logowanie do platformy CodeLinesJS - dołącz do nas i rozwijaj swoje umiejętności w przyjaznym środowisku."}};return e.jsxs("main",{className:"min-h-screen w-full flex",children:[e.jsxs(oe,{children:[e.jsx("title",{children:i()}),e.jsx("meta",{name:"description",content:l()})]}),e.jsx("div",{className:"hidden lg:flex w-1/2 bg-black items-center justify-center p-8",children:e.jsx(ze,{})}),e.jsx("div",{className:"w-full lg:w-1/2 bg-[#f7df1e] flex items-center justify-center p-4 sm:p-6 lg:p-8",children:e.jsx(ie,{children:e.jsx(Ie,{children:s,title:t,subtitle:a,mode:r})})})]})},Re=({errorMessage:s})=>e.jsxs(f.div,{initial:{opacity:0},animate:{opacity:1},className:"space-y-6",children:[e.jsxs("div",{className:"p-4 bg-red-500/10 border border-red-500/20 rounded-lg text-red-400 text-sm flex items-start",children:[e.jsx(V,{className:"mr-2 mt-0.5 flex-shrink-0"}),e.jsxs("div",{children:[e.jsx("p",{className:"font-semibold mb-1",children:"Błąd tokenu resetowania hasła"}),e.jsx("p",{children:s||"Nieprawidłowy lub wygasły token resetowania hasła. Sprawdź, czy link jest poprawny lub spróbuj ponownie zresetować hasło."})]})]}),e.jsxs("div",{className:"text-center mt-6",children:[e.jsx("p",{className:"text-gray-400 mb-4",children:"Możesz spróbować ponownie zresetować hasło:"}),e.jsx(S,{to:"/logowanie",className:"text-js hover:underline",children:"Wróć do strony logowania"})]})]}),He=({message:s})=>s?e.jsxs(e.Fragment,{children:[e.jsx(le,{message:s,title:"Hasło zostało zmienione"}),e.jsxs("div",{className:"text-center mt-4",children:[e.jsx("p",{className:"text-gray-400 mb-2",children:"Za chwilę zostaniesz przekierowany do strony logowania."}),e.jsx(S,{to:"/logowanie",className:"text-js hover:underline",children:"Przejdź do logowania"})]})]}):null,Be=()=>{const{token:s}=_(),{resetPassword:t,loading:a,error:o}=v(),r=R({initialError:o}),[i,l]=x.useState(!1),{register:d,handleSubmit:c,formState:{errors:m},reset:n}=A({resolver:L(Ee),mode:"onSubmit"});return x.useEffect(()=>{s?s.length<10&&(l(!0),r.setError("Token resetowania hasła jest nieprawidłowy. Sprawdź, czy link jest poprawny.")):(l(!0),r.setError("Brak tokenu resetowania hasła. Sprawdź, czy link jest poprawny."))},[s,r]),w({register:d,errors:m,loading:a,tokenError:i,handleSubmit:c(g=>p(void 0,null,function*(){if(!s){r.setError("Brak tokenu resetowania hasła. Sprawdź, czy link jest poprawny.");return}try{r.resetStatus();const y=yield t(s,g.password,g.confirmPassword);r.setSuccess(y),n()}catch(y){const h=r.handleError(y);typeof h=="string"&&(h.includes("token")||h.includes("Token")||h.includes("wygasł"))&&l(!0)}}))},r)},Te=()=>{var d,c;const{register:s,errors:t,loading:a,handleSubmit:o,errorMessage:r,successMessage:i,tokenError:l}=Be();return l?e.jsx(Re,{errorMessage:r}):e.jsxs(I,{onSubmit:o,children:[r&&e.jsx(ce,{message:r,title:"Błąd resetowania hasła"}),e.jsx(He,{message:i}),!i&&e.jsxs(e.Fragment,{children:[e.jsx("p",{className:"text-gray-400 text-sm mb-6",children:"Wprowadź nowe hasło dla swojego konta. Hasło musi zawierać co najmniej 8 znaków, w tym jedną wielką literę, jedną małą literę i jedną cyfrę."}),e.jsx(P,w({label:"Nowe hasło",error:(d=t.password)==null?void 0:d.message},s("password"))),e.jsx(P,w({label:"Potwierdź nowe hasło",error:(c=t.confirmPassword)==null?void 0:c.message},s("confirmPassword"))),e.jsx(de,{type:"submit",className:"w-full",disabled:a,children:a?"Resetowanie hasła...":"Zresetuj hasło"})]})]})},Ue=()=>e.jsxs(f.div,{initial:{opacity:0},animate:{opacity:1},className:"space-y-6",children:[e.jsxs("div",{className:"p-4 bg-red-500/10 border border-red-500/20 rounded-lg text-red-400 text-sm flex items-start",children:[e.jsx(V,{className:"mr-2 mt-0.5 flex-shrink-0"}),e.jsxs("div",{children:[e.jsx("p",{className:"font-semibold mb-1",children:"Błąd tokenu resetowania hasła"}),e.jsx("p",{children:"Nieprawidłowy lub wygasły token resetowania hasła. Sprawdź, czy link jest poprawny lub spróbuj ponownie zresetować hasło."})]})]}),e.jsxs("div",{className:"text-center mt-6",children:[e.jsx("p",{className:"text-gray-400 mb-4",children:"Możesz spróbować ponownie zresetować hasło na stronie logowania:"}),e.jsxs(S,{to:"/logowanie",className:"inline-flex items-center text-js hover:underline",children:[e.jsx(ne,{className:"mr-2"})," Wróć do strony logowania"]})]})]}),Je=()=>{const{token:s}=_();return s?e.jsx(G,{title:"Resetowanie hasła",subtitle:"Wprowadź nowe hasło dla swojego konta",children:e.jsx(Te,{})}):e.jsx(G,{title:"Nieprawidłowy link",subtitle:"Link do resetowania hasła jest nieprawidłowy lub wygasł.",children:e.jsx(Ue,{})})},qe=Object.freeze(Object.defineProperty({__proto__:null,ResetPasswordPage:Je},Symbol.toStringTag,{value:"Module"}));export{J as A,$e as a,G as b,qe as i,v as u};
