--- server/public/assets/app-components-DCNmHqgN.js
@@ -1,58 +1,0 @@
-var Fe=Object.defineProperty,Ee=Object.defineProperties;var Le=Object.getOwnPropertyDescriptors;var S=Object.getOwnPropertySymbols;var J=Object.prototype.hasOwnProperty,_=Object.prototype.propertyIsEnumerable;var E=(t,s,a)=>s in t?Fe(t,s,{enumerable:!0,configurable:!0,writable:!0,value:a}):t[s]=a,j=(t,s)=>{for(var a in s||(s={}))J.call(s,a)&&E(t,a,s[a]);if(S)for(var a of S(s))_.call(s,a)&&E(t,a,s[a]);return t},b=(t,s)=>Ee(t,Le(s));var C=(t,s)=>{var a={};for(var i in t)J.call(t,i)&&s.indexOf(i)<0&&(a[i]=t[i]);if(t!=null&&S)for(var i of S(t))s.indexOf(i)<0&&_.call(t,i)&&(a[i]=t[i]);return a};var L=(t,s,a)=>E(t,typeof s!="symbol"?s+"":s,a);var B=(t,s,a)=>new Promise((i,r)=>{var l=x=>{try{c(a.next(x))}catch(m){r(m)}},d=x=>{try{c(a.throw(x))}catch(m){r(m)}},c=x=>x.done?i(x.value):Promise.resolve(x.value).then(l,d);c((a=a.apply(t,s)).next())});import{r as o,j as e,l as Be,m as De,n as $e,u as He,o as M,p as ee,g as N,L as f,N as Re,b as We,c as Ve,d as Je,q as _e,s as Oe,I as Ge,t as R,v as Ze,w as te,x as Ke,y as Ue,z as se,A as ae,B as Ye,C as ie,D as qe,E as re,J as ne,K as Qe,M as Xe,O as et}from"./vendor-react-DZE_uCEP.js";import{m as n,A as I}from"./vendor-framer-motion-D6A4MpK3.js";import{u as tt,a as st,A as oe}from"./app-auth-BOk5gCNS.js";const O=o.memo(({size:t="md",fullScreen:s=!1,text:a="Ładowanie..."})=>{const i={sm:"h-4 w-4",md:"h-5 w-5",lg:"h-8 w-8"},r={sm:"text-sm",md:"text-base",lg:"text-xl"},l=s?"min-h-screen w-full flex items-center justify-center bg-gradient-to-b from-dark via-dark-medium to-dark":"flex items-center justify-center p-4";return e.jsx("div",{className:l,role:"status","aria-label":a,children:e.jsxs(n.div,{initial:{opacity:0},animate:{opacity:1},className:`font-space text-js ${r[t]}`,children:[e.jsx("span",{className:"sr-only",children:a}),e.jsxs("div",{className:"flex items-center gap-2",children:[e.jsxs("svg",{className:`animate-spin ${i[t]}`,xmlns:"http://www.w3.org/2000/svg",fill:"none",viewBox:"0 0 24 24","aria-hidden":"true",children:[e.jsx("circle",{className:"opacity-25",cx:"12",cy:"12",r:"10",stroke:"currentColor",strokeWidth:"4"}),e.jsx("path",{className:"opacity-75",fill:"currentColor",d:"M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4z"})]}),a]})]})})}),Hs=Object.freeze(Object.defineProperty({__proto__:null,LoadingSpinner:O,default:O},Symbol.toStringTag,{value:"Module"})),le=o.memo(({currentQuestion:t,totalQuestions:s,score:a})=>e.jsxs("div",{className:"flex justify-between items-center mb-4",children:[e.jsxs("span",{className:"text-gray-400",children:["Pytanie ",t+1,"/",s]}),e.jsxs("span",{className:"text-js font-bold",children:["Wynik: ",a]})]}));le.displayName="QuestionHeader";const ce=o.memo(({code:t})=>e.jsx("pre",{className:"bg-dark/50 p-4 rounded-lg mb-6 overflow-x-auto border border-js/10",children:e.jsx("code",{className:"text-js/90 font-mono",children:t})}));ce.displayName="CodeBlock";const de=o.memo(({option:t,index:s,selectedAnswer:a,isCorrect:i,correctAnswer:r,onSelect:l})=>e.jsx(n.button,{whileHover:{scale:1.02},whileTap:{scale:.98},disabled:a!==null,onClick:()=>l(s),className:`
-      p-4 rounded-lg text-left transition-all
-      ${a===null?"bg-dark/50 hover:bg-dark/70 text-js border border-js/10":a===s?i?"bg-green-500/20 border-2 border-green-500/50":"bg-red-500/20 border-2 border-red-500/50":s===r&&!i?"bg-green-500/20 border-2 border-green-500/50":"bg-dark/50 opacity-50 border border-js/10"}
-    `,children:e.jsxs("div",{className:"flex items-center justify-between",children:[e.jsx("span",{children:t}),a===s&&e.jsx(n.span,{initial:{scale:0},animate:{scale:1},transition:{type:"spring",stiffness:200},children:i?e.jsx(Be,{className:"text-green-500"}):e.jsx(De,{className:"text-red-500"})})]})}));de.displayName="AnswerOption";const T=[{id:1,question:"Co zwróci poniższy kod?",code:"typeof typeof 42",options:["'number'","'string'","'undefined'","'object'"],correctAnswer:1},{id:2,question:"Jaki będzie wynik?",code:"[1, 2, 3].map(num => num * 2)",options:["[2, 4, 6]","[1, 2, 3]","undefined","Error"],correctAnswer:0}],at=t=>{const[s,a]=o.useState(0),[i,r]=o.useState(0),[l,d]=o.useState(null),[c,x]=o.useState(null),m=o.useCallback(h=>{d(h);const u=h===T[s].correctAnswer;x(u),u&&r(k=>k+100),setTimeout(()=>{s<T.length-1?(a(k=>k+1),d(null),x(null)):t(i+(u?100:0))},1500)},[s,t,i]);return{currentQuestion:s,score:i,selectedAnswer:l,isCorrect:c,handleAnswer:m}},xe=o.memo(({onComplete:t})=>{const{currentQuestion:s,score:a,selectedAnswer:i,isCorrect:r,handleAnswer:l}=at(t),d=T[s];return e.jsxs(n.div,{initial:{opacity:0,y:20},animate:{opacity:1,y:0},exit:{opacity:0,y:-20},className:"bg-dark/50 backdrop-blur-lg rounded-xl p-6 border border-js/20",children:[e.jsxs("div",{className:"mb-6",children:[e.jsx(le,{currentQuestion:s,totalQuestions:T.length,score:a}),e.jsx("h3",{className:"text-xl text-js mb-4",children:d.question}),e.jsx(ce,{code:d.code})]}),e.jsx("div",{className:"grid grid-cols-1 gap-4",children:d.options.map((c,x)=>e.jsx(de,{option:c,index:x,selectedAnswer:i,isCorrect:r,correctAnswer:d.correctAnswer,onSelect:l},x))})]})});xe.displayName="JSGame";const it=()=>e.jsxs("div",{className:"absolute inset-0",children:[e.jsx("div",{className:"absolute top-0 left-0 w-full h-full bg-dark opacity-90"}),e.jsx("div",{className:"absolute -top-40 -right-40 w-96 h-96 bg-js rounded-full blur-[150px] opacity-20"}),e.jsx("div",{className:"absolute -bottom-40 -left-40 w-96 h-96 bg-js rounded-full blur-[150px] opacity-20"}),e.jsx("div",{className:"absolute inset-0 opacity-[0.02]",style:{backgroundImage:`url("data:image/svg+xml,%3Csvg width='60' height='60' viewBox='0 0 60 60' xmlns='http://www.w3.org/2000/svg'%3E%3Cg fill='none' fill-rule='evenodd'%3E%3Cg fill='%23f7df1e' fill-opacity='1'%3E%3Cpath d='M36 34v-4h-2v4h-4v2h4v4h2v-4h4v-2h-4zm0-30V0h-2v4h-4v2h4v4h2V6h4V4h-4zM6 34v-4H4v4H0v2h4v4h2v-4h4v-2H6zM6 4V0H4v4H0v2h4v4h2V6h4V4H6z'/%3E%3C/g%3E%3C/g%3E%3C/svg%3E")`,backgroundSize:"30px 30px"}})]}),me=o.memo(()=>e.jsxs("div",{className:"text-center mb-8",children:[e.jsx(n.div,{initial:{scale:0},animate:{scale:1},transition:{type:"spring",stiffness:200,damping:15},className:"inline-block",children:e.jsx($e,{className:"text-[150px] text-js"})}),e.jsx(n.h1,{className:"text-7xl font-bold text-js mb-4",initial:{x:-100,opacity:0},animate:{x:0,opacity:1},transition:{delay:.2},children:"404"}),e.jsx(n.p,{className:"text-xl text-gray-400 mb-8",initial:{x:100,opacity:0},animate:{x:0,opacity:1},transition:{delay:.3},children:"Ups! Wygląda na to, że zabłądziłeś w kodzie..."})]}));me.displayName="ErrorHeader";const pe=o.memo(({score:t,onStartGame:s})=>{const a=He();return e.jsx("div",{className:"space-y-4",children:e.jsxs(n.div,{initial:{opacity:0},animate:{opacity:1},transition:{delay:.4},className:"bg-dark/50 backdrop-blur-lg rounded-xl p-6 border border-js/20",children:[e.jsxs("h2",{className:"text-xl text-js mb-4 flex items-center gap-2",children:[e.jsx(M,{className:"text-js"}),"Zagraj w mini-grę JavaScript!"]}),e.jsx("p",{className:"text-gray-400 mb-4",children:"Rozwiąż zadania JavaScript i zdobądź punkty. Może nawet pobijesz swój rekord?"}),t>0&&e.jsxs("p",{className:"text-lg text-js font-bold mb-4",children:["Twój ostatni wynik: ",t," pkt"]}),e.jsxs("div",{className:"flex gap-4",children:[e.jsxs(n.button,{whileHover:{scale:1.05},whileTap:{scale:.95},onClick:s,className:`px-6 py-3 rounded-lg bg-js text-dark font-bold flex items-center gap-2 \r
-                     hover:bg-js/90 transition-all`,children:[e.jsx(M,{}),t>0?"Zagraj ponownie":"Rozpocznij grę"]}),e.jsxs(n.button,{whileHover:{scale:1.05},whileTap:{scale:.95},onClick:()=>a("/"),className:`px-6 py-3 rounded-lg border border-js/20 text-js font-bold flex items-center gap-2 \r
-                     hover:bg-js/10 transition-all`,children:[e.jsx(ee,{}),"Wróć do strony głównej"]})]})]})})});pe.displayName="GameIntro";const he=o.memo(()=>{const[t,s]=o.useState(!1),[a,i]=o.useState(0),r=o.useCallback(l=>{i(l),s(!1)},[]);return e.jsxs("div",{className:"min-h-screen bg-gradient-to-b from-dark via-dark-medium to-dark flex items-center justify-center p-4 relative overflow-hidden",children:[e.jsx(it,{}),e.jsxs(n.div,{initial:{opacity:0,y:20},animate:{opacity:1,y:0},className:"max-w-2xl w-full relative z-10",children:[e.jsx(me,{}),t?e.jsx(xe,{onComplete:r}):e.jsx(pe,{score:a,onStartGame:()=>s(!0)})]})]})});he.displayName="ErrorPage";const Rs=Object.freeze(Object.defineProperty({__proto__:null,ErrorPage:he},Symbol.toStringTag,{value:"Module"})),rt=768,p=()=>{const[t,s]=o.useState(!1),a=o.useCallback(()=>{s(window.innerWidth<rt)},[]);return o.useEffect(()=>{a();let i;const r=()=>{clearTimeout(i),i=setTimeout(a,100)};return window.addEventListener("resize",r),()=>{window.removeEventListener("resize",r),clearTimeout(i)}},[a]),t};class Ws extends o.Component{constructor(){super(...arguments);L(this,"state",{hasError:!1})}static getDerivedStateFromError(a){return{hasError:!0,error:a}}componentDidCatch(a,i){console.error("Błąd w komponencie:",a),console.error("Szczegóły błędu:",i)}render(){return this.state.hasError?this.props.fallback||e.jsxs("div",{className:"flex flex-col items-center justify-center p-8 text-center",children:[e.jsx(N,{className:"text-4xl text-red-500 mb-4"}),e.jsx("h2",{className:"text-xl font-bold text-gray-200 mb-2",children:"Ups! Coś poszło nie tak"}),e.jsx("p",{className:"text-gray-400 mb-4",children:"Wystąpił błąd podczas ładowania tej sekcji."}),e.jsx("button",{onClick:()=>window.location.reload(),className:"px-4 py-2 rounded-lg bg-gradient-to-r from-indigo-500 to-purple-500 text-white font-medium hover:from-indigo-600 hover:to-purple-600 transition-colors",children:"Odśwież stronę"})]}):this.props.children}}const nt=()=>e.jsxs("div",{className:"flex flex-col items-center justify-center p-8 text-center",children:[e.jsx(N,{className:"text-4xl text-red-500 mb-4"}),e.jsx("h2",{className:"text-xl font-bold text-gray-200 mb-2",children:"Ups! Coś poszło nie tak"}),e.jsx("p",{className:"text-gray-400 mb-4",children:"Wystąpił nieoczekiwany błąd podczas ładowania formularza autoryzacji."}),e.jsxs("div",{className:"flex flex-col sm:flex-row gap-4 mt-2",children:[e.jsx(f,{to:"/",className:"px-4 py-2 rounded-lg bg-gray-700 text-white font-medium hover:bg-gray-600 transition-colors",children:"Wróć do strony głównej"}),e.jsx("button",{onClick:()=>window.location.reload(),className:"px-4 py-2 rounded-lg bg-gradient-to-r from-indigo-500 to-purple-500 text-white font-medium hover:from-indigo-600 hover:to-purple-600 transition-colors",children:"Odśwież stronę"})]})]});class Vs extends o.Component{constructor(){super(...arguments);L(this,"state",{hasError:!1})}static getDerivedStateFromError(a){return{hasError:!0,error:a}}componentDidCatch(a,i){console.error("Błąd w komponencie Auth:",a),console.error("Szczegóły błędu:",i)}render(){return this.state.hasError?this.props.fallback||e.jsx(nt,{}):this.props.children}}const ot=({children:t})=>{const{isAuthenticated:s,loading:a,isAuthChecking:i}=tt();return a||i?e.jsx("div",{className:"flex items-center justify-center min-h-screen",children:e.jsx("div",{className:"animate-spin rounded-full h-8 w-8 border-t-2 border-b-2 border-primary"})}):s?e.jsx(e.Fragment,{children:t}):e.jsx(Re,{to:"/logowanie",replace:!0})},Js=Object.freeze(Object.defineProperty({__proto__:null,ProtectedRoute:ot},Symbol.toStringTag,{value:"Module"})),lt=o.memo(()=>e.jsx("div",{className:"min-h-screen w-full flex items-center justify-center bg-gradient-to-b from-dark via-dark-medium to-dark",role:"status","aria-label":"Ładowanie strony",children:e.jsxs(n.div,{initial:{opacity:0},animate:{opacity:1},className:"text-2xl font-space text-js",children:[e.jsx("span",{className:"sr-only",children:"Ładowanie..."}),e.jsxs("div",{className:"flex items-center gap-2",children:[e.jsxs("svg",{className:"animate-spin h-5 w-5",xmlns:"http://www.w3.org/2000/svg",fill:"none",viewBox:"0 0 24 24","aria-hidden":"true",children:[e.jsx("circle",{className:"opacity-25",cx:"12",cy:"12",r:"10",stroke:"currentColor",strokeWidth:"4"}),e.jsx("path",{className:"opacity-75",fill:"currentColor",d:"M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4z"})]}),"Ładowanie..."]})]})})),ct=o.memo(({level:t,rewards:s,onClose:a,isVisible:i})=>e.jsx(I,{children:i&&e.jsxs(n.div,{initial:{opacity:0,y:50},animate:{opacity:1,y:0},exit:{opacity:0,y:50},className:"fixed bottom-4 right-4 bg-dark/95 border border-js/20 rounded-lg p-6 shadow-lg max-w-sm",children:[e.jsxs("div",{className:"flex items-center gap-4 mb-4",children:[e.jsx("div",{className:"w-12 h-12 rounded-full bg-js/20 flex items-center justify-center",children:e.jsx("span",{className:"text-2xl",children:"🎉"})}),e.jsxs("div",{children:[e.jsx("h3",{className:"text-js font-bold text-lg",children:"Nowy Poziom!"}),e.jsxs("p",{className:"text-gray-400",children:["Osiągnięto poziom ",t]})]})]}),s&&e.jsxs("div",{className:"space-y-2",children:[s.badges.length>0&&e.jsx("div",{className:"flex gap-2",children:s.badges.map(r=>e.jsxs("div",{className:"flex items-center gap-2 bg-dark/50 rounded-lg px-3 py-2",children:[e.jsx("span",{children:r.icon}),e.jsx("span",{className:"text-js text-sm",children:r.name})]},r.id))}),s.bonusPoints>0&&e.jsxs("p",{className:"text-emerald-400 text-sm",children:["+",s.bonusPoints," punktów bonusowych"]})]}),e.jsx("button",{onClick:a,className:"mt-4 text-gray-400 hover:text-white text-sm transition-colors",children:"Zamknij"})]})}));ct.displayName="LevelUpNotification";const _s=({title:t,subtitle:s,className:a="",titleClassName:i="",subtitleClassName:r=""})=>e.jsxs("div",{className:`max-w-3xl ${a}`,children:[e.jsx("h2",{className:`text-4xl md:text-5xl font-bold font-space mb-4 ${i}`,children:t}),e.jsx("p",{className:`text-lg md:text-xl ${r}`,children:s})]}),ue=o.forwardRef((c,d)=>{var x=c,{label:t,error:s,icon:a,rightIcon:i,className:r=""}=x,l=C(x,["label","error","icon","rightIcon","className"]);return e.jsxs("div",{className:"space-y-1 sm:space-y-2",children:[e.jsx("label",{className:"block text-xs sm:text-sm font-medium text-gray-400",children:t}),e.jsxs("div",{className:"relative",children:[a&&e.jsx("div",{className:"absolute inset-y-0 left-0 pl-2 sm:pl-3 flex items-center pointer-events-none text-gray-500",children:a}),e.jsx("input",j({ref:d,className:`
-            w-full py-2 sm:py-3 text-sm sm:text-base bg-dark-medium
-            ${a?"pl-8 sm:pl-10":"pl-3 sm:pl-4"} 
-            ${i?"pr-8 sm:pr-10":"pr-3 sm:pr-4"}
-             border rounded-lg border-js/10
-            text-gray-500 placeholder-gray-400
-            focus:outline-none focus:ring-2
-            transition-all duration-200
-            ${s?"border-red-300 focus:border-red-500 focus:ring-red-100":"border-gray-300 focus:border-[#f7df1e] focus:ring-[#f7df1e]/20 hover:border-gray-400"}
-            ${r}
-          `},l)),i&&e.jsx("div",{className:"absolute inset-y-0 right-0 pr-2 sm:pr-3 flex items-center text-gray-500",children:i})]}),s&&e.jsx(n.p,{initial:{opacity:0,y:-10},animate:{opacity:1,y:0},className:"text-xs sm:text-sm text-red-600 mt-1","data-testid":`${l.name}-error`,children:s})]})});ue.displayName="FormInput";const fe=i=>{var r=i,{children:t,className:s}=r,a=C(r,["children","className"]);return e.jsx(n.button,b(j({whileHover:{scale:1.05},whileTap:{scale:.95},className:`px-6 py-2 rounded-lg bg-[#f7df1e] text-black font-bold 
-                hover:bg-[#f7df1e]/90 transition-all shadow-lg 
-                shadow-[#f7df1e]/25 ${s}`},a),{children:t}))},dt=o.memo(({message:t,withIcon:s=!1,variant:a="standard",title:i="Błąd",className:r=""})=>!t||t==="Brak tokenu autoryzacji. Zaloguj się, aby uzyskać dostęp."?null:a==="alert"?e.jsxs(n.div,{initial:{opacity:0},animate:{opacity:1},className:`p-4 bg-red-50 border border-red-200 rounded-lg text-red-700 text-sm flex items-start ${r}`,children:[e.jsx(N,{className:"mr-2 mt-0.5 flex-shrink-0"}),e.jsxs("div",{children:[e.jsx("p",{className:"font-semibold mb-1",children:i}),e.jsx("p",{children:t})]})]}):e.jsxs(n.div,{initial:{opacity:0},animate:{opacity:1},className:`p-3 rounded-lg bg-red-50 border border-red-200 text-red-700 text-sm ${r} ${s?"flex items-center":""}`,children:[s&&e.jsx(N,{className:"mr-2 flex-shrink-0"}),e.jsx("span",{children:t})]}));dt.displayName="FormError";const Os=l=>{var d=l,{isLoading:t,loadingText:s="Przetwarzanie...",children:a,className:i=""}=d,r=C(d,["isLoading","loadingText","children","className"]);return e.jsx("button",b(j({type:"submit",className:`relative w-full bg-[#f7df1e] hover:bg-[#f7df1e]/90 text-black font-semibold py-2 sm:py-3 text-sm sm:text-base rounded-lg transition-all transform hover:scale-[1.02] active:scale-[0.98] disabled:opacity-70 disabled:cursor-not-allowed disabled:hover:scale-100 ${i}`,disabled:t||r.disabled},r),{children:t?e.jsxs("div",{className:"flex items-center justify-center",children:[e.jsxs("svg",{className:"animate-spin -ml-1 mr-2 h-3 w-3 sm:h-4 sm:w-4 text-black",xmlns:"http://www.w3.org/2000/svg",fill:"none",viewBox:"0 0 24 24",children:[e.jsx("circle",{className:"opacity-25",cx:"12",cy:"12",r:"10",stroke:"currentColor",strokeWidth:"4"}),e.jsx("path",{className:"opacity-75",fill:"currentColor",d:"M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"})]}),e.jsx("span",{className:"text-sm sm:text-base",children:s})]}):a}))},Gs=({children:t,onSubmit:s,className:a=""})=>e.jsx("form",{onSubmit:s,className:`space-y-6 ${a}`,children:t}),xt=o.forwardRef((d,l)=>{var c=d,{label:t,placeholder:s="••••••••",error:a,showIcon:i=!0}=c,r=C(c,["label","placeholder","error","showIcon"]);const{isVisible:x,toggleVisibility:m,inputType:h}=st();return e.jsx(ue,j({ref:l,type:h,label:t,placeholder:s,error:a,icon:i?e.jsx(We,{}):void 0,rightIcon:e.jsx("button",{type:"button",onClick:m,className:"bg-white focus:outline-none text-gray-500 hover:text-gray-600 transition-colors","aria-label":x?"Ukryj hasło":"Pokaż hasło",children:x?e.jsx(Ve,{}):e.jsx(Je,{})})},r))});xt.displayName="PasswordInput";const mt=({message:t,title:s="Błąd"})=>t?e.jsxs(n.div,{initial:{opacity:0},animate:{opacity:1},className:"p-4 bg-red-500/10 border border-red-500/20 rounded-lg text-red-400 text-sm flex items-start",children:[e.jsx(N,{className:"mr-2 mt-0.5 flex-shrink-0"}),e.jsxs("div",{children:[e.jsx("p",{className:"font-semibold mb-1",children:s}),e.jsx("p",{children:t})]})]}):null,Zs=({message:t,title:s="Sukces!"})=>t?e.jsxs(n.div,{initial:{opacity:0},animate:{opacity:1},className:"p-4 bg-green-500/10 border border-green-500/20 rounded-lg text-green-400 text-sm flex items-start",children:[e.jsx(_e,{className:"mr-2 mt-0.5 flex-shrink-0"}),e.jsxs("div",{children:[e.jsx("p",{className:"font-semibold mb-1",children:s}),e.jsx("p",{children:t})]})]}):null,g=({children:t,className:s=""})=>e.jsx("div",{className:`max-w-[2000px] mx-auto px-4 sm:px-6 md:px-8 lg:px-20 xl:px-24 2xl:px-48 ${s}`,children:t}),pt=()=>e.jsx(n.div,{whileHover:{scale:1.02},className:"flex items-center z-50 relative",children:e.jsxs(f,{to:"/",className:`text-3xl font-bold font-space text-[#f7df1e] hover:text-[#f7df1e]/90 \r
-                 transition-colors flex items-center gap-2`,children:[e.jsx("span",{className:"text-2xl bg-[#f7df1e] text-black px-2 py-1 rounded",children:"JS"}),e.jsx("span",{children:"CodeLines"})]})}),D=[{href:"#nowosci",label:"Co nowego ?"},{href:"#gra",label:"O grze"},{href:"#dlaczego-warto",label:"Dlaczego warto ?"},{href:"#opinie",label:"Opinie"}],je=o.memo(({href:t,children:s,className:a=""})=>e.jsxs(n.a,{href:t,whileHover:{scale:1.05},className:`relative text-base text-gray-400 hover:text-js font-mono 
-               transition-colors duration-200 py-2 group ${a}`,children:[s,e.jsx("span",{className:`absolute bottom-0 left-0 w-0 h-0.5 bg-js \r
-                    group-hover:w-full transition-all duration-300`})]}));je.displayName="NavLink";const ht=()=>e.jsx("nav",{className:"flex items-center gap-10",children:D.map(({href:t,label:s})=>e.jsx(je,{href:t,children:s},t))}),ut=()=>e.jsx("div",{className:"absolute left-1/2 -translate-x-1/2 hidden xl:block",children:e.jsx(ht,{})}),ft={hidden:{opacity:0,y:20},visible:t=>({opacity:1,y:0,transition:{delay:t*.1}}),exit:{opacity:0,y:20}},jt=({href:t,label:s,index:a,onClick:i})=>e.jsx(n.a,{href:t,custom:a,variants:ft,initial:"hidden",animate:"visible",exit:"exit",whileHover:{scale:1.05},onClick:i,className:"text-2xl text-gray-300 hover:text-js font-mono transition-colors duration-200",children:s}),gt=({onClose:t})=>e.jsxs(n.div,{initial:{opacity:0,y:20},animate:{opacity:1,y:0},exit:{opacity:0,y:20},className:"h-screen flex flex-col items-center justify-center gap-8 px-4",children:[D.map((s,a)=>e.jsx(jt,b(j({},s),{index:a,onClick:t}),s.href)),e.jsx(n.div,{initial:{opacity:0,y:20},animate:{opacity:1,y:0,transition:{delay:D.length*.1}},exit:{opacity:0,y:20},children:e.jsx(f,{to:"/logowanie",children:e.jsx(fe,{onClick:t,className:"mt-4 text-xl px-12",children:"Zaloguj się"})})})]}),bt=({isOpen:t,onClose:s})=>e.jsx(I,{children:t&&e.jsx(n.div,{initial:{opacity:0},animate:{opacity:1},exit:{opacity:0},children:e.jsx(gt,{onClose:s})})}),ge=o.memo(({isOpen:t,onClick:s})=>e.jsxs(n.button,{whileTap:{scale:.95},onClick:s,className:`xl:hidden z-50 w-8 h-8 sm:w-10 sm:h-10 flex items-center justify-center rounded-lg \r
-               border border-js-light bg-dark-medium hover:bg-js-light \r
-               backdrop-blur-sm transition-all relative overflow-hidden`,"aria-label":t?"Zamknij menu":"Otwórz menu",children:[e.jsx(n.div,{animate:{rotate:t?90:0,opacity:t?0:1,scale:t?.5:1},transition:{duration:.2},className:"absolute inset-0 flex items-center justify-center",children:e.jsx(Oe,{className:"w-5 h-5 text-js"})}),e.jsx(n.div,{animate:{rotate:t?0:-90,opacity:t?1:0,scale:t?1:.5},transition:{duration:.2},className:"absolute inset-0 flex items-center justify-center",children:e.jsx(Ge,{className:"w-5 h-5 text-js"})})]}));ge.displayName="MobileMenuButton";const yt=()=>e.jsx(n.a,{href:"https://github.com/ol1mowski/CodeLinesJS",target:"_blank",rel:"noopener noreferrer",whileHover:{scale:1.1,rotate:8},className:"text-gray-400 hover:text-js transition-colors","aria-label":"GitHub Repository",children:e.jsx(R,{className:"w-6 h-6"})}),wt=()=>e.jsx(f,{to:"/logowanie",children:e.jsx(fe,{className:"hidden xl:flex text-sm px-5 py-2.5",children:"Zaloguj się"})}),be=o.memo(({isMobileMenuOpen:t,onMobileMenuToggle:s})=>e.jsxs("div",{className:"flex items-center gap-10",children:[e.jsx(yt,{}),e.jsx(wt,{}),e.jsx(ge,{isOpen:t,onClick:s})]}));be.displayName="HeaderActions";const Ks=()=>{const[t,s]=o.useState(!1),a=()=>s(!t),i=()=>s(!1);return e.jsx(n.header,{initial:{y:-100},animate:{y:0},className:`fixed w-full z-50 transition-all duration-500 
-        bg-dark-medium/90 backdrop-blur-xl border-b border-js-light py-5`,children:e.jsxs(g,{children:[e.jsxs("nav",{className:"flex items-center justify-between relative",children:[e.jsx(pt,{}),e.jsx(ut,{}),e.jsx(be,{isMobileMenuOpen:t,onMobileMenuToggle:a})]}),e.jsx(bt,{isOpen:t,onClose:i})]})})},G=({icon:t,title:s,description:a,delay:i})=>p()?e.jsxs("div",{className:"flex items-start gap-3",children:[e.jsx("div",{className:"w-10 h-10 flex items-center justify-center text-lg bg-[#f7df1e]/10 text-[#f7df1e] rounded-lg shrink-0",children:t}),e.jsxs("div",{children:[e.jsx("h3",{className:"text-lg font-bold text-white mb-1",children:s}),e.jsx("p",{className:"text-gray-400 text-sm",children:a})]})]}):e.jsxs(n.div,{initial:{opacity:0,x:-20},whileInView:{opacity:1,x:0},transition:{delay:i,duration:.5},viewport:{once:!0},className:"flex items-start gap-3",children:[e.jsx("div",{className:"w-10 h-10 flex items-center justify-center text-lg bg-[#f7df1e]/10 text-[#f7df1e] rounded-lg shrink-0",children:t}),e.jsxs("div",{children:[e.jsx("h3",{className:"text-lg font-bold text-white mb-1",children:s}),e.jsx("p",{className:"text-gray-400 text-sm",children:a})]})]}),vt=()=>{const t=p(),s=[{icon:e.jsx(Ze,{}),title:"Interaktywne Lekcje",description:"Ucz się w swoim tempie dzięki przystępnie napisanym lekcjom z przykładami kodu.",delay:.2},{icon:e.jsx(te,{}),title:"Praktyczne Przykłady",description:"Każda koncepcja jest zilustrowana praktycznymi przykładami, które możesz uruchomić i zmodyfikować.",delay:.3},{icon:e.jsx(Ke,{}),title:"Śledzenie Postępów",description:"Monitoruj swoje postępy i wracaj do lekcji, które wymagają powtórzenia.",delay:.4}];return t?e.jsx("div",{className:"space-y-5",children:e.jsx("div",{className:"space-y-4",children:s.map((a,i)=>e.jsx(G,{icon:a.icon,title:a.title,description:a.description,delay:a.delay},i))})}):e.jsx(n.div,{initial:{opacity:0,x:-20},whileInView:{opacity:1,x:0},transition:{duration:.6},viewport:{once:!0},className:"space-y-5",children:e.jsx("div",{className:"space-y-4",children:s.map((a,i)=>e.jsx(G,{icon:a.icon,title:a.title,description:a.description,delay:a.delay},i))})})},P=({position:t="right",title:s,className:a="",showFrame:i=!1,isAbsolute:r=!0,image:l="https://res.cloudinary.com/dbbuav0rj/image/upload/v1742282966/CodeLinesJs/learn_cljs_nttqge.png"})=>{const d=p(),c=t==="right"?50:-50,x=t==="right"?12:-12,m=t==="right"?-10:10,h=t==="right"?{top:{right:"-5px"},bottom:{left:"-5px"}}:{top:{left:"-5px"},bottom:{right:"-5px"}},u=()=>e.jsxs("div",{className:"relative w-[220px] md:w-[240px] lg:w-[260px]",children:[e.jsxs("div",{className:"relative z-10 w-full aspect-[9/19] bg-black rounded-[30px] border-[6px] border-gray-800 shadow-2xl overflow-hidden",style:{transform:r?`rotateX(5deg) rotateY(${m}deg)`:"none",boxShadow:"0 25px 50px -12px rgba(0, 0, 0, 0.5), 0 0 20px rgba(247, 223, 30, 0.2)"},children:[e.jsx("div",{className:"absolute top-0 left-1/2 transform -translate-x-1/2 w-1/3 h-6 bg-black rounded-b-xl z-20"}),e.jsx("div",{className:"w-full h-full bg-gradient-to-br from-[#2a2a2a] to-[#1a1a1a] flex items-center justify-center",children:l?e.jsx("div",{className:"w-full h-full overflow-hidden",children:e.jsx("img",{src:l,alt:s,className:"w-full h-full object-cover"})}):e.jsxs("div",{className:"text-center p-4",children:[e.jsx("div",{className:"w-14 h-14 bg-[#f7df1e] rounded-xl flex items-center justify-center mx-auto mb-3",children:e.jsx("span",{className:"text-black text-xl font-bold",children:"JS"})}),e.jsx("h4",{className:"text-white font-bold mb-1 text-sm",children:s}),e.jsx("p",{className:"text-gray-400 text-xs",children:"Tutaj pojawi się zrzut ekranu z aplikacji"})]})}),e.jsx("div",{className:"absolute top-0 left-0 w-full h-full bg-gradient-to-br from-white/10 to-transparent opacity-30 rounded-[30px] z-20 pointer-events-none"})]}),r&&e.jsxs(e.Fragment,{children:[e.jsx("div",{className:"absolute top-1/3 w-16 h-16 bg-[#f7df1e]/20 rounded-full blur-xl",style:{[h.top.right?"right":"left"]:h.top.right||h.top.left}}),e.jsx("div",{className:"absolute bottom-1/3 w-16 h-16 bg-[#f7df1e]/20 rounded-full blur-xl",style:{[h.bottom.right?"right":"left"]:h.bottom.right||h.bottom.left}})]})]});return i?e.jsx("div",{className:a,children:e.jsx("div",{className:"relative w-[200px]",children:e.jsxs("div",{className:"relative z-10 w-full aspect-[9/19] bg-black rounded-[30px] border-[6px] border-gray-800 shadow-2xl overflow-hidden",children:[e.jsx("div",{className:"absolute top-0 left-1/2 transform -translate-x-1/2 w-1/3 h-6 bg-black rounded-b-xl z-20"}),e.jsx("div",{className:"w-full h-full bg-gradient-to-br from-[#2a2a2a] to-[#1a1a1a] flex items-center justify-center",children:l?e.jsx("div",{className:"w-full h-full overflow-hidden",children:e.jsx("img",{src:l,alt:s,className:"w-full h-full object-cover"})}):e.jsxs("div",{className:"text-center p-4",children:[e.jsx("div",{className:"w-14 h-14 bg-[#f7df1e] rounded-xl flex items-center justify-center mx-auto mb-3",children:e.jsx("span",{className:"text-black text-xl font-bold",children:"JS"})}),e.jsx("h4",{className:"text-white font-bold mb-1 text-sm",children:s}),e.jsx("p",{className:"text-gray-400 text-xs",children:"Tutaj pojawi się zrzut ekranu z aplikacji"})]})})]})})}):d?e.jsx("div",{className:a,style:{perspective:"1000px"},children:e.jsx(u,{})}):e.jsx(n.div,{initial:{opacity:0,x:c,rotate:0},whileInView:{opacity:1,x:0,rotate:x},transition:{duration:.8,ease:"easeOut"},viewport:{once:!0},className:a,style:{perspective:"1000px"},children:e.jsx(u,{})})},ye=({text:t,description:s,href:a,isExternal:i=!1,className:r="",icon:l=e.jsx(Ue,{})})=>{const d=a.startsWith("#"),c=p(),x=()=>e.jsxs(e.Fragment,{children:[e.jsx("span",{children:t}),e.jsx("span",{className:"ml-2",children:l})]}),m=()=>i?e.jsx("a",{href:a,target:"_blank",rel:"noopener noreferrer",className:"inline-flex items-center gap-2 px-6 py-3 bg-[#f7df1e] text-black font-bold rounded-lg hover:bg-[#f7df1e]/90 transition-all shadow-lg shadow-[#f7df1e]/25",children:e.jsx(x,{})}):d?e.jsx("a",{href:a,className:"inline-flex items-center gap-2 px-6 py-3 bg-[#f7df1e] text-black font-bold rounded-lg hover:bg-[#f7df1e]/90 transition-all shadow-lg shadow-[#f7df1e]/25",children:e.jsx(x,{})}):e.jsx(f,{to:a,className:"inline-flex items-center gap-2 px-6 py-3 bg-[#f7df1e] text-black font-bold rounded-lg hover:bg-[#f7df1e]/90 transition-all shadow-lg shadow-[#f7df1e]/25",children:e.jsx(x,{})});return c?e.jsxs("div",{className:`pt-2 ${r}`,children:[m(),s&&e.jsx("p",{className:"mt-3 text-gray-400 text-sm",children:s})]}):e.jsxs(n.div,{initial:{opacity:0,y:20},whileInView:{opacity:1,y:0},transition:{delay:.5,duration:.5},viewport:{once:!0},className:`pt-2 ${r}`,children:[m(),s&&e.jsx("p",{className:"mt-3 text-gray-400 text-sm",children:s})]})},Nt=()=>e.jsx("div",{className:"w-full max-w-5xl mx-auto px-4",children:e.jsxs("div",{className:"relative pt-16 md:pt-24 lg:pt-20",children:[e.jsx(P,{position:"right",title:"Teoria JavaScript",className:"absolute -right-4 top-0 md:top-[-20px] md:right-0 lg:right-10 xl:right-20 z-10 hidden md:block"}),e.jsxs("div",{className:"md:max-w-[60%] lg:max-w-[55%] space-y-6 py-6",children:[e.jsx(P,{title:"Teoria JavaScript",className:"flex justify-center mb-8 md:hidden",showFrame:!0,isAbsolute:!1}),e.jsx(vt,{}),e.jsx(ye,{href:"#practice",text:"Przejdź do Praktyki",description:"Gotowy na wyzwania? Sprawdź swoje umiejętności w praktyce!"})]})]})}),z=({children:t,motionProps:s,className:a=""})=>p()?e.jsx("div",{className:a,children:t}):e.jsx(n.div,b(j({className:a},s),{children:t})),zt=({children:t,motionProps:s,className:a=""})=>p()?e.jsx("h2",{className:a,children:t}):e.jsx(n.h2,b(j({className:a},s),{children:t})),kt=({children:t,motionProps:s,className:a=""})=>p()?e.jsx("p",{className:a,children:t}):e.jsx(n.p,b(j({className:a},s),{children:t})),we=({badge:t,title:s,subtitle:a,id:i,className:r="",titleClassName:l="",subtitleClassName:d=""})=>{const c=p();return e.jsxs("div",{className:`w-full flex flex-col items-center ${r}`,children:[t&&e.jsx(z,{className:"mb-2",motionProps:{initial:{opacity:0,y:-20},whileInView:{opacity:1,y:0},transition:{duration:.5},viewport:{once:!0}},children:e.jsx("div",{className:"px-4 py-1.5 bg-[#f7df1e]/10 border border-[#f7df1e]/20 rounded-full text-[#f7df1e] font-semibold text-sm inline-block",children:t})}),e.jsxs(z,{className:"max-w-3xl text-center px-4",motionProps:{initial:{opacity:0,y:20},whileInView:{opacity:1,y:0},transition:{duration:.5,delay:c?0:.1},viewport:{once:!0}},children:[e.jsx("h2",{id:i,className:`text-4xl md:text-5xl font-bold font-space mb-4 text-[#f7df1e] drop-shadow-lg ${l}`,children:s}),e.jsx("p",{className:`text-lg md:text-xl text-gray-400 ${d}`,children:a})]})]})},Us=()=>e.jsx("section",{id:"theory",className:"min-h-screen w-full bg-gradient-to-b from-[#1a1a1a] via-[#242424] to-[#1a1a1a] py-12 md:py-16 relative overflow-hidden","aria-labelledby":"theory-title",children:e.jsx(g,{className:"relative z-10",children:e.jsxs("div",{className:"flex flex-col items-center gap-8 md:gap-12",children:[e.jsx(we,{badge:"Krok 1: Teoria",title:"Zbuduj Solidne Fundamenty",subtitle:"Rozpocznij swoją przygodę z JavaScript od zrozumienia kluczowych koncepcji",id:"theory-title"}),e.jsx(Nt,{})]})})}),Z=({icon:t,title:s,description:a,delay:i})=>p()?e.jsxs("div",{className:"flex items-start gap-3",children:[e.jsx("div",{className:"w-10 h-10 flex items-center justify-center text-lg bg-[#f7df1e]/10 text-[#f7df1e] rounded-lg shrink-0",children:t}),e.jsxs("div",{children:[e.jsx("h3",{className:"text-lg font-bold text-white mb-1",children:s}),e.jsx("p",{className:"text-gray-400 text-sm",children:a})]})]}):e.jsxs(n.div,{initial:{opacity:0,x:20},whileInView:{opacity:1,x:0},transition:{delay:i,duration:.5},viewport:{once:!0},className:"flex items-start gap-3",children:[e.jsx("div",{className:"w-10 h-10 flex items-center justify-center text-lg bg-[#f7df1e]/10 text-[#f7df1e] rounded-lg shrink-0",children:t}),e.jsxs("div",{children:[e.jsx("h3",{className:"text-lg font-bold text-white mb-1",children:s}),e.jsx("p",{className:"text-gray-400 text-sm",children:a})]})]}),Ct=()=>{const t=p(),s=[{icon:e.jsx(M,{}),title:"Interaktywne Gry",description:"Ucz się JavaScript poprzez wciągające gry, które sprawdzą Twoją wiedzę w praktyce.",delay:.2},{icon:e.jsx(se,{}),title:"Wyzwania Kodowania",description:"Rozwiązuj praktyczne problemy programistyczne o różnym poziomie trudności.",delay:.3},{icon:e.jsx(ae,{}),title:"Rywalizacja z Innymi",description:"Porównuj swoje wyniki z innymi uczestnikami i wspólnie się rozwijajcie.",delay:.4}];return t?e.jsx("div",{className:"space-y-5",children:e.jsx("div",{className:"space-y-4",children:s.map((a,i)=>e.jsx(Z,{icon:a.icon,title:a.title,description:a.description,delay:a.delay},i))})}):e.jsx(n.div,{initial:{opacity:0,x:20},whileInView:{opacity:1,x:0},transition:{duration:.6},viewport:{once:!0},className:"space-y-5",children:e.jsx("div",{className:"space-y-4",children:s.map((a,i)=>e.jsx(Z,{icon:a.icon,title:a.title,description:a.description,delay:a.delay},i))})})},St=()=>e.jsx("div",{className:"w-full max-w-5xl mx-auto px-4",children:e.jsxs("div",{className:"relative pt-16 md:pt-24 lg:pt-20",children:[e.jsx(P,{position:"left",image:"https://res.cloudinary.com/dbbuav0rj/image/upload/v1742282966/CodeLinesJs/game_cljs_snynmf.png",title:"Praktyka JavaScript",className:"absolute -left-4 top-0 md:top-[-20px] md:left-0 lg:left-10 xl:left-20 z-10 hidden md:block"}),e.jsxs("div",{className:"md:ml-auto md:max-w-[60%] lg:max-w-[55%] space-y-6 py-6",children:[e.jsx(P,{title:"Praktyka JavaScript",className:"flex justify-center mb-8 md:hidden",showFrame:!0,isAbsolute:!1}),e.jsx(Ct,{}),e.jsx(ye,{href:"#challenges",text:"Rozpocznij Wyzwania",description:"Gotowy do działania? Sprawdź swoje umiejętności w praktycznych wyzwaniach!"})]})]})}),Ys=()=>e.jsx("section",{id:"practice",className:"min-h-screen w-full bg-gradient-to-b from-[#1a1a1a] via-[#242424] to-[#1a1a1a] py-12 md:py-16 relative overflow-hidden","aria-labelledby":"practice-title",children:e.jsx(g,{className:"relative z-10",children:e.jsxs("div",{className:"flex flex-col items-center gap-8 md:gap-12",children:[e.jsx(we,{badge:"Krok 2: Praktyka",title:"Zastosuj Wiedzę w Praktyce",subtitle:"Sprawdź swoje umiejętności w interaktywnych grach i wyzwaniach",id:"practice-title"}),e.jsx(St,{})]})})}),Mt=200,K=.8,y={comment:"text-[#6A9955]",function:"text-[#569CD6]",constant:"text-[#4EC9B0]",return:"text-[#C586C0]",arrow:"text-[#DCDCAA]",string:"text-[#CE9178]",console:"text-[#DCDCAA]",default:"text-[#9CDCFE]"},$=[{line:"// Witaj w CodeLinesJS! 🚀",indent:0},{line:"",indent:0},{line:"function calculateScore(challenges) {",indent:0},{line:"const points = challenges.reduce((total, challenge) => {",indent:1},{line:"return total + challenge.difficulty * 100;",indent:2},{line:"}, 0);",indent:1},{line:"",indent:0},{line:"const bonus = points > 1000 ? points * 0.1 : 0;",indent:1},{line:"return points + bonus;",indent:1},{line:"}",indent:0},{line:"",indent:0},{line:"// Oblicz wynik gracza",indent:0},{line:"const playerScore = calculateScore(completedChallenges);",indent:0},{line:"console.log(`Twój wynik: ${playerScore} punktów! 🎮`);",indent:0}],Tt=()=>{const[t,s]=o.useState(0),[a,i]=o.useState([]),[r,l]=o.useState(!1),d=p(),c=o.useMemo(()=>d?$.filter((m,h)=>h%2===0||h<3):$,[d]),x=o.useCallback(()=>{r&&(s(0),i([]),l(!1))},[r]);return o.useEffect(()=>{if(d){i(c),s(c.length),l(!0);return}const h=setInterval(()=>{t<c.length?(i(u=>[...u,c[t]]),s(u=>u+1)):(clearInterval(h),l(!0))},Mt);return()=>clearInterval(h)},[t,c,d]),{visibleLines:a,currentLineIndex:t,isAnimationComplete:r,resetAnimation:x}},Pt=({number:t})=>e.jsxs(e.Fragment,{children:[e.jsx("span",{className:"text-gray-500 select-none w-8 inline-block",children:t}),e.jsx("span",{className:"text-gray-400 select-none",children:"│ "})]}),At=t=>t.startsWith("//")?y.comment:t.startsWith("function")?y.function:t.startsWith("const")?y.constant:t.includes("return")?y.return:t.includes("=>")?y.arrow:t.includes("`")?y.string:t.includes("console.log")?y.console:y.default,It=({code:t})=>e.jsx(n.span,{initial:{opacity:0},animate:{opacity:1},transition:{duration:.2},style:{paddingLeft:`${t.indent*20}px`},className:At(t.line),children:t.line}),Ft=()=>{const t=p(),s=t?K*1.5:K;return t?e.jsx("span",{className:"absolute bottom-0 ml-1 h-[2px] w-3 bg-[#f7df1e]"}):e.jsx(n.span,{initial:{opacity:0},animate:{opacity:[0,1,0]},transition:{duration:s,repeat:1/0},className:"absolute bottom-0 ml-1 h-[2px] w-3 bg-[#f7df1e]"})},Et=({lineNumber:t,code:s,showCursor:a})=>{const i=p();return e.jsxs(z,{className:"whitespace-pre relative",motionProps:{initial:{opacity:0},animate:{opacity:1},transition:{duration:.2}},children:[e.jsx(Pt,{number:t}),e.jsx(It,{code:s}),a&&!i&&e.jsx(Ft,{})]})},Lt=()=>{const{visibleLines:t,currentLineIndex:s}=Tt();return e.jsx("div",{className:"p-6 font-mono text-sm md:text-base h-full overflow-y-auto custom-scrollbar",children:e.jsx("pre",{className:"flex flex-col",children:t.map((a,i)=>e.jsx(Et,{lineNumber:i+1,code:a,showCursor:i===t.length-1&&s<$.length},i))})})},Bt=()=>{const t=p();return e.jsx(z,{className:"w-full xl:w-1/2 relative px-4 md:px-0 h-[45vh] flex items-center",motionProps:{initial:{opacity:0,scale:.9},whileInView:{opacity:1,scale:1},viewport:{once:!0},transition:{duration:t?.3:.8}},children:e.jsxs("div",{className:"rounded-xl border border-[#f7df1e]/20 bg-[#1E1E1E] overflow-hidden shadow-2xl w-full h-full flex flex-col",children:[e.jsxs("div",{className:"bg-[#2D2D2D] px-4 py-3 flex items-center gap-2 border-b border-black/50",children:[e.jsxs("div",{className:"flex gap-2",children:[e.jsx("div",{className:"w-3 h-3 rounded-full bg-[#FF5F56]"}),e.jsx("div",{className:"w-3 h-3 rounded-full bg-[#FFBD2E]"}),e.jsx("div",{className:"w-3 h-3 rounded-full bg-[#27C93F]"})]}),e.jsx("span",{className:"text-gray-400 text-sm ml-2",children:"script.js"})]}),e.jsx("div",{className:"flex-1 overflow-hidden",children:e.jsx(Lt,{})})]})})},Dt=()=>e.jsx(zt,{className:"text-4xl md:text-5xl lg:text-6xl font-bold font-space text-[#f7df1e] mb-6",motionProps:{initial:{opacity:0,y:20},whileInView:{opacity:1,y:0},viewport:{once:!0}},children:"Wejdź do Świata JavaScript"}),$t=()=>e.jsx(kt,{className:"w-full m-auto text-lg text-center md:text-xl xl:ml-0 lg:text-left text-gray-300 font-inter mb-8 max-w-2xl",motionProps:{initial:{opacity:0,y:20},whileInView:{opacity:1,y:0},viewport:{once:!0},transition:{delay:.1}},children:"Podejmij wyzwanie i zostań mistrzem JavaScript poprzez wciągającą grę. Rozwiązuj zagadki, zdobywaj osiągnięcia i rywalizuj z innymi graczami."}),Ht=()=>{const t=p();return e.jsxs(z,{className:"flex gap-4 flex-wrap justify-center xl:justify-start mb-12",motionProps:{initial:{opacity:0,y:20},whileInView:{opacity:1,y:0},viewport:{once:!0},transition:{delay:.2}},children:[e.jsx(Rt,{isMobile:t}),e.jsx(Wt,{isMobile:t})]})},Rt=({isMobile:t})=>{const s=()=>e.jsxs(e.Fragment,{children:[e.jsx(Ye,{className:"text-lg sm:text-xl"}),"Zacznij Grę"]});return e.jsx(f,{to:"/dashboard",children:t?e.jsx("button",{className:"px-4 sm:px-8 py-3 sm:py-4 rounded-xl bg-[#f7df1e] text-black font-bold font-space flex items-center justify-center gap-2 hover:bg-[#f7df1e]/90 transition-all shadow-lg shadow-[#f7df1e]/25 text-sm sm:text-base",children:e.jsx(s,{})}):e.jsx(n.button,{whileHover:{scale:1.05},whileTap:{scale:.95},className:"px-4 sm:px-8 py-3 sm:py-4 rounded-xl bg-[#f7df1e] text-black font-bold font-space flex items-center justify-center gap-2 hover:bg-[#f7df1e]/90 transition-all shadow-lg shadow-[#f7df1e]/25 text-sm sm:text-base",children:e.jsx(s,{})})})},Wt=({isMobile:t})=>{const s=()=>e.jsxs(e.Fragment,{children:[e.jsx(te,{className:"text-lg sm:text-xl"}),"Dowiedz się więcej"]});return e.jsx("a",{href:"#gra",children:t?e.jsx("button",{className:"px-4 sm:px-8 py-3 sm:py-4 rounded-xl border-2 border-[#f7df1e]/50 text-[#f7df1e] font-bold font-space flex items-center justify-center gap-2 hover:bg-[#f7df1e]/10 transition-all text-sm sm:text-base",children:e.jsx(s,{})}):e.jsx(n.button,{whileHover:{scale:1.05},whileTap:{scale:.95},className:"px-4 sm:px-8 py-3 sm:py-4 rounded-xl border-2 border-[#f7df1e]/50 text-[#f7df1e] font-bold font-space flex items-center justify-center gap-2 hover:bg-[#f7df1e]/10 transition-all text-sm sm:text-base",children:e.jsx(s,{})})})},Vt=()=>B(void 0,null,function*(){try{const t=yield fetch(`${oe}stats/general`);if(!t.ok)throw new Error("Błąd podczas pobierania statystyk");const s=yield t.json();return s.data||s}catch(t){throw console.error("Błąd podczas pobierania statystyk:",t),t}}),Jt=["generalStats"],_t=()=>{const{data:t,isLoading:s,error:a}=ie({queryKey:Jt,queryFn:Vt,staleTime:3e5,refetchOnWindowFocus:!0});return{stats:t?[{value:t.lessons.value,label:t.lessons.label},{value:t.games.value,label:t.games.label},{value:t.users.value,label:t.users.label}]:[],isLoading:s,error:a,data:t}},ve=o.memo(()=>{const{stats:t,isLoading:s}=_t();return e.jsx("div",{className:"grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-4",children:t.map((a,i)=>e.jsx(Ne,b(j({},a),{index:i,isLoading:s}),a.label))})}),Ne=o.memo(({value:t,label:s,index:a,isLoading:i})=>{const r=p();return e.jsx(z,{className:"text-center p-4 rounded-lg bg-black/20 border border-[#f7df1e]/10",motionProps:{initial:{opacity:0,y:20},whileInView:{opacity:1,y:0},viewport:{once:!0},transition:{duration:.5,delay:r?0:a*.1}},children:i?e.jsx("span",{className:"animate-pulse",children:"..."}):e.jsxs(e.Fragment,{children:[e.jsx("p",{className:"text-3xl md:text-4xl font-bold font-space text-[#f7df1e]",children:t}),e.jsx("p",{className:"text-sm md:text-base text-gray-400 font-inter",children:s})]})})});ve.displayName="StatsGrid";Ne.displayName="StatItem";const ze=o.memo(()=>e.jsx("div",{className:"w-screen xl:w-1/2 px-4 md:px-0",children:e.jsxs(n.div,{className:"text-center xl:text-left",children:[e.jsx(Dt,{}),e.jsx($t,{}),e.jsx(Ht,{}),e.jsx(ve,{})]})}));ze.displayName="HeroContent";const qs=()=>e.jsx("section",{className:"relative min-h-screen pt-16 w-full bg-gradient-to-b from-dark via-dark/90 to-dark overflow-hidden flex items-center",children:e.jsx(g,{className:"relative w-full z-10 h-full py-8 pt-24 md:pt-32",children:e.jsxs("div",{className:"flex flex-col xl:flex-row items-center justify-center h-full gap-8 md:gap-12 xl:gap-16",children:[e.jsx(ze,{}),e.jsx(Bt,{})]})})}),ke=o.memo(()=>p()?e.jsxs("div",{className:"lg:col-span-2",children:[e.jsx("h3",{className:"text-xl font-bold text-[#f7df1e] mb-4",children:"CodeLinesJs"}),e.jsx("p",{className:"text-gray-400 mb-6 leading-relaxed",children:"Platforma do nauki JavaScript stworzona z myślą o społeczności programistów. Dołącz do nas i rozwijaj swoje umiejętności w przyjaznym środowisku."}),e.jsxs("div",{className:"flex items-center gap-4",children:[e.jsx(f,{to:"/dashboard",children:e.jsx("button",{className:`px-6 py-2 bg-[#f7df1e] text-black font-medium rounded-lg \r
-                       hover:bg-[#f7df1e]/90 transition-colors`,children:"Rozpocznij Naukę"})}),e.jsx(f,{to:"/logowanie",children:e.jsx("button",{className:`px-6 py-2 border border-[#f7df1e]/20 text-[#f7df1e] font-medium rounded-lg\r
-                       hover:bg-[#f7df1e]/10 transition-colors`,children:"Dołącz do Nas"})})]})]}):e.jsxs(n.div,{initial:{opacity:0,y:20},whileInView:{opacity:1,y:0},viewport:{once:!0},className:"lg:col-span-2",children:[e.jsx("h3",{className:"text-xl font-bold text-[#f7df1e] mb-4",children:"CodeLinesJs"}),e.jsx("p",{className:"text-gray-400 mb-6 leading-relaxed",children:"Platforma do nauki JavaScript stworzona z myślą o społeczności programistów. Dołącz do nas i rozwijaj swoje umiejętności w przyjaznym środowisku."}),e.jsxs("div",{className:"flex items-center gap-4",children:[e.jsx(f,{to:"/dashboard",children:e.jsx(n.button,{whileHover:{scale:1.05},whileTap:{scale:.95},className:`px-6 py-2 bg-[#f7df1e] text-black font-medium rounded-lg \r
-                   hover:bg-[#f7df1e]/90 transition-colors`,children:"Rozpocznij Naukę"})}),e.jsx(f,{to:"/logowanie",children:e.jsx(n.button,{whileHover:{scale:1.05},whileTap:{scale:.95},className:`px-6 py-2 border border-[#f7df1e]/20 text-[#f7df1e] font-medium rounded-lg\r
-                   hover:bg-[#f7df1e]/10 transition-colors`,children:"Dołącz do Nas"})})]})]}));ke.displayName="FooterIntro";const H=o.memo(({title:t,links:s,delay:a=0})=>p()?e.jsxs("div",{children:[e.jsx("h4",{className:"text-lg font-bold text-[#f7df1e] mb-4",children:t}),e.jsx("ul",{className:"space-y-3",children:s.map(r=>e.jsx("li",{children:e.jsxs("a",{href:r.href,className:"flex items-center gap-3 text-gray-400 hover:text-[#f7df1e] transition-colors",children:[e.jsx(r.icon,{className:"w-5 h-5"}),r.label]})},r.label))})]}):e.jsxs(n.div,{initial:{opacity:0,y:20},whileInView:{opacity:1,y:0},viewport:{once:!0},transition:{delay:a},children:[e.jsx("h4",{className:"text-lg font-bold text-[#f7df1e] mb-4",children:t}),e.jsx("ul",{className:"space-y-3",children:s.map(r=>e.jsx(n.li,{whileHover:{x:5},children:e.jsxs("a",{href:r.href,className:"flex items-center gap-3 text-gray-400 hover:text-[#f7df1e] transition-colors",children:[e.jsx(r.icon,{className:"w-5 h-5"}),r.label]})},r.label))})]}));H.displayName="FooterLinks";const U={resources:[{label:"Polityka prywatności",href:"/polityka-prywatnosci",icon:qe}],social:[{label:"GitHub",href:"https://github.com/ol1mowski/CodeLinesJS",icon:R},{label:"YouTube",href:"https://www.youtube.com/@oliwiermarkiewicz",icon:re}]},Ot=()=>e.jsxs("div",{className:"grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-12",children:[e.jsx(ke,{}),e.jsx(H,{title:"Prywatność",links:U.resources,delay:.2}),e.jsx(H,{title:"Social Media",links:U.social,delay:.3})]}),Gt=()=>e.jsxs("div",{className:"absolute inset-0",children:[e.jsx("div",{className:"absolute inset-0 bg-[#1a1a1a] opacity-95"}),e.jsx("div",{className:"absolute left-0 top-0 w-1/4 h-full bg-[#f7df1e] opacity-30",style:{clipPath:"polygon(0 0, 80% 0, 60% 100%, 0 100%)"}}),e.jsx("div",{className:"absolute right-0 bottom-0 w-1/5 h-1/2 bg-[#f7df1e] opacity-25",style:{clipPath:"polygon(40% 0, 100% 0, 100% 100%, 0 100%)"}}),e.jsx("div",{className:"absolute top-1/4 right-1/4 w-6 h-6 bg-[#f7df1e] opacity-20",style:{clipPath:"polygon(50% 0%, 0% 100%, 100% 100%)"}}),e.jsx("div",{className:"absolute bottom-1/3 left-1/3 w-4 h-4 bg-[#f7df1e] opacity-20",style:{clipPath:"polygon(50% 0%, 0% 100%, 100% 100%)"}}),e.jsx("div",{className:"absolute top-1/2 right-1/3 w-3 h-3 bg-[#f7df1e] opacity-20",style:{clipPath:"polygon(50% 0%, 0% 100%, 100% 100%)"}}),e.jsx("div",{className:"absolute -top-40 -left-40 w-96 h-96 bg-[#f7df1e] rounded-full blur-[150px] opacity-10"}),e.jsx("div",{className:"absolute -bottom-40 -right-40 w-96 h-96 bg-[#f7df1e] rounded-full blur-[150px] opacity-10"}),e.jsx("div",{className:"absolute top-1/4 left-1/4 w-32 h-32 bg-[#f7df1e] rounded-full blur-[100px] opacity-5"}),e.jsx("div",{className:"absolute bottom-1/4 right-1/4 w-32 h-32 bg-[#f7df1e] rounded-full blur-[100px] opacity-5"}),e.jsx("div",{className:"absolute top-0 right-0 w-64 h-64 opacity-20 pointer-events-none",style:{background:"radial-gradient(circle at center, rgba(26, 26, 26, 0.8) 0%, transparent 70%)",filter:"blur(20px)"}}),e.jsx("div",{className:"absolute inset-0 opacity-[0.01]",style:{backgroundImage:`url("data:image/svg+xml,%3Csvg width='60' height='60' viewBox='0 0 60 60' xmlns='http://www.w3.org/2000/svg'%3E%3Cg fill='none' fill-rule='evenodd'%3E%3Cg fill='%23f7df1e' fill-opacity='1'%3E%3Cpath d='M36 34v-4h-2v4h-4v2h4v4h2v-4h4v-2h-4zm0-30V0h-2v4h-4v2h4v4h2V6h4V4h-4zM6 34v-4H4v4H0v2h4v4h2v-4h4v-2H6zM6 4V0H4v4H0v2h4v4h2V6h4V4H6z'/%3E%3C/g%3E%3C/g%3E%3C/svg%3E")`,backgroundSize:"30px 30px"}})]}),Zt=()=>e.jsxs(n.p,{initial:{opacity:0},whileInView:{opacity:1},viewport:{once:!0},className:"text-sm text-gray-400",children:["© ",new Date().getFullYear()," Aplikację wykonał"," ",e.jsx("a",{href:"https://oliwiermarkiewicz.pl/",target:"_blank",rel:"noopener noreferrer",className:"text-[#f7df1e] hover:text-[#f7df1e]/80",children:"Oliwier Markiewicz"}),". ",e.jsx("span",{children:"Wszelkie prawa zastrzeżone."})]}),Ce=o.memo(({links:t})=>p()?e.jsx("div",{className:"flex items-center gap-3",children:t.map(a=>e.jsx("a",{href:a.href,target:"_blank",rel:"noopener noreferrer",className:`p-2 rounded-lg ${a.bgColor} ${a.textColor} 
-                       ${a.hoverBg} transition-all duration-300 group`,children:e.jsx(a.icon,{className:`w-5 h-5 transition-colors duration-300 ${a.iconColor}`})},a.href))}):e.jsx("div",{className:"flex items-center gap-3",children:t.map((a,i)=>e.jsx(n.a,{href:a.href,target:"_blank",rel:"noopener noreferrer",initial:{opacity:0,scale:.5},whileInView:{opacity:1,scale:1},whileHover:{scale:1.1},whileTap:{scale:.9},transition:{delay:i*.1},viewport:{once:!0},className:`p-2 rounded-lg ${a.bgColor} ${a.textColor} 
-                     ${a.hoverBg} transition-all duration-300 group`,children:e.jsx(a.icon,{className:`w-5 h-5 transition-colors duration-300 ${a.iconColor}`})},a.href))}));Ce.displayName="SocialLinks";const Kt=[{icon:R,href:"https://github.com/ol1mowski",bgColor:"bg-[#333]/20",textColor:"text-white",hoverBg:"hover:bg-[#333]/20",iconColor:"group-hover:text-white"},{icon:re,href:"https://youtube.com/@oliwier.markiewicz",bgColor:"bg-[#FF0000]/10",textColor:"text-[#FF0000]",hoverBg:"hover:bg-[#FF0000]/20",iconColor:"group-hover:text-white"},{icon:ne,href:"https://www.linkedin.com/in/oliwier-markiewicz-47857228a/",bgColor:"bg-[#0077B5]/10",textColor:"text-[#0077B5]",hoverBg:"hover:bg-[#0077B5]/20",iconColor:"group-hover:text-white"}],Qs=()=>e.jsxs("footer",{className:"w-full relative py-16 md:py-24 bg-gradient-to-b from-[#242424] via-[#242424] to-[#242424] overflow-hidden",children:[e.jsx(Gt,{}),e.jsx(g,{className:"relative z-10",children:e.jsxs("div",{className:"flex flex-col gap-16",children:[e.jsx(Ot,{}),e.jsx("div",{className:"pt-8 border-t border-gray-800",children:e.jsxs("div",{className:"flex flex-col md:flex-row items-center justify-between gap-4",children:[e.jsx(Zt,{}),e.jsx(Ce,{links:Kt})]})})]})})]}),W=()=>o.useMemo(()=>({header:{initial:{opacity:0,y:-20},animate:{opacity:1,y:0},transition:{duration:.5}},testimonial:{initial:{opacity:0,y:30},animate:{opacity:1,y:0},transition:t=>({duration:.5,delay:t*.2})},cta:{initial:{opacity:0,y:30},animate:{opacity:1,y:0},transition:{duration:.5,delay:.4}}}),[]),Se=o.memo(({content:t,author:s,role:a,avatar:i,index:r,linkedIn:l})=>{const d=p(),c=W(),x=()=>e.jsx("div",{className:"flex mt-2",children:Array(5).fill(0).map((h,u)=>e.jsx(Xe,{className:"text-js"},u))}),m=()=>e.jsxs("div",{className:"relative rounded-xl p-6 md:p-8 bg-dark-light/30 backdrop-blur-sm border border-js/10 h-full flex flex-col",children:[e.jsxs("div",{className:"flex items-start mb-6",children:[e.jsx("div",{className:"shrink-0 mr-4",children:e.jsx("div",{className:"w-16 h-16 rounded-full overflow-hidden border-2 border-js",children:e.jsx("img",{src:i,alt:s,className:"w-full h-full object-cover"})})}),e.jsxs("div",{children:[e.jsxs("div",{className:"flex items-center",children:[e.jsx("h3",{className:"text-lg font-bold text-white",children:s}),l&&e.jsx("a",{href:l,target:"_blank",rel:"noopener noreferrer",className:"ml-2 text-blue-400 hover:text-blue-300 transition-colors","aria-label":`Profil LinkedIn ${s}`,children:e.jsx(ne,{})})]}),e.jsx("p",{className:"text-gray-400 text-sm",children:a}),e.jsx(x,{})]})]}),e.jsxs("div",{className:"relative mb-6 flex-grow",children:[e.jsx(Qe,{className:"absolute -top-2 -left-1 text-js/20 text-4xl","aria-hidden":"true"}),e.jsx("p",{className:"text-gray-300 italic pl-6 relative z-10",children:t})]}),e.jsx("div",{className:"w-1/4 h-1 bg-js/30 rounded-full","aria-hidden":"true"})]});return d?e.jsx(m,{}):e.jsx(n.div,{initial:c.testimonial.initial,whileInView:c.testimonial.animate,transition:c.testimonial.transition(r),viewport:{once:!0},children:e.jsx(m,{})})});Se.displayName="TestimonialCard";const Me=o.memo(({title:t,subtitle:s})=>{const a=W();return e.jsx("div",{className:"text-center mb-12 relative z-10",children:e.jsxs(n.div,{initial:a.header.initial,whileInView:a.header.animate,transition:a.header.transition,viewport:{once:!0},children:[e.jsx("span",{className:"text-js font-semibold uppercase tracking-wider",children:"Opinie uczestników"}),e.jsx("h2",{className:"text-3xl md:text-4xl font-bold text-white mt-2 mb-4",children:t}),s&&e.jsx("p",{className:"text-gray-400 max-w-2xl mx-auto",children:s})]})})});Me.displayName="SectionHeader";const Te=o.memo(({text:t,url:s})=>{const a=W();return e.jsx(n.div,{initial:a.cta.initial,whileInView:a.cta.animate,transition:a.cta.transition,viewport:{once:!0},className:"mt-16 text-center",children:e.jsx(f,{to:s,className:"inline-flex items-center justify-center px-8 py-3 rounded-lg bg-js text-dark font-semibold hover:bg-js/90 transition-all duration-300 transform hover:scale-105",children:t})})});Te.displayName="CallToAction";const Ut=()=>o.useMemo(()=>[{content:"Jako właściciel serwisu chciałbym podkreślić, że nasza platforma to doskonałe miejsce do nauki teorii oraz praktycznego utrwalania wiedzy poprzez angażujące gry edukacyjne. Użytkownicy mogą również śledzić swoje postępy, co motywuje do dalszej nauki. Nieustannie rozwijamy nasz serwis i już wkrótce wprowadzimy nowe funkcje oparte na sztucznej inteligencji, które jeszcze bardziej usprawnią naukę JavaScriptu!",author:"Oliwier Markiewicz",role:"Twórca CodeLinesJS",avatar:"https://res.cloudinary.com/dbbuav0rj/image/upload/v1731069658/Portfolio/images/me_gmqeii.webp",linkedIn:"https://www.linkedin.com/in/oliwier-markiewicz-47857228a/"}],[]),Yt=o.memo(()=>{const t=Ut();return e.jsx("section",{id:"opinie",className:"py-20 overflow-hidden bg-gradient-to-b from-dark via-dark-medium to-dark relative",children:e.jsxs(g,{children:[e.jsx(Me,{title:"Co mówią nasi użytkownicy"}),e.jsx("div",{className:"grid grid-cols-1 md:grid-cols-auto gap-8",children:t.map((s,a)=>e.jsx(Se,{content:s.content,author:s.author,role:s.role,avatar:s.avatar,index:a,linkedIn:s.linkedIn},s.author))}),e.jsx(Te,{text:"Dołącz do nas",url:"/logowanie"})]})})});Yt.displayName="TestimonialsSection";const qt=({icon:t,title:s,subtitle:a,description:i,delay:r})=>{const l=p(),d=e.jsx("div",{className:"bg-black/20 border border-[#f7df1e]/10 rounded-xl p-6 hover:border-[#f7df1e]/20 transition-all duration-300 hover:bg-black/30 group",children:e.jsxs("div",{className:"flex items-start gap-4",children:[e.jsx("div",{className:"w-12 h-12 bg-[#f7df1e]/10 rounded-lg flex items-center justify-center text-[#f7df1e] text-xl shrink-0 group-hover:bg-[#f7df1e]/20 transition-all duration-300",children:t}),e.jsxs("div",{className:"flex-1",children:[e.jsx("h3",{className:"text-lg md:text-xl font-bold text-[#f7df1e] mb-1 font-space",children:s}),e.jsx("p",{className:"text-sm text-gray-300 mb-3 font-medium",children:a}),e.jsx("p",{className:"text-gray-400 leading-relaxed text-sm",children:i})]})]})});return l?d:e.jsx(n.div,{initial:{opacity:0,y:20},whileInView:{opacity:1,y:0},transition:{duration:.5,delay:r},viewport:{once:!0},className:"h-full",children:d})},Qt=()=>{const t=p(),s=[{icon:e.jsx(M,{}),title:"Interaktywna Rozgrywka",subtitle:"Prawdziwe wyzwania programistyczne",description:"Rozwiązuj rzeczywiste problemy programistyczne w formie gry. Każdy poziom to nowa wyzwanie, które rozwija Twoje umiejętności.",delay:.1},{icon:e.jsx(ae,{}),title:"Społeczność",subtitle:"Ucz się razem z innymi",description:"Dołącz do społeczności programistów, dziel się rozwiązaniami i ucz się od najlepszych.",delay:.2},{icon:e.jsx(se,{}),title:"Postęp",subtitle:"Śledź swoje osiągnięcia",description:"System osiągnięć i rankingów motywuje do ciągłego rozwoju i doskonalenia umiejętności.",delay:.3}];return e.jsx("div",{className:"space-y-6",children:s.map((a,i)=>e.jsx(qt,{icon:a.icon,title:a.title,subtitle:a.subtitle,description:a.description,delay:t?0:a.delay},i))})},Xt=()=>e.jsx("div",{className:"w-full max-w-7xl mx-auto px-4",children:e.jsxs("div",{className:"grid grid-cols-1 lg:grid-cols-2 gap-8 lg:gap-16 items-start mb-16",children:[e.jsxs("div",{className:"space-y-6 text-left",children:[e.jsx("h2",{className:"text-4xl md:text-5xl font-bold text-[#f7df1e] font-space text-left mb-4",children:"O Grze"}),e.jsx("p",{className:"text-lg md:text-xl text-gray-400 leading-relaxed text-left mb-8",children:"Zanurz się w fascynującym świecie programowania z naszą interaktywną grą JavaScript! Ucz się, koduj i rozwiązuj wyzwania w dynamicznym środowisku, które łączy zabawę z nauką."}),e.jsx("h3",{className:"text-2xl md:text-3xl font-bold text-white font-space text-left",children:"Dlaczego nasza gra?"}),e.jsx("p",{className:"text-lg text-gray-400 leading-relaxed text-left mb-8",children:"Stworzona przez programistów dla programistów, nasza gra oferuje unikalne doświadczenie nauki JavaScript poprzez praktyczne wyzwania i interaktywne scenariusze. Każdy poziom to nowa wyzwanie, które rozwija Twoje umiejętności."}),e.jsxs("div",{className:"grid grid-cols-1 md:grid-cols-2 gap-4",children:[e.jsxs("div",{className:"bg-black/20 border border-[#f7df1e]/10 rounded-xl p-4 hover:border-[#f7df1e]/20 transition-all duration-300",children:[e.jsxs("div",{className:"flex items-center gap-3 mb-2",children:[e.jsx("div",{className:"w-8 h-8 bg-[#f7df1e]/10 rounded-lg flex items-center justify-center text-[#f7df1e] text-sm",children:"⚡"}),e.jsx("h4",{className:"text-lg font-bold text-[#f7df1e] font-space",children:"Szybka Nauka"})]}),e.jsx("p",{className:"text-sm text-gray-400",children:"Efektywne metody uczenia się przez praktykę"})]}),e.jsxs("div",{className:"bg-black/20 border border-[#f7df1e]/10 rounded-xl p-4 hover:border-[#f7df1e]/20 transition-all duration-300",children:[e.jsxs("div",{className:"flex items-center gap-3 mb-2",children:[e.jsx("div",{className:"w-8 h-8 bg-[#f7df1e]/10 rounded-lg flex items-center justify-center text-[#f7df1e] text-sm",children:"🏆"}),e.jsx("h4",{className:"text-lg font-bold text-[#f7df1e] font-space",children:"Wyzwania"})]}),e.jsx("p",{className:"text-sm text-gray-400",children:"Progresywne poziomy trudności"})]})]})]}),e.jsx("div",{className:"lg:pl-8",children:e.jsx(Qt,{})})]})}),Xs=()=>e.jsx("section",{id:"gra",className:"min-h-screen w-full bg-gradient-to-b from-[#1a1a1a] via-[#242424] to-[#1a1a1a] pt-40 md:pt-32 pb-12 md:pb-16 relative overflow-hidden","aria-labelledby":"about-game-title",children:e.jsx(g,{className:"relative z-10",children:e.jsx(Xt,{})})}),es=()=>{const t=p(),s=e.jsxs("div",{className:"text-right",children:[e.jsxs("div",{className:"flex justify-end items-center gap-2 mb-4",children:[e.jsx("div",{className:"w-3 h-3 bg-[#f7df1e] rounded-full"}),e.jsx("div",{className:"w-3 h-3 bg-[#f7df1e] rounded-full"}),e.jsx("div",{className:"w-3 h-3 bg-[#f7df1e] rounded-full"})]}),e.jsx("h2",{id:"latest-news-title",className:"text-6xl md:text-7xl lg:text-8xl xl:text-9xl font-bold text-[#f7df1e] font-space uppercase tracking-wider transform -skew-x-12",style:{textShadow:"2px 2px 0px rgba(26, 26, 26, 0.5)"},children:"Co Nowego"})]});return t?s:e.jsx(n.div,{initial:{opacity:0,x:50},whileInView:{opacity:1,x:0},transition:{duration:.6},viewport:{once:!0},children:s})},ts=()=>B(void 0,null,function*(){try{const t=yield fetch(`${oe}latest-features?limit=2&isActive=true&sortBy=releaseDate&sortOrder=desc`);if(!t.ok)throw new Error("Błąd podczas pobierania najnowszych funkcji");return(yield t.json()).data||[]}catch(t){throw console.error("Błąd podczas pobierania najnowszych funkcji:",t),t}}),ss=["latestFeatures"],as=()=>{const{data:t,isLoading:s,error:a}=ie({queryKey:ss,queryFn:ts,staleTime:3e5,refetchOnWindowFocus:!1});return{features:t||[],isLoading:s,error:a,data:t}},is=t=>new Date(t).toLocaleDateString("pl-PL",{year:"numeric",month:"long",day:"numeric"}),rs=({version:t,title:s,description:a,category:i,categoryColor:r="#f7df1e",releaseDate:l,tags:d,delay:c})=>{const x=p(),m=e.jsxs("div",{className:"group cursor-pointer",children:[e.jsxs("div",{className:"relative bg-[#1a1a1a] border border-[#f7df1e]/20 rounded-lg mb-6 p-6 group-hover:border-[#f7df1e]/40 transition-all duration-300",children:[e.jsxs("div",{className:"text-center mb-4",children:[e.jsx("div",{className:"text-4xl md:text-5xl font-bold text-[#f7df1e] font-space mb-2",children:t}),e.jsx("div",{className:"text-sm text-gray-300 uppercase tracking-wider",children:is(l)})]}),e.jsxs("div",{className:"text-center",children:[e.jsx("h3",{className:"text-lg md:text-xl font-bold text-white leading-tight font-space uppercase tracking-wide group-hover:text-[#f7df1e] transition-colors duration-300 mb-4",children:s}),e.jsx("p",{className:"text-gray-300 text-sm leading-relaxed mb-4 overflow-hidden",style:{display:"-webkit-box",WebkitLineClamp:3,WebkitBoxOrient:"vertical"},children:a}),d.length>0&&e.jsx("div",{className:"flex flex-wrap gap-2 justify-center",children:d.slice(0,3).map((h,u)=>e.jsxs("span",{className:"text-xs px-2 py-1 bg-[#f7df1e]/10 text-[#f7df1e] rounded border border-[#f7df1e]/20",children:["#",h]},u))})]}),e.jsx("div",{className:"absolute top-4 left-4",children:e.jsx("span",{className:"text-[#1a1a1a] px-3 py-1 text-sm font-bold uppercase tracking-wider font-space",style:{backgroundColor:r},children:i})}),e.jsx("div",{className:"absolute bottom-4 right-4 w-4 h-4",style:{backgroundColor:r,clipPath:"polygon(50% 0%, 0% 100%, 100% 100%)"}})]}),e.jsxs("div",{className:"flex items-center justify-center gap-1",children:[e.jsx("div",{className:"w-2 h-2 bg-[#f7df1e] rotate-45"}),e.jsx("div",{className:"w-2 h-2 bg-[#f7df1e] rotate-45"}),e.jsx("div",{className:"w-2 h-2 bg-[#f7df1e] rotate-45"})]})]});return x?m:e.jsx(n.div,{initial:{opacity:0,y:30},whileInView:{opacity:1,y:0},transition:{duration:.6,delay:c},viewport:{once:!0},children:m})},ns=t=>({feature:"Nowość",update:"Aktualizacja",improvement:"Ulepszenie",bugfix:"Poprawka"})[t]||t,os=t=>({feature:"#f7df1e",update:"#3b82f6",improvement:"#10b981",bugfix:"#ef4444"})[t]||"#f7df1e",ls=()=>{const t=p(),{features:s,isLoading:a,error:i}=as();return a?e.jsx(lt,{}):i?e.jsx(mt,{message:"Wystąpił błąd podczas ładowania najnowszych funkcji"}):!s||s.length===0?e.jsx("div",{className:"text-center py-12",children:e.jsx("p",{className:"text-gray-400 text-lg",children:"Brak najnowszych funkcji do wyświetlenia"})}):e.jsx("div",{className:"grid grid-cols-1 md:grid-cols-2 gap-8 md:gap-12",children:s.map((r,l)=>e.jsx(rs,{id:r._id,version:r.version||`v${r.priority}.0`,title:r.title.toUpperCase(),description:r.description,category:ns(r.category),categoryColor:os(r.category),releaseDate:r.releaseDate,tags:r.tags,delay:t?0:(l+1)*.1},r._id))})},cs=()=>e.jsx("div",{className:"w-full max-w-6xl mx-auto",children:e.jsxs("div",{className:"space-y-12",children:[e.jsx(es,{}),e.jsx(ls,{})]})}),ea=()=>e.jsxs("section",{id:"nowosci",className:"relative w-full bg-gradient-to-b from-[#1a1a1a] via-[#242424] to-[#1a1a1a] py-16 md:py-24 overflow-hidden","aria-labelledby":"latest-news-title",children:[e.jsxs("div",{className:"absolute inset-0",children:[e.jsx("div",{className:"absolute inset-0 bg-[#1a1a1a] opacity-90"}),e.jsx("div",{className:"absolute left-0 top-0 w-1/4 h-full bg-[#f7df1e]",style:{clipPath:"polygon(0 0, 80% 0, 60% 100%, 0 100%)"}}),e.jsx("div",{className:"absolute right-0 bottom-0 w-1/5 h-1/2 bg-[#f7df1e]",style:{clipPath:"polygon(40% 0, 100% 0, 100% 100%, 0 100%)"}}),e.jsx("div",{className:"absolute top-1/4 right-1/4 w-6 h-6 bg-[#f7df1e]",style:{clipPath:"polygon(50% 0%, 0% 100%, 100% 100%)"}}),e.jsx("div",{className:"absolute bottom-1/3 left-1/3 w-4 h-4 bg-[#f7df1e]",style:{clipPath:"polygon(50% 0%, 0% 100%, 100% 100%)"}}),e.jsx("div",{className:"absolute top-1/2 right-1/3 w-3 h-3 bg-[#f7df1e]",style:{clipPath:"polygon(50% 0%, 0% 100%, 100% 100%)"}}),e.jsx("div",{className:"absolute -top-40 -left-40 w-96 h-96 bg-[#f7df1e] rounded-full blur-[150px] opacity-20"}),e.jsx("div",{className:"absolute -bottom-40 -right-40 w-96 h-96 bg-[#f7df1e] rounded-full blur-[150px] opacity-20"})]}),e.jsx(g,{className:"relative z-10",children:e.jsx(cs,{})})]}),ds=()=>{const t=p(),s=e.jsxs("div",{className:"space-y-6",children:[e.jsx("h2",{id:"why-this-game-title",className:"text-3xl md:text-4xl lg:text-5xl font-bold text-white font-space leading-tight",children:"Dlaczego nasza gra zmieni Twoje podejście do programowania?"}),e.jsx("p",{className:"text-lg md:text-xl text-gray-400 leading-relaxed",children:"Stworzona przez doświadczonych programistów i edukatorów, nasza gra oferuje unikalne doświadczenie nauki JavaScript poprzez praktyczne wyzwania, interaktywne projekty i realistyczne scenariusze."}),e.jsx("p",{className:"text-lg md:text-xl text-gray-400 leading-relaxed",children:"Każdy poziom został starannie zaprojektowany, aby wprowadzić nowe koncepty na podstawie solidnych fundamentów programistycznych. Na tyle szczegółowo, że - Rozwiązuj prawdziwe problemy i rozwijaj umiejętności krok po kroku."}),e.jsxs("div",{className:"space-y-4 mt-8",children:[e.jsxs("div",{className:"flex items-start gap-3",children:[e.jsx("div",{className:"w-2 h-2 bg-[#f7df1e] rounded-full mt-3 shrink-0"}),e.jsxs("p",{className:"text-gray-300",children:[e.jsx("span",{className:"text-[#f7df1e] font-semibold",children:"Praktyczne wyzwania"})," - Rozwiązuj rzeczywiste problemy programistyczne"]})]}),e.jsxs("div",{className:"flex items-start gap-3",children:[e.jsx("div",{className:"w-2 h-2 bg-[#f7df1e] rounded-full mt-3 shrink-0"}),e.jsxs("p",{className:"text-gray-300",children:[e.jsx("span",{className:"text-[#f7df1e] font-semibold",children:"Interaktywne środowisko"})," - Testuj kod w czasie rzeczywistym"]})]}),e.jsxs("div",{className:"flex items-start gap-3",children:[e.jsx("div",{className:"w-2 h-2 bg-[#f7df1e] rounded-full mt-3 shrink-0"}),e.jsxs("p",{className:"text-gray-300",children:[e.jsx("span",{className:"text-[#f7df1e] font-semibold",children:"Progresywna nauka"})," - Od podstaw do zaawansowanych konceptów"]})]})]})]});return t?s:e.jsx(n.div,{initial:{opacity:0,x:-50},whileInView:{opacity:1,x:0},transition:{duration:.8},viewport:{once:!0},children:s})},xs=()=>e.jsxs("div",{className:"bg-[#1e1e1e] rounded-lg border border-gray-700 overflow-hidden shadow-2xl",children:[e.jsxs("div",{className:"bg-[#2d2d2d] px-4 py-3 flex items-center gap-3 border-b border-gray-700",children:[e.jsxs("div",{className:"flex items-center gap-2",children:[e.jsx("div",{className:"w-3 h-3 bg-red-500 rounded-full"}),e.jsx("div",{className:"w-3 h-3 bg-yellow-500 rounded-full"}),e.jsx("div",{className:"w-3 h-3 bg-green-500 rounded-full"})]}),e.jsx("div",{className:"text-gray-300 text-sm font-mono",children:"game-challenge.js"})]}),e.jsx("div",{className:"p-4 font-mono text-sm",children:e.jsxs("div",{className:"flex",children:[e.jsx("div",{className:"text-gray-500 select-none mr-4 text-right",style:{minWidth:"2rem"},children:`function solveProblem(level) {
-  // Test kod tutaj
-  const result = processData(level.data);
-  
-  return validateSolution(result);
-}
-
-// Przykład wyzwania z poziomu 15
-const challenge = {
-  title: "Sortowanie Algorytmów",
-  description: "Zaimplementuj quicksort",
-  difficulty: "intermediate"
-};`.split(`
-`).map((s,a)=>e.jsx("div",{className:"leading-6",children:a+1},a))}),e.jsx("div",{className:"flex-1",children:e.jsx("pre",{className:"text-gray-300 leading-6",children:e.jsxs("code",{children:[e.jsx("span",{className:"text-purple-400",children:"function"})," ",e.jsx("span",{className:"text-blue-400",children:"solveProblem"}),e.jsx("span",{className:"text-gray-300",children:"("}),e.jsx("span",{className:"text-orange-400",children:"level"}),e.jsxs("span",{className:"text-gray-300",children:[") ","{"]}),`
-`,e.jsx("span",{className:"text-gray-500",children:"  // Test kod tutaj"}),`
-`,e.jsx("span",{className:"text-purple-400",children:"  const"})," ",e.jsx("span",{className:"text-blue-400",children:"result"})," ",e.jsx("span",{className:"text-gray-300",children:"= "}),e.jsx("span",{className:"text-yellow-400",children:"processData"}),e.jsx("span",{className:"text-gray-300",children:"(level.data);"}),`
-`,`
-`,e.jsx("span",{className:"text-purple-400",children:"  return"})," ",e.jsx("span",{className:"text-yellow-400",children:"validateSolution"}),e.jsx("span",{className:"text-gray-300",children:"(result);"}),`
-`,e.jsx("span",{className:"text-gray-300",children:"}"}),`
-`,`
-`,e.jsx("span",{className:"text-gray-500",children:"// Przykład wyzwania z poziomu 15"}),`
-`,e.jsx("span",{className:"text-purple-400",children:"const"})," ",e.jsx("span",{className:"text-blue-400",children:"challenge"})," ",e.jsxs("span",{className:"text-gray-300",children:["= ","{"]}),`
-`,e.jsx("span",{className:"text-red-400",children:"  title"}),e.jsx("span",{className:"text-gray-300",children:": "}),e.jsx("span",{className:"text-green-400",children:'"Sortowanie Algorytmów"'}),e.jsx("span",{className:"text-gray-300",children:","}),`
-`,e.jsx("span",{className:"text-red-400",children:"  description"}),e.jsx("span",{className:"text-gray-300",children:": "}),e.jsx("span",{className:"text-green-400",children:'"Zaimplementuj quicksort"'}),e.jsx("span",{className:"text-gray-300",children:","}),`
-`,e.jsx("span",{className:"text-red-400",children:"  difficulty"}),e.jsx("span",{className:"text-gray-300",children:": "}),e.jsx("span",{className:"text-green-400",children:'"intermediate"'}),`
-`,e.jsx("span",{className:"text-gray-300",children:"};"})]})})})]})})]}),Y=()=>e.jsxs("div",{className:"relative",children:[e.jsx("div",{className:"w-64 h-[500px] bg-gray-900 rounded-[2.5rem] p-2 shadow-2xl border-4 border-gray-800",children:e.jsxs("div",{className:"w-full h-full bg-[#1a1a1a] rounded-[2rem] overflow-hidden relative",children:[e.jsxs("div",{className:"bg-[#2d2d2d] px-4 py-2 flex justify-between items-center text-white text-xs",children:[e.jsx("span",{children:"9:41"}),e.jsx("div",{className:"flex items-center gap-1",children:e.jsx("div",{className:"w-4 h-2 border border-white rounded-sm",children:e.jsx("div",{className:"w-3 h-1 bg-green-400 rounded-sm"})})})]}),e.jsxs("div",{className:"p-4 space-y-4",children:[e.jsxs("div",{className:"text-center",children:[e.jsx("h3",{className:"text-[#f7df1e] font-bold text-lg font-space",children:"JavaScript Game"}),e.jsx("p",{className:"text-gray-400 text-sm",children:"Poziom 15"})]}),e.jsxs("div",{className:"bg-[#2d2d2d] rounded-lg p-3 border border-[#f7df1e]/20",children:[e.jsx("h4",{className:"text-white font-semibold text-sm mb-2",children:"Wyzwanie: Sortowanie"}),e.jsx("p",{className:"text-gray-300 text-xs leading-relaxed",children:"Rozwiązuj rzeczywiste problemy programistyczne w formie gry. Każdy poziom to nowe wyzwanie, które rozwija Twoje umiejętności."})]}),e.jsxs("div",{className:"space-y-2",children:[e.jsxs("div",{className:"flex justify-between text-xs",children:[e.jsx("span",{className:"text-gray-400",children:"Postęp"}),e.jsx("span",{className:"text-[#f7df1e]",children:"75%"})]}),e.jsx("div",{className:"w-full bg-gray-700 rounded-full h-2",children:e.jsx("div",{className:"bg-[#f7df1e] h-2 rounded-full",style:{width:"75%"}})})]}),e.jsxs("div",{className:"space-y-2",children:[e.jsx("button",{className:"w-full bg-[#f7df1e] text-[#1a1a1a] py-2 rounded-lg font-semibold text-sm",children:"Uruchom Kod"}),e.jsx("button",{className:"w-full border border-[#f7df1e] text-[#f7df1e] py-2 rounded-lg font-semibold text-sm",children:"Sprawdź Rozwiązanie"})]}),e.jsx("div",{className:"bg-[#f7df1e]/10 border border-[#f7df1e]/20 rounded-lg p-3",children:e.jsx("p",{className:"text-[#f7df1e] text-xs",children:"💡 Wskazówka: Użyj rekurencji do podziału tablicy"})})]})]})}),e.jsx("div",{className:"absolute inset-0 bg-[#f7df1e] rounded-[2.5rem] blur-xl opacity-20 -z-10"})]}),ms=()=>{const t=p(),s=e.jsxs("div",{className:"relative",children:[e.jsx("div",{className:"relative z-10",children:e.jsx(xs,{})}),e.jsx("div",{className:"absolute -bottom-8 -right-8 z-20 hidden lg:block",children:e.jsx(Y,{})}),e.jsx("div",{className:"mt-8 lg:hidden",children:e.jsx(Y,{})})]});return t?s:e.jsx(n.div,{initial:{opacity:0,x:50},whileInView:{opacity:1,x:0},transition:{duration:.8,delay:.2},viewport:{once:!0},children:s})},q=({icon:t,title:s,description:a,variant:i})=>{const r="rounded-2xl p-8 h-full min-h-[400px] flex flex-col justify-between hover:scale-105 transition-all duration-300 cursor-pointer group",l=i==="yellow"?"bg-[#f7df1e] text-[#1a1a1a]":"bg-[#1a1a1a] text-white border border-gray-700";return e.jsxs("div",{className:`${r} ${l}`,children:[e.jsx("div",{className:"text-5xl mb-6 group-hover:scale-110 transition-transform duration-300",children:t}),e.jsxs("div",{children:[e.jsx("h3",{className:"text-3xl font-bold font-space mb-4 leading-tight",children:s}),e.jsx("p",{className:"text-lg leading-relaxed opacity-80",children:a})]})]})},Q=({icon:t,title:s,description:a})=>e.jsx("div",{className:"bg-[#1a1a1a] text-white rounded-2xl p-8 min-h-[200px] flex items-center hover:scale-105 transition-all duration-300 cursor-pointer group border border-gray-700",children:e.jsxs("div",{className:"flex items-center gap-6 w-full",children:[e.jsx("div",{className:"text-4xl group-hover:scale-110 transition-transform duration-300",children:t}),e.jsxs("div",{className:"flex-1",children:[e.jsx("h3",{className:"text-2xl font-bold font-space mb-2",children:s}),e.jsx("p",{className:"text-base opacity-80",children:a})]})]})}),v=({icon:t,title:s,description:a})=>e.jsxs("div",{className:"bg-[#1a1a1a] text-white rounded-2xl p-6 min-h-[150px] flex flex-col justify-between hover:scale-105 transition-all duration-300 cursor-pointer group border border-gray-700",children:[e.jsx("div",{className:"text-3xl group-hover:scale-110 transition-transform duration-300",children:t}),e.jsxs("div",{children:[e.jsx("h3",{className:"text-lg font-bold font-space mb-2",children:s}),e.jsx("p",{className:"text-sm opacity-80",children:a})]})]}),X=({position:t})=>t==="top"?e.jsxs("div",{className:"grid grid-cols-1 lg:grid-cols-3 gap-6 mb-6",children:[e.jsx("div",{className:"lg:row-span-2",children:e.jsx(q,{icon:"🎯",title:"Interaktywne Wyzwania",description:"Rozwiązuj rzeczywiste problemy programistyczne w dynamicznym środowisku nauki",variant:"yellow"})}),e.jsxs("div",{className:"lg:col-span-2 space-y-6",children:[e.jsx(Q,{icon:"📊",title:"Śledzenie Postępów",description:"Monitoruj swoje osiągnięcia i zobacz jak rozwijają się Twoje umiejętności"}),e.jsxs("div",{className:"grid grid-cols-1 md:grid-cols-3 gap-6",children:[e.jsx(v,{icon:"👥",title:"Społeczność",description:"Dołącz do programistów"}),e.jsx(v,{icon:"🏆",title:"Certyfikaty",description:"Zdobywaj osiągnięcia"}),e.jsx(v,{icon:"💬",title:"Wsparcie 24/7",description:"Pomoc przez całą dobę"})]})]})]}):e.jsxs("div",{className:"grid grid-cols-1 lg:grid-cols-3 gap-6",children:[e.jsxs("div",{className:"lg:col-span-2 space-y-6",children:[e.jsx(Q,{icon:"💻",title:"Praktyczne Projekty",description:"Twórz rzeczywiste aplikacje podczas nauki programowania"}),e.jsxs("div",{className:"grid grid-cols-1 md:grid-cols-3 gap-6",children:[e.jsx(v,{icon:"🎓",title:"Mentoring",description:"Wsparcie ekspertów"}),e.jsx(v,{icon:"📁",title:"Portfolio",description:"Buduj swoje projekty"}),e.jsx(v,{icon:"🚀",title:"Kariera",description:"Rozwój zawodowy"})]})]}),e.jsx("div",{className:"lg:row-span-2",children:e.jsx(q,{icon:"🧠",title:"Zaawansowane Algorytmy",description:"Poznaj złożone struktury danych i algorytmy w przystępny sposób",variant:"yellow"})})]}),ps=()=>e.jsxs("div",{className:"w-full",children:[e.jsx(X,{position:"top"}),e.jsx(X,{position:"bottom"})]}),hs=()=>e.jsxs("div",{className:"w-full max-w-7xl mx-auto space-y-20",children:[e.jsxs("div",{className:"grid grid-cols-1 lg:grid-cols-2 gap-12 lg:gap-16 items-center",children:[e.jsx("div",{className:"order-2 lg:order-1",children:e.jsx(ds,{})}),e.jsx("div",{className:"order-1 lg:order-2",children:e.jsx(ms,{})})]}),e.jsx(ps,{})]}),ta=()=>e.jsx("section",{id:"dlaczego-warto",className:"min-h-screen w-full bg-gradient-to-b from-[#1a1a1a] via-[#242424] to-[#1a1a1a] py-16 md:py-24 relative overflow-hidden","aria-labelledby":"dlaczego-warto-title",children:e.jsx(g,{className:"relative z-10",children:e.jsx(hs,{})})}),us=({icon:t})=>e.jsx("div",{className:"text-6xl mb-6 group-hover:scale-110 transition-transform duration-300 text-center",children:t}),fs=({title:t,subtitle:s,description:a})=>e.jsxs("div",{className:"text-center space-y-4",children:[e.jsx("h2",{className:"text-3xl md:text-4xl font-bold font-space leading-tight",children:t}),e.jsx("p",{className:"text-lg md:text-xl opacity-90 max-w-3xl mx-auto",children:s}),a&&e.jsx("p",{className:"text-base opacity-80 max-w-2xl mx-auto",children:a})]}),js=({primaryButton:t,secondaryButton:s,buttonStyles:a})=>e.jsxs("div",{className:"flex flex-col sm:flex-row gap-4 justify-center items-center pt-6",children:[e.jsx(f,{to:t.to,className:`${a.button} px-8 py-4 rounded-xl text-lg transition-all duration-300 hover:scale-105 shadow-lg inline-block text-center`,children:t.text}),s&&a.secondaryButton&&e.jsx(f,{to:s.to,className:`${a.secondaryButton} px-8 py-4 rounded-xl text-lg transition-all duration-300 hover:scale-105 inline-block text-center`,children:s.text})]}),gs=({note:t,features:s})=>!t&&!s?null:e.jsxs("div",{className:"space-y-4",children:[t&&e.jsx("p",{className:"text-sm opacity-70 pt-4 text-center",children:t}),s&&e.jsx("div",{className:"flex flex-wrap justify-center gap-4 pt-4",children:s.map((a,i)=>e.jsx("span",{className:"text-sm opacity-80 bg-black/20 px-3 py-1 rounded-full",children:a},i))})]}),bs=t=>{switch(t){case"primary":return{container:"bg-gradient-to-br from-amber-600 via-yellow-500 to-amber-400 text-black",button:"bg-[#f7df1e] hover:bg-yellow-400 text-black font-bold",secondaryButton:"bg-black/20 hover:bg-black/30 text-black border border-black/30"};case"challenge":return{container:"bg-gradient-to-br from-red-900 via-red-800 to-red-700 text-white border border-red-600",button:"bg-red-600 hover:bg-red-500 text-white font-bold"};case"beginner":return{container:"bg-gradient-to-br from-green-900 via-green-800 to-green-700 text-white border border-green-600",button:"bg-green-600 hover:bg-green-500 text-white font-bold"};case"secondary":return{container:"bg-gradient-to-br from-blue-900 via-blue-800 to-blue-700 text-white border border-blue-600",button:"bg-[#f7df1e] hover:bg-yellow-400 text-black font-bold"};default:return{container:"bg-[#1a1a1a] text-white border border-gray-700",button:"bg-[#f7df1e] hover:bg-yellow-400 text-black font-bold"}}},A=({variant:t,icon:s,title:a,subtitle:i,description:r,primaryButton:l,secondaryButton:d,note:c,features:x})=>{const m=bs(t);return e.jsxs("div",{className:`${m.container} rounded-2xl p-8 hover:scale-105 transition-all duration-300 cursor-pointer group relative overflow-hidden`,children:[e.jsx(us,{icon:s}),e.jsxs("div",{className:"space-y-4",children:[e.jsx(fs,{title:a,subtitle:i,description:r}),e.jsx(js,{primaryButton:l,secondaryButton:d,buttonStyles:{button:m.button,secondaryButton:m.secondaryButton}}),e.jsx(gs,{note:c,features:x})]}),e.jsx("div",{className:"absolute inset-0 bg-gradient-to-r from-transparent via-white/5 to-transparent opacity-0 group-hover:opacity-100 transition-opacity duration-500 pointer-events-none"})]})},ys=()=>e.jsx("div",{className:"mb-8",children:e.jsx(A,{variant:"primary",icon:"🎮",title:"Rozpocznij Swoją Przygodę z Kodem!",subtitle:"Dołącz do 75,000+ graczy którzy już opanowują JavaScript przez zabawę!",primaryButton:{text:"⚡ ZACZNIJ GRAĆ ZA DARMO",to:"/dashboard"},secondaryButton:{text:"👁️ ZOBACZ GAMEPLAY",to:"/dashboard"},note:"✨ Bez rejestracji • Najwspanialszy start • 100% za darmo"})}),ws=()=>e.jsxs("div",{className:"grid grid-cols-1 lg:grid-cols-2 gap-8",children:[e.jsx(A,{variant:"challenge",icon:"🔥",title:"Wyzwanie dla Pro Graczy",subtitle:"Myślisz, że jesteś gotowy?",description:"Tylko 3% graczy ukończyło wszystkie poziomy. Czy dołączysz do elit?",primaryButton:{text:"🏆 PRZYJMIJ WYZWANIE",to:"/dashboard"}}),e.jsx(A,{variant:"beginner",icon:"🚀",title:"Szybki Start dla Początkujących",subtitle:"Pierwszy raz z kodem? Idealnie!",description:"Naucz się podstaw w 30 minut i zacznij budować swoje pierwsze aplikacje!",primaryButton:{text:"🌟 ZACZNIJ OD PODSTAW",to:"/dashboard"}})]}),vs=()=>e.jsx("div",{className:"mt-8",children:e.jsx(A,{variant:"secondary",icon:"💎",title:"Twoja Przygoda z Kodem Zaczyna Się TERAZ",subtitle:"Każdy ekspert był kiedyś początkującym. Każda linia kodu przybliża Cię do celu.",description:"Nie czekaj - zacznij dziś!",primaryButton:{text:"🎯 ROZPOCZNIJ PRZYGODĘ!",to:"/dashboard"},features:["🔒 Bez zobowiązań","🎯 Aktualny kiedyś chcesz","🎮 Wsparcie 24/7"]})}),Ns=()=>{const t=p(),s=e.jsxs("div",{className:"w-full max-w-7xl mx-auto",children:[e.jsx(ys,{}),e.jsx(ws,{}),e.jsx(vs,{})]});return t?s:e.jsx(n.div,{initial:{opacity:0,y:50},whileInView:{opacity:1,y:0},transition:{duration:.8},viewport:{once:!0},children:s})},sa=()=>e.jsx("section",{id:"cta",className:"min-h-screen w-full bg-gradient-to-b from-[#1a1a1a] via-[#242424] to-[#1a1a1a] py-16 md:py-24 relative overflow-hidden","aria-labelledby":"cta-title",children:e.jsx(g,{className:"relative z-10",children:e.jsx(Ns,{})})}),zs=({isOpen:t,onClick:s,hasNewMessages:a})=>e.jsxs(n.button,{onClick:s,className:"relative w-16 h-16 rounded-full shadow-lg transition-all duration-300 bg-[#f7df1e] hover:bg-yellow-400 flex items-center justify-center group font-space border-2 border-black",whileHover:{scale:1.1},whileTap:{scale:.95},initial:{scale:0},animate:{scale:1},transition:{type:"spring",stiffness:260,damping:20},children:[e.jsx(n.div,{animate:{rotate:t?45:0},transition:{duration:.3},className:"text-2xl text-black",children:t?"✕":"💬"}),a&&!t&&e.jsx(n.div,{initial:{scale:0},animate:{scale:1},className:"absolute -top-1 -right-1 w-6 h-6 bg-red-500 rounded-full border-2 border-white flex items-center justify-center",children:e.jsx("span",{className:"text-white text-xs font-bold font-space",children:"1"})}),!t&&e.jsx("div",{className:"absolute inset-0 rounded-full bg-[#f7df1e] opacity-30 animate-ping"}),e.jsxs("div",{className:"absolute bottom-full right-0 mb-2 px-3 py-1 bg-gray-900 text-white text-sm rounded-lg opacity-0 group-hover:opacity-100 transition-opacity duration-300 whitespace-nowrap font-space",children:[t?"Zamknij chat":"Potrzebujesz pomocy?",e.jsx("div",{className:"absolute top-full right-4 w-0 h-0 border-l-4 border-r-4 border-t-4 border-transparent border-t-gray-900"})]})]}),ks=({onClose:t})=>e.jsxs("div",{className:"bg-gradient-to-r from-[#f7df1e] to-yellow-400 p-4 flex items-center justify-between",children:[e.jsxs("div",{className:"flex items-center gap-3",children:[e.jsx("div",{className:"w-10 h-10 bg-[#1a1a1a] rounded-full flex items-center justify-center shadow-lg",children:e.jsx("span",{className:"text-[#f7df1e] text-sm font-bold font-space",children:"JS"})}),e.jsxs("div",{children:[e.jsx("h3",{className:"font-bold text-[#1a1a1a] font-space text-lg",children:"Asystent Kodowania"}),e.jsxs("div",{className:"flex items-center gap-2",children:[e.jsx("div",{className:"w-2 h-2 bg-orange-500 rounded-full animate-pulse"}),e.jsx("span",{className:"text-xs text-[#1a1a1a] opacity-80 font-space",children:"W fazie rozwoju"})]})]})]}),e.jsx("button",{onClick:t,className:"w-8 h-8 rounded-full bg-[#1a1a1a]/10 hover:bg-[#1a1a1a]/20 flex items-center justify-center transition-all duration-200 group",children:e.jsx("span",{className:"text-[#1a1a1a] text-lg group-hover:scale-110 transition-transform duration-200",children:"✕"})})]}),Cs=({message:t})=>{const s=t.sender==="user";return e.jsx("div",{className:`flex ${s?"justify-end":"justify-start"}`,children:e.jsxs("div",{className:`flex items-start gap-2 max-w-[80%] ${s?"flex-row-reverse":"flex-row"}`,children:[e.jsx("div",{className:`w-8 h-8 rounded-full flex items-center justify-center flex-shrink-0 shadow-sm ${s?"bg-gradient-to-br from-[#f7df1e] to-yellow-400 text-[#1a1a1a]":"bg-gradient-to-br from-gray-600 to-gray-700 text-white"}`,children:e.jsx("span",{className:"text-xs font-bold font-space",children:s?"TY":"JS"})}),e.jsxs("div",{className:`rounded-2xl px-4 py-3 shadow-sm ${s?"bg-gradient-to-br from-[#f7df1e] to-yellow-400 text-[#1a1a1a] rounded-br-md":"bg-white text-gray-800 rounded-bl-md border border-gray-100"}`,children:[e.jsx("p",{className:"text-sm leading-relaxed whitespace-pre-wrap font-space",children:t.text}),e.jsx("div",{className:`text-xs mt-2 font-space ${s?"text-[#1a1a1a]/60":"text-gray-500"}`,children:t.timestamp.toLocaleTimeString("pl-PL",{hour:"2-digit",minute:"2-digit"})})]})]})})},Ss=()=>e.jsx("div",{className:"flex justify-start",children:e.jsxs("div",{className:"flex items-start gap-2 max-w-[80%]",children:[e.jsx("div",{className:"w-8 h-8 rounded-full bg-gradient-to-br from-gray-600 to-gray-700 text-white flex items-center justify-center flex-shrink-0 shadow-sm",children:e.jsx("span",{className:"text-xs font-bold font-space",children:"JS"})}),e.jsx("div",{className:"bg-white text-gray-800 rounded-2xl rounded-bl-md shadow-sm border border-gray-100 px-4 py-3",children:e.jsxs("div",{className:"flex items-center gap-2",children:[e.jsxs("div",{className:"flex gap-1",children:[e.jsx("div",{className:"w-2 h-2 bg-[#f7df1e] rounded-full animate-bounce",style:{animationDelay:"0ms"}}),e.jsx("div",{className:"w-2 h-2 bg-[#f7df1e] rounded-full animate-bounce",style:{animationDelay:"150ms"}}),e.jsx("div",{className:"w-2 h-2 bg-[#f7df1e] rounded-full animate-bounce",style:{animationDelay:"300ms"}})]}),e.jsx("span",{className:"text-xs text-gray-500 ml-1 font-space",children:"pisze..."})]})})]})}),Ms=({messages:t,isTyping:s,messagesEndRef:a})=>e.jsxs("div",{className:"flex-1 overflow-y-auto p-4 space-y-4 bg-gray-50",children:[t.map((i,r)=>e.jsx(n.div,{initial:{opacity:0,y:20},animate:{opacity:1,y:0},transition:{duration:.3,delay:r*.1},children:e.jsx(Cs,{message:i})},i.id)),s&&e.jsx(n.div,{initial:{opacity:0,y:20},animate:{opacity:1,y:0},transition:{duration:.3},children:e.jsx(Ss,{})}),e.jsx("div",{ref:a})]}),Ts=({value:t,onChange:s,onSend:a,onKeyPress:i})=>e.jsxs("div",{className:"p-4 bg-white border-t border-gray-100",children:[e.jsxs("div",{className:"flex items-end gap-3",children:[e.jsx("div",{className:"flex-1 relative",children:e.jsx("textarea",{value:t,onChange:r=>s(r.target.value),onKeyPress:i,placeholder:"Napisz wiadomość...",className:"w-full resize-none rounded-xl border-0 px-4 py-3 text-sm focus:outline-none focus:ring-2 focus:ring-[#f7df1e] max-h-20 min-h-[44px] font-space placeholder-gray-400 transition-all duration-200 text-gray-800 bg-gray-50",rows:1,style:{height:"auto",minHeight:"44px",maxHeight:"80px",color:"#374151",backgroundColor:"#f9fafb",textDecoration:"none",outline:"none",border:"none"},onInput:r=>{const l=r.target;l.style.height="auto",l.style.height=Math.min(l.scrollHeight,80)+"px"},spellCheck:!1})}),e.jsx(n.button,{onClick:a,disabled:!t.trim(),className:`w-12 h-12 rounded-xl flex items-center justify-center transition-all duration-200 shadow-sm ${t.trim()?"bg-gradient-to-br from-[#f7df1e] to-yellow-400 hover:from-yellow-400 hover:to-yellow-500 text-[#1a1a1a] shadow-lg":"bg-gray-200 text-gray-400 cursor-not-allowed"}`,whileHover:t.trim()?{scale:1.05}:{},whileTap:t.trim()?{scale:.95}:{},children:e.jsxs("svg",{width:"20",height:"20",viewBox:"0 0 24 24",fill:"none",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round",children:[e.jsx("line",{x1:"22",y1:"2",x2:"11",y2:"13"}),e.jsx("polygon",{points:"22,2 15,22 11,13 2,9 22,2"})]})})]}),e.jsxs("div",{className:"text-xs text-gray-600 mt-2 text-center font-space",children:["Naciśnij ",e.jsx("span",{className:"font-bold text-[#f7df1e]",children:"Enter"})," aby wysłać, ",e.jsx("span",{className:"font-bold text-[#f7df1e]",children:"Shift+Enter"})," dla nowej linii"]})]}),Ps=({messages:t,onSendMessage:s,onClose:a,isTyping:i})=>{const[r,l]=o.useState(""),d=o.useRef(null),c=p(),x=()=>{var u;(u=d.current)==null||u.scrollIntoView({behavior:"smooth"})};o.useEffect(()=>{x()},[t,i]);const m=()=>{r.trim()&&(s(r.trim()),l(""))},h=u=>{u.key==="Enter"&&!u.shiftKey&&(u.preventDefault(),m())};return e.jsxs(n.div,{className:`
-        bg-white rounded-2xl shadow-2xl border border-gray-200 flex flex-col overflow-hidden
-        ${c?"fixed inset-4 top-16 bottom-20 w-auto h-auto":"w-80 h-96"}
-      `,initial:{opacity:0,y:20},animate:{opacity:1,y:0},transition:{duration:.3},children:[e.jsx(ks,{onClose:a}),e.jsx(Ms,{messages:t,isTyping:i,messagesEndRef:d}),e.jsx(Ts,{value:r,onChange:l,onSend:m,onKeyPress:h})]})},As=({message:t,isVisible:s,onClose:a})=>e.jsx(I,{children:s&&e.jsx(n.div,{initial:{opacity:0,scale:.8,y:10},animate:{opacity:1,scale:1,y:0},exit:{opacity:0,scale:.8,y:10},transition:{type:"spring",stiffness:400,damping:25,duration:.3},className:"absolute bottom-full right-0 mb-3 max-w-64 z-[9999]",onClick:a,style:{zIndex:9999},children:e.jsxs("div",{className:"bg-white rounded-2xl shadow-2xl border border-gray-200 p-3 relative cursor-pointer hover:shadow-2xl transition-shadow duration-200",style:{boxShadow:"0 25px 50px -12px rgba(0, 0, 0, 0.25), 0 0 0 1px rgba(0, 0, 0, 0.05), 0 10px 25px rgba(0, 0, 0, 0.15)",zIndex:9999},children:[e.jsxs("div",{className:"flex items-center gap-2 mb-2",children:[e.jsx("div",{className:"w-6 h-6 rounded-full bg-gradient-to-br from-gray-600 to-gray-700 text-white flex items-center justify-center",children:e.jsx("span",{className:"text-xs font-bold font-space",children:"JS"})}),e.jsx("span",{className:"text-xs font-bold text-gray-800 font-space",children:"Asystent"}),e.jsx("button",{onClick:i=>{i.stopPropagation(),a()},className:"ml-auto w-4 h-4 rounded-full bg-gray-100 hover:bg-gray-200 flex items-center justify-center transition-colors duration-200",children:e.jsx("span",{className:"text-gray-500 text-xs",children:"✕"})})]}),e.jsx("p",{className:"text-sm text-gray-800 font-space leading-relaxed",children:t.length>60?`${t.substring(0,60)}...`:t}),e.jsx("div",{className:"text-xs text-[#f7df1e] mt-1 font-space font-bold",children:"Kliknij aby otworzyć"}),e.jsx("div",{className:"absolute top-full right-4 w-0 h-0 border-l-4 border-r-4 border-t-4 border-transparent border-t-white",style:{zIndex:9999}}),e.jsx("div",{className:"absolute top-full right-4 w-0 h-0 border-l-4 border-r-4 border-t-4 border-transparent border-t-gray-200 transform translate-y-px",style:{zIndex:9999}}),e.jsx("div",{className:"absolute inset-0 bg-gradient-to-r from-transparent via-[#f7df1e]/5 to-transparent opacity-0 hover:opacity-100 transition-opacity duration-300 rounded-2xl pointer-events-none"}),e.jsx("div",{className:"absolute -inset-0.5 bg-[#f7df1e] rounded-2xl opacity-10 animate-pulse pointer-events-none"})]})})}),Is=()=>{const[t,s]=o.useState([{id:"1",text:"Cześć! 👋 Jestem asystentem kodowania. Aktualnie jestem w fazie rozwoju, ale już wkrótce będę mógł Ci pomóc w nauce programowania!",sender:"bot",timestamp:new Date}]),[a,i]=o.useState(!1),r=o.useCallback(l=>{const d={id:Date.now().toString(),text:l,sender:"user",timestamp:new Date};s(c=>[...c,d]),i(!0),setTimeout(()=>{const c=Fs(l),x={id:(Date.now()+1).toString(),text:c,sender:"bot",timestamp:new Date};s(m=>[...m,x]),i(!1)},1e3+Math.random()*2e3)},[]);return{messages:t,addMessage:r,isTyping:a}},Fs=t=>{const s=t.toLowerCase(),a=["Dziękuję za wiadomość! 🚧 Aktualnie jestem w fazie rozwoju i nie mogę jeszcze w pełni odpowiadać na pytania. Wkrótce będę gotowy do pomocy!","Cześć! 👷‍♂️ Jestem jeszcze w budowie, ale już niedługo będę mógł Ci pomóc z programowaniem. Dziękuję za cierpliwość!","Hej! 🔧 Chat jest jeszcze w fazie testowej. Już wkrótce będę w pełni funkcjonalny i pomogę Ci w nauce kodowania!","Witaj! ⚙️ Jestem w trakcie rozwoju. Niedługo będę gotowy, żeby odpowiadać na wszystkie Twoje pytania o programowanie!"];return s.includes("kiedy")||s.includes("gdy")||s.includes("gotowy")?"Pracujemy nad tym! 🚀 Chat będzie w pełni funkcjonalny już wkrótce. Śledź nasze aktualizacje!":s.includes("dziękuję")||s.includes("dzięki")||s.includes("ok")?"Nie ma za co! 😊 Cieszę się, że jesteś zainteresowany. Wróć wkrótce - będę gotowy do pomocy!":s.includes("cześć")||s.includes("hej")||s.includes("witaj")||s.includes("hello")?"Cześć! 👋 Miło Cię poznać! Jestem jeszcze w fazie rozwoju, ale już niedługo będę mógł Ci pomóc z programowaniem!":a[Math.floor(Math.random()*a.length)]},aa=()=>{const[t,s]=o.useState(!1),[a,i]=o.useState(!1),[r,l]=o.useState("Cześć! 👋 Sprawdź nowego asystenta kodowania!"),[d,c]=o.useState(!1),[x,m]=o.useState(!1),{messages:h,addMessage:u,isTyping:k}=Is(),F=p(),V=()=>{s(!t),m(!0),t||(i(!1),c(!1))},Pe=()=>{c(!1),i(!1),s(!0),m(!0)};o.useEffect(()=>{if(!t&&!x){const w=setTimeout(()=>{c(!0),setTimeout(()=>{c(!1)},4e3)},3e3),Ie=setTimeout(()=>{i(!0)},5e3);return()=>{clearTimeout(w),clearTimeout(Ie)}}},[t,x]),o.useEffect(()=>{if(h.length>1&&!t&&!x){const w=h[h.length-1];w.sender==="bot"&&(setTimeout(()=>{l(w.text),c(!0),setTimeout(()=>{c(!1)},4e3)},1e3),setTimeout(()=>{i(!0)},2e3))}},[h,t,x]);const Ae=()=>{s(!1)};return e.jsxs(e.Fragment,{children:[t&&e.jsx(n.div,{initial:{opacity:0},animate:{opacity:1},exit:{opacity:0},className:`fixed inset-0 z-[9998] ${F?"bg-black/50":"bg-transparent"}`,onClick:Ae}),e.jsxs("div",{className:`fixed z-[9999] ${F?"bottom-4 right-4":"bottom-6 right-6"}`,style:{zIndex:9999},children:[e.jsx(As,{message:r,isVisible:d&&!t,onClose:Pe}),e.jsx(I,{children:t&&e.jsx(n.div,{initial:{opacity:0,scale:.8,y:20},animate:{opacity:1,scale:1,y:0},exit:{opacity:0,scale:.8,y:20},transition:{duration:.3,ease:"easeOut"},className:F?"":"mb-4",onClick:w=>w.stopPropagation(),style:{zIndex:9999},children:e.jsx(Ps,{messages:h,onSendMessage:u,onClose:V,isTyping:k})})}),e.jsx(zs,{isOpen:t,onClick:V,hasNewMessages:a})]})]})},Es=({error:t,resetErrorBoundary:s})=>{const a=()=>{s?s():window.location.reload()};return e.jsx("div",{className:"min-h-screen flex items-center justify-center bg-dark p-4",children:e.jsx("div",{className:"bg-dark-medium border border-red-500/20 rounded-lg p-8 max-w-md w-full shadow-lg",children:e.jsxs("div",{className:"flex flex-col items-center text-center",children:[e.jsx(N,{className:"text-5xl text-red-500 mb-4","aria-hidden":"true"}),e.jsx("h1",{className:"text-2xl font-bold text-white mb-2",children:"Wystąpił nieoczekiwany błąd"}),e.jsx("p",{className:"text-gray-400 mb-6",children:"Przepraszamy za niedogodności. Nasz zespół został powiadomiony o problemie."}),t&&e.jsx("div",{className:"w-full mb-6 p-3 bg-red-900/20 border border-red-500/10 rounded overflow-auto text-left",children:e.jsx("p",{className:"text-sm text-red-400 font-mono",children:t.message||"Nieznany błąd aplikacji"})}),e.jsxs("div",{className:"flex flex-col sm:flex-row gap-4 w-full",children:[e.jsxs("button",{onClick:a,className:"flex-1 flex items-center justify-center gap-2 py-3 px-4 bg-js hover:bg-js/90 text-dark font-medium rounded-lg transition-colors",children:[e.jsx(et,{className:"text-sm","aria-hidden":"true"}),e.jsx("span",{children:"Spróbuj ponownie"})]}),e.jsx(f,{to:"/dashboard",className:"flex-1",children:e.jsxs("button",{className:"w-full flex items-center justify-center gap-2 py-3 px-4 bg-gray-800 hover:bg-gray-700 text-white font-medium rounded-lg transition-colors",children:[e.jsx(ee,{className:"text-sm","aria-hidden":"true"}),e.jsx("span",{children:"Wróć do Home"})]})})]})]})})})},ia=Object.freeze(Object.defineProperty({__proto__:null,default:Es},Symbol.toStringTag,{value:"Module"}));export{Vs as A,fe as B,sa as C,ia as D,mt as E,dt as F,Ks as H,O as L,rt as M,xt as P,Zs as S,Us as T,ta as W,ue as a,Os as b,Gs as c,Ws as d,lt as e,ct as f,_s as g,qs as h,ea as i,Xs as j,Ys as k,Yt as l,Qs as m,aa as n,g as o,Hs as p,Rs as q,Js as r,p as u};
