+++ server/public/assets/app-dashboard-D-SLglbn.js
@@ -1,5 +1,5 @@
-const __vite__mapDeps=(i,m=__vite__mapDeps,d=(m.f||(m.f=["assets/app-components-DCNmHqgN.js","assets/vendor-react-DZE_uCEP.js","assets/vendor-other-DT6kVT6s.js","assets/vendor-monaco-C9aCaKYu.js","assets/vendor-framer-motion-D6A4MpK3.js","assets/app-auth-BOk5gCNS.js","assets/app-services-MwNJNg2u.js"])))=>i.map(i=>d[i]);
-var Vi=Object.defineProperty,Zi=Object.defineProperties;var Ki=Object.getOwnPropertyDescriptors;var Ds=Object.getOwnPropertySymbols;var qi=Object.prototype.hasOwnProperty,Xi=Object.prototype.propertyIsEnumerable;var Fs=(s,t,a)=>t in s?Vi(s,t,{enumerable:!0,configurable:!0,writable:!0,value:a}):s[t]=a,S=(s,t)=>{for(var a in t||(t={}))qi.call(t,a)&&Fs(s,a,t[a]);if(Ds)for(var a of Ds(t))Xi.call(t,a)&&Fs(s,a,t[a]);return s},z=(s,t)=>Zi(s,Ki(t));var k=(s,t,a)=>new Promise((r,n)=>{var o=l=>{try{d(a.next(l))}catch(x){n(x)}},c=l=>{try{d(a.throw(l))}catch(x){n(x)}},d=l=>l.done?r(l.value):Promise.resolve(l.value).then(o,c);d((a=a.apply(s,t)).next())});import{r as i,P as He,j as e,Q as Ms,H as te,R as Yi,S as Os,w as A,p as en,T as sn,o as ps,v as Te,A as hs,U as Vs,V as Zs,u as je,W as Ks,C as J,X as ye,Y as ce,Z as js,q as tn,z as K,L as Le,_ as an,$ as qs,M,a0 as ys,a1 as Xs,a2 as ie,g as Z,f as Ys,l as H,a3 as F,b as se,a4 as et,a5 as Ge,a6 as gs,a7 as rn,a8 as st,a9 as tt,aa as ue,ab as nn,x as bs,ac as We,ad as on,i as de,ae as ln,af as fs,m as Ue,ag as ws,ah as Ns,ai as at,c as cn,d as dn,aj as vs,ak as mn,al as xn,am as ne,an as rt,ao as it,ap as P,aq as nt,ar as un,h as ks,N as ot,e as lt,as as Ss,a as Qe,F as pn,at as zs,au as hn,av as jn,aw as ct,ax as dt,ay as yn,az as gn,aA as bn,aB as fn,aC as wn,aD as Nn,aE as vn,aF as Cs,aG as Pe,aH as kn,aI as Sn,aJ as zn,aK as Cn,aL as Tn,aM as Ln,aN as Pn,aO as En,aP as mt,aQ as ls}from"./vendor-react-DZE_uCEP.js";import{L as cs,u as xt,M as An,d as Je,e as q,f as $n,g as Rn,a as he,B as ze}from"./app-components-DCNmHqgN.js";import{u as _,A as $}from"./app-auth-BOk5gCNS.js";import{m as u,A as G,u as Dn}from"./vendor-framer-motion-D6A4MpK3.js";import{f as ut,p as pt,z as O,t as Ve,a as ds}from"./vendor-other-DT6kVT6s.js";import{R as Fn,L as Mn,C as On,X as _n,Y as Bn,T as In,a as Hn}from"./vendor-charts-BQN0K2I5.js";import{h as X}from"./app-services-MwNJNg2u.js";import{l as Gn}from"./vendor-monaco-C9aCaKYu.js";const Wn="modulepreload",Un=function(s){return"/"+s},_s={},ee=function(t,a,r){let n=Promise.resolve();if(a&&a.length>0){document.getElementsByTagName("link");const c=document.querySelector("meta[property=csp-nonce]"),d=(c==null?void 0:c.nonce)||(c==null?void 0:c.getAttribute("nonce"));n=Promise.allSettled(a.map(l=>{if(l=Un(l),l in _s)return;_s[l]=!0;const x=l.endsWith(".css"),m=x?'[rel="stylesheet"]':"";if(document.querySelector(`link[href="${l}"]${m}`))return;const p=document.createElement("link");if(p.rel=x?"stylesheet":Wn,x||(p.as="script"),p.crossOrigin="",p.href=l,d&&p.setAttribute("nonce",d),document.head.appendChild(p),x)return new Promise((h,j)=>{p.addEventListener("load",h),p.addEventListener("error",()=>j(new Error(`Unable to preload CSS for ${l}`)))})}))}function o(c){const d=new Event("vite:preloadError",{cancelable:!0});if(d.payload=c,window.dispatchEvent(d),!d.defaultPrevented)throw c}return n.then(c=>{for(const d of c||[])d.status==="rejected"&&o(d.reason);return t().catch(o)})},Qn={activeView:"community"},Jn=(s,t)=>{switch(t.type){case"SET_ACTIVE_VIEW":return z(S({},s),{activeView:t.payload});default:return s}},Vn=()=>{const[s,t]=i.useReducer(Jn,Qn),a=i.useCallback(r=>{t({type:"SET_ACTIVE_VIEW",payload:r})},[]);return{state:s,setActiveView:a}},ht=i.memo(()=>{const t=He().pathname,a=t==="/dashboard/community"||t==="/dashboard/community/"||/^\/dashboard\/community\/?$/.test(t),r=t.includes("/dashboard/community/ranking");return e.jsxs("div",{className:"flex gap-4 mb-8 pb-2",children:[e.jsxs(Ms,{to:"/dashboard/community",className:`flex items-center gap-2 px-6 py-3 font-semibold rounded-lg transition-all ${a?"bg-js/10 text-js border-b-2 border-js shadow-md":"text-gray-400 hover:text-js hover:bg-js/5"}`,children:[e.jsx("i",{className:"fas fa-newspaper"}),"Aktualności"]}),e.jsxs(Ms,{to:"/dashboard/community/ranking",className:`flex items-center gap-2 px-6 py-3 font-semibold rounded-lg transition-all ${r?"bg-js/10 text-js border-b-2 border-js shadow-md":"text-gray-400 hover:text-js hover:bg-js/5"}`,children:[e.jsx("i",{className:"fas fa-trophy"}),"Ranking"]})]})});ht.displayName="CommunityNavigation";const Zn=i.lazy(()=>ee(()=>Promise.resolve().then(()=>zc),void 0)),Kn=i.lazy(()=>ee(()=>Promise.resolve().then(()=>Lc),void 0)),jt=i.memo(()=>{const s=He(),{state:{activeView:t},setActiveView:a}=Vn();return i.useEffect(()=>{const r=s.pathname.split("community").pop()||"community";r!==t&&a(r)},[s.pathname,t,a]),e.jsxs("div",{className:"p-8 w-full min-h-screen bg-dark/50 backdrop-blur-sm",children:[e.jsxs(te,{children:[e.jsx("title",{children:"Community | CodeLinesJS"}),e.jsx("meta",{name:"description",content:"Community CodeLinesJS - dołącz do nas i rozwijaj swoje umiejętności w przyjaznym środowisku."})]}),e.jsx("div",{className:"max-w-7xl mx-auto",children:e.jsxs("div",{className:"bg-dark/30 backdrop-blur-sm rounded-xl border border-js/10 p-6 shadow-lg",children:[e.jsx(ht,{}),e.jsxs(Yi,{children:[e.jsx(Os,{index:!0,element:e.jsx(i.Suspense,{fallback:e.jsx(cs,{text:"Ładowanie aktualności..."}),children:e.jsx(Zn,{})})}),e.jsx(Os,{path:"ranking",element:e.jsx(i.Suspense,{fallback:e.jsx(cs,{text:"Ładowanie rankingu..."}),children:e.jsx(Kn,{})})})]})]})})]})});jt.displayName="CommunitySection";const Sd=Object.freeze(Object.defineProperty({__proto__:null,CommunitySection:jt},Symbol.toStringTag,{value:"Module"})),qn=()=>{const[s,t]=i.useState(!1),[a,r]=i.useState("dashboard"),n=He();return i.useEffect(()=>{const c=n.pathname.split("/").pop()||"dashboard";r(c)},[n]),i.useEffect(()=>{const o=()=>{window.innerWidth<768&&t(!1)};return window.addEventListener("resize",o),o(),()=>window.removeEventListener("resize",o)},[]),{isExpanded:s,setIsExpanded:t,activeItem:a,setActiveItem:r}},Xn={expanded:{width:"240px",transition:{type:"spring",stiffness:300,damping:30}},collapsed:{width:"100px",transition:{type:"spring",stiffness:300,damping:30}}},Ts={expanded:{opacity:1,x:0,transition:{type:"spring",stiffness:300,damping:30}},collapsed:{opacity:0,x:-10,transition:{duration:.2}}},yt=i.memo(({isExpanded:s})=>e.jsx("div",{className:"px-4 mb-8",children:e.jsxs(u.div,{initial:{opacity:0},animate:{opacity:1},className:"flex items-center gap-3",children:[e.jsx("div",{className:"w-10 h-10 min-w-[40px] rounded-xl bg-js/20 flex items-center justify-center",children:e.jsx(A,{className:"text-xl text-js"})}),e.jsx(G,{mode:"wait",children:s&&e.jsx(u.span,{variants:Ts,initial:"collapsed",animate:"expanded",exit:"collapsed",className:"font-bold text-js font-space",children:"CodeLinesJS"})})]})}));yt.displayName="NavigationLogo";const Ls=i.memo(({icon:s,label:t,isActive:a=!1,isExpanded:r,onClick:n,variant:o="default",id:c})=>e.jsxs(u.button,{onClick:n,"data-testid":`nav-button-${t==null?void 0:t.toLowerCase().replace(/\s/g,"-")}`,className:`
+const __vite__mapDeps=(i,m=__vite__mapDeps,d=(m.f||(m.f=["assets/app-components-L1wo8hYs.js","assets/vendor-react-DZE_uCEP.js","assets/vendor-other-DT6kVT6s.js","assets/vendor-monaco-C9aCaKYu.js","assets/vendor-framer-motion-D6A4MpK3.js","assets/app-auth-C1CjYxGY.js","assets/app-services-vJYhj3YV.js"])))=>i.map(i=>d[i]);
+var Vi=Object.defineProperty,Zi=Object.defineProperties;var Ki=Object.getOwnPropertyDescriptors;var Ds=Object.getOwnPropertySymbols;var qi=Object.prototype.hasOwnProperty,Xi=Object.prototype.propertyIsEnumerable;var Fs=(s,t,a)=>t in s?Vi(s,t,{enumerable:!0,configurable:!0,writable:!0,value:a}):s[t]=a,S=(s,t)=>{for(var a in t||(t={}))qi.call(t,a)&&Fs(s,a,t[a]);if(Ds)for(var a of Ds(t))Xi.call(t,a)&&Fs(s,a,t[a]);return s},z=(s,t)=>Zi(s,Ki(t));var k=(s,t,a)=>new Promise((r,n)=>{var o=l=>{try{d(a.next(l))}catch(x){n(x)}},c=l=>{try{d(a.throw(l))}catch(x){n(x)}},d=l=>l.done?r(l.value):Promise.resolve(l.value).then(o,c);d((a=a.apply(s,t)).next())});import{r as i,P as He,j as e,Q as Ms,H as te,R as Yi,S as Os,w as A,p as en,T as sn,o as ps,v as Te,A as hs,U as Vs,V as Zs,u as je,W as Ks,C as J,X as ye,Y as ce,Z as js,q as tn,z as K,L as Le,_ as an,$ as qs,M,a0 as ys,a1 as Xs,a2 as ie,g as Z,f as Ys,l as H,a3 as F,b as se,a4 as et,a5 as Ge,a6 as gs,a7 as rn,a8 as st,a9 as tt,aa as ue,ab as nn,x as bs,ac as We,ad as on,i as de,ae as ln,af as fs,m as Ue,ag as ws,ah as Ns,ai as at,c as cn,d as dn,aj as vs,ak as mn,al as xn,am as ne,an as rt,ao as it,ap as P,aq as nt,ar as un,h as ks,N as ot,e as lt,as as Ss,a as Qe,F as pn,at as zs,au as hn,av as jn,aw as ct,ax as dt,ay as yn,az as gn,aA as bn,aB as fn,aC as wn,aD as Nn,aE as vn,aF as Cs,aG as Pe,aH as kn,aI as Sn,aJ as zn,aK as Cn,aL as Tn,aM as Ln,aN as Pn,aO as En,aP as mt,aQ as ls}from"./vendor-react-DZE_uCEP.js";import{L as cs,u as xt,M as An,d as Je,e as q,f as $n,g as Rn,a as he,B as ze}from"./app-components-L1wo8hYs.js";import{u as _,A as $}from"./app-auth-C1CjYxGY.js";import{m as u,A as G,u as Dn}from"./vendor-framer-motion-D6A4MpK3.js";import{f as ut,p as pt,z as O,t as Ve,a as ds}from"./vendor-other-DT6kVT6s.js";import{R as Fn,L as Mn,C as On,X as _n,Y as Bn,T as In,a as Hn}from"./vendor-charts-BQN0K2I5.js";import{h as X}from"./app-services-vJYhj3YV.js";import{l as Gn}from"./vendor-monaco-C9aCaKYu.js";const Wn="modulepreload",Un=function(s){return"/"+s},_s={},ee=function(t,a,r){let n=Promise.resolve();if(a&&a.length>0){document.getElementsByTagName("link");const c=document.querySelector("meta[property=csp-nonce]"),d=(c==null?void 0:c.nonce)||(c==null?void 0:c.getAttribute("nonce"));n=Promise.allSettled(a.map(l=>{if(l=Un(l),l in _s)return;_s[l]=!0;const x=l.endsWith(".css"),m=x?'[rel="stylesheet"]':"";if(document.querySelector(`link[href="${l}"]${m}`))return;const p=document.createElement("link");if(p.rel=x?"stylesheet":Wn,x||(p.as="script"),p.crossOrigin="",p.href=l,d&&p.setAttribute("nonce",d),document.head.appendChild(p),x)return new Promise((h,j)=>{p.addEventListener("load",h),p.addEventListener("error",()=>j(new Error(`Unable to preload CSS for ${l}`)))})}))}function o(c){const d=new Event("vite:preloadError",{cancelable:!0});if(d.payload=c,window.dispatchEvent(d),!d.defaultPrevented)throw c}return n.then(c=>{for(const d of c||[])d.status==="rejected"&&o(d.reason);return t().catch(o)})},Qn={activeView:"community"},Jn=(s,t)=>{switch(t.type){case"SET_ACTIVE_VIEW":return z(S({},s),{activeView:t.payload});default:return s}},Vn=()=>{const[s,t]=i.useReducer(Jn,Qn),a=i.useCallback(r=>{t({type:"SET_ACTIVE_VIEW",payload:r})},[]);return{state:s,setActiveView:a}},ht=i.memo(()=>{const t=He().pathname,a=t==="/dashboard/community"||t==="/dashboard/community/"||/^\/dashboard\/community\/?$/.test(t),r=t.includes("/dashboard/community/ranking");return e.jsxs("div",{className:"flex gap-4 mb-8 pb-2",children:[e.jsxs(Ms,{to:"/dashboard/community",className:`flex items-center gap-2 px-6 py-3 font-semibold rounded-lg transition-all ${a?"bg-js/10 text-js border-b-2 border-js shadow-md":"text-gray-400 hover:text-js hover:bg-js/5"}`,children:[e.jsx("i",{className:"fas fa-newspaper"}),"Aktualności"]}),e.jsxs(Ms,{to:"/dashboard/community/ranking",className:`flex items-center gap-2 px-6 py-3 font-semibold rounded-lg transition-all ${r?"bg-js/10 text-js border-b-2 border-js shadow-md":"text-gray-400 hover:text-js hover:bg-js/5"}`,children:[e.jsx("i",{className:"fas fa-trophy"}),"Ranking"]})]})});ht.displayName="CommunityNavigation";const Zn=i.lazy(()=>ee(()=>Promise.resolve().then(()=>zc),void 0)),Kn=i.lazy(()=>ee(()=>Promise.resolve().then(()=>Lc),void 0)),jt=i.memo(()=>{const s=He(),{state:{activeView:t},setActiveView:a}=Vn();return i.useEffect(()=>{const r=s.pathname.split("community").pop()||"community";r!==t&&a(r)},[s.pathname,t,a]),e.jsxs("div",{className:"p-8 w-full min-h-screen bg-dark/50 backdrop-blur-sm",children:[e.jsxs(te,{children:[e.jsx("title",{children:"Community | CodeLinesJS"}),e.jsx("meta",{name:"description",content:"Community CodeLinesJS - dołącz do nas i rozwijaj swoje umiejętności w przyjaznym środowisku."})]}),e.jsx("div",{className:"max-w-7xl mx-auto",children:e.jsxs("div",{className:"bg-dark/30 backdrop-blur-sm rounded-xl border border-js/10 p-6 shadow-lg",children:[e.jsx(ht,{}),e.jsxs(Yi,{children:[e.jsx(Os,{index:!0,element:e.jsx(i.Suspense,{fallback:e.jsx(cs,{text:"Ładowanie aktualności..."}),children:e.jsx(Zn,{})})}),e.jsx(Os,{path:"ranking",element:e.jsx(i.Suspense,{fallback:e.jsx(cs,{text:"Ładowanie rankingu..."}),children:e.jsx(Kn,{})})})]})]})})]})});jt.displayName="CommunitySection";const Sd=Object.freeze(Object.defineProperty({__proto__:null,CommunitySection:jt},Symbol.toStringTag,{value:"Module"})),qn=()=>{const[s,t]=i.useState(!1),[a,r]=i.useState("dashboard"),n=He();return i.useEffect(()=>{const c=n.pathname.split("/").pop()||"dashboard";r(c)},[n]),i.useEffect(()=>{const o=()=>{window.innerWidth<768&&t(!1)};return window.addEventListener("resize",o),o(),()=>window.removeEventListener("resize",o)},[]),{isExpanded:s,setIsExpanded:t,activeItem:a,setActiveItem:r}},Xn={expanded:{width:"240px",transition:{type:"spring",stiffness:300,damping:30}},collapsed:{width:"100px",transition:{type:"spring",stiffness:300,damping:30}}},Ts={expanded:{opacity:1,x:0,transition:{type:"spring",stiffness:300,damping:30}},collapsed:{opacity:0,x:-10,transition:{duration:.2}}},yt=i.memo(({isExpanded:s})=>e.jsx("div",{className:"px-4 mb-8",children:e.jsxs(u.div,{initial:{opacity:0},animate:{opacity:1},className:"flex items-center gap-3",children:[e.jsx("div",{className:"w-10 h-10 min-w-[40px] rounded-xl bg-js/20 flex items-center justify-center",children:e.jsx(A,{className:"text-xl text-js"})}),e.jsx(G,{mode:"wait",children:s&&e.jsx(u.span,{variants:Ts,initial:"collapsed",animate:"expanded",exit:"collapsed",className:"font-bold text-js font-space",children:"CodeLinesJS"})})]})}));yt.displayName="NavigationLogo";const Ls=i.memo(({icon:s,label:t,isActive:a=!1,isExpanded:r,onClick:n,variant:o="default",id:c})=>e.jsxs(u.button,{onClick:n,"data-testid":`nav-button-${t==null?void 0:t.toLowerCase().replace(/\s/g,"-")}`,className:`
       ${c==="code"&&"hidden lg:flex"}
       w-full flex items-center gap-3 px-4 py-3 
       rounded-lg transition-all duration-200
@@ -47,7 +47,7 @@ var Vi=Object.defineProperty,Zi=Object.defineProperties;var Ki=Object.getOwnProp
     max-w-[1000px]
     mx-auto
     relative
-  `,actions:"flex items-center gap-3 absolute right-6",gradients:{text:"bg-gradient-to-r from-js to-js/80 bg-clip-text text-transparent",hover:"hover:bg-js/5",base:"bg-gradient-to-br from-js/5 to-js/10"},borders:{base:"border-js/20",hover:"hover:border-js/30"},transitions:{base:"transition-all duration-200"},text:{heading:"text-lg font-bold text-js",primary:"text-white",secondary:"text-gray-400"}},no=()=>i.useMemo(()=>({container:{initial:{y:-20,opacity:0},animate:{y:0,opacity:1,transition:{duration:.3,ease:"easeOut"}}}}),[]),oo=()=>k(void 0,null,function*(){const s=yield fetch(`${$}users/profile`,{headers:{"Content-Type":"application/json"},credentials:"include"});if(!s.ok)throw new Error("Nie udało się pobrać danych użytkownika");return s.json()}),lo=["userProfile"],co=()=>{const{isAuthenticated:s,isAuthChecking:t}=_(),a=J({queryKey:lo,queryFn:()=>oo(),enabled:s&&!t}),r=a.isLoading||t;return z(S({},a),{isLoading:r})},zt=i.memo(()=>{const s=no(),{data:t,isLoading:a,error:r}=co(),n=i.useMemo(()=>r?"Użytkowniku":a?"Ładowanie...":(t==null?void 0:t.data.user.username)||"Użytkowniku",[t==null?void 0:t.data.user.username,a,r]);return e.jsx(Je,{children:e.jsx(u.nav,{variants:s.container,initial:"initial",animate:"animate",className:ss.container,"aria-label":"Górna nawigacja",children:e.jsxs("div",{className:ss.content,children:[e.jsx(wt,{username:n}),e.jsx("div",{className:ss.actions,children:e.jsx(i.Suspense,{fallback:e.jsx("div",{className:"w-10 h-10 rounded-full bg-gray-700 animate-pulse"}),children:e.jsx(St,{})})})]})})})});zt.displayName="TopNavigation";const mo=i.lazy(()=>ee(()=>import("./app-components-DCNmHqgN.js").then(s=>s.D),__vite__mapDeps([0,1,2,3,4,5,6]))),xo=i.lazy(()=>ee(()=>import("./app-components-DCNmHqgN.js").then(s=>s.p),__vite__mapDeps([0,1,2,3,4,5,6])).then(s=>({default:s.LoadingSpinner}))),Ct=i.memo(()=>{const s=bt(),t=xt(),{logout:a}=_();return e.jsx(Je,{fallback:e.jsx(i.Suspense,{fallback:e.jsx("div",{className:"flex items-center justify-center min-h-screen",children:"Ładowanie..."}),children:e.jsx(mo,{})}),children:e.jsxs("main",{className:"bg-gradient-to-b from-dark via-dark-medium to-dark backdrop-blur-lg flex min-h-screen",children:[e.jsxs(te,{children:[e.jsx("title",{children:"Dashboard | CodeLinesJS"}),e.jsx("meta",{name:"description",content:"Panel użytkownika CodeLinesJS - dostęp do kursów, gier i statystyk nauki."})]}),e.jsx(ft,{}),e.jsxs("div",{className:`flex-1 ${!s&&!t?"ml-[100px]":""} md:ml-[100px]`,children:[e.jsx(zt,{}),e.jsx("div",{className:"mt-20 p-6",children:e.jsxs("div",{className:`${s||t?"flex":""} flex-col justify-center items-center gap-4 md:block`,children:[s||t?e.jsxs("div",{className:"flex flex-col gap-2 mb-4 md:hidden",children:[e.jsx(Le,{to:"/dashboard",children:e.jsx("button",{"aria-label":"Wróć do strony głównej",className:"bg-js text-dark px-4 py-2 rounded-md w-full flex items-center justify-center",children:"Wróć do Home"})}),e.jsxs("button",{onClick:a,"aria-label":"Wyloguj się",className:"bg-red-600 text-white px-4 py-2 rounded-md flex items-center justify-center gap-2",children:[e.jsx(Ks,{"aria-hidden":"true"}),"Wyloguj się"]})]}):null,e.jsx(i.Suspense,{fallback:e.jsx(xo,{text:"Ładowanie zawartości..."}),children:e.jsx(an,{})})]})})]})]})})});Ct.displayName="Dashboard";const zd=Object.freeze(Object.defineProperty({__proto__:null,default:Ct},Symbol.toStringTag,{value:"Module"})),ve={text:{heading:"font-bold text-js"},grid:"grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8",card:{base:"bg-dark/50 backdrop-blur-sm border border-js/10 shadow-xl",community:`
+  `,actions:"flex items-center gap-3 absolute right-6",gradients:{text:"bg-gradient-to-r from-js to-js/80 bg-clip-text text-transparent",hover:"hover:bg-js/5",base:"bg-gradient-to-br from-js/5 to-js/10"},borders:{base:"border-js/20",hover:"hover:border-js/30"},transitions:{base:"transition-all duration-200"},text:{heading:"text-lg font-bold text-js",primary:"text-white",secondary:"text-gray-400"}},no=()=>i.useMemo(()=>({container:{initial:{y:-20,opacity:0},animate:{y:0,opacity:1,transition:{duration:.3,ease:"easeOut"}}}}),[]),oo=()=>k(void 0,null,function*(){const s=yield fetch(`${$}users/profile`,{headers:{"Content-Type":"application/json"},credentials:"include"});if(!s.ok)throw new Error("Nie udało się pobrać danych użytkownika");return s.json()}),lo=["userProfile"],co=()=>{const{isAuthenticated:s,isAuthChecking:t}=_(),a=J({queryKey:lo,queryFn:()=>oo(),enabled:s&&!t}),r=a.isLoading||t;return z(S({},a),{isLoading:r})},zt=i.memo(()=>{const s=no(),{data:t,isLoading:a,error:r}=co(),n=i.useMemo(()=>r?"Użytkowniku":a?"Ładowanie...":(t==null?void 0:t.data.user.username)||"Użytkowniku",[t==null?void 0:t.data.user.username,a,r]);return e.jsx(Je,{children:e.jsx(u.nav,{variants:s.container,initial:"initial",animate:"animate",className:ss.container,"aria-label":"Górna nawigacja",children:e.jsxs("div",{className:ss.content,children:[e.jsx(wt,{username:n}),e.jsx("div",{className:ss.actions,children:e.jsx(i.Suspense,{fallback:e.jsx("div",{className:"w-10 h-10 rounded-full bg-gray-700 animate-pulse"}),children:e.jsx(St,{})})})]})})})});zt.displayName="TopNavigation";const mo=i.lazy(()=>ee(()=>import("./app-components-L1wo8hYs.js").then(s=>s.D),__vite__mapDeps([0,1,2,3,4,5,6]))),xo=i.lazy(()=>ee(()=>import("./app-components-L1wo8hYs.js").then(s=>s.p),__vite__mapDeps([0,1,2,3,4,5,6])).then(s=>({default:s.LoadingSpinner}))),Ct=i.memo(()=>{const s=bt(),t=xt(),{logout:a}=_();return e.jsx(Je,{fallback:e.jsx(i.Suspense,{fallback:e.jsx("div",{className:"flex items-center justify-center min-h-screen",children:"Ładowanie..."}),children:e.jsx(mo,{})}),children:e.jsxs("main",{className:"bg-gradient-to-b from-dark via-dark-medium to-dark backdrop-blur-lg flex min-h-screen",children:[e.jsxs(te,{children:[e.jsx("title",{children:"Dashboard | CodeLinesJS"}),e.jsx("meta",{name:"description",content:"Panel użytkownika CodeLinesJS - dostęp do kursów, gier i statystyk nauki."})]}),e.jsx(ft,{}),e.jsxs("div",{className:`flex-1 ${!s&&!t?"ml-[100px]":""} md:ml-[100px]`,children:[e.jsx(zt,{}),e.jsx("div",{className:"mt-20 p-6",children:e.jsxs("div",{className:`${s||t?"flex":""} flex-col justify-center items-center gap-4 md:block`,children:[s||t?e.jsxs("div",{className:"flex flex-col gap-2 mb-4 md:hidden",children:[e.jsx(Le,{to:"/dashboard",children:e.jsx("button",{"aria-label":"Wróć do strony głównej",className:"bg-js text-dark px-4 py-2 rounded-md w-full flex items-center justify-center",children:"Wróć do Home"})}),e.jsxs("button",{onClick:a,"aria-label":"Wyloguj się",className:"bg-red-600 text-white px-4 py-2 rounded-md flex items-center justify-center gap-2",children:[e.jsx(Ks,{"aria-hidden":"true"}),"Wyloguj się"]})]}):null,e.jsx(i.Suspense,{fallback:e.jsx(xo,{text:"Ładowanie zawartości..."}),children:e.jsx(an,{})})]})})]})]})})});Ct.displayName="Dashboard";const zd=Object.freeze(Object.defineProperty({__proto__:null,default:Ct},Symbol.toStringTag,{value:"Module"})),ve={text:{heading:"font-bold text-js"},grid:"grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8",card:{base:"bg-dark/50 backdrop-blur-sm border border-js/10 shadow-xl",community:`
       col-span-1 md:col-span-2 lg:col-span-1
       order-2 lg:order-1
     `,game:`
