+++ server/public/assets/app-services-vJYhj3YV.js
@@ -1,0 +1,1 @@
+var k=Object.defineProperty,T=Object.defineProperties;var E=Object.getOwnPropertyDescriptors;var z=Object.getOwnPropertySymbols;var N=Object.prototype.hasOwnProperty,P=Object.prototype.propertyIsEnumerable;var g=(t,e,s)=>e in t?k(t,e,{enumerable:!0,configurable:!0,writable:!0,value:s}):t[e]=s,u=(t,e)=>{for(var s in e||(e={}))N.call(e,s)&&g(t,s,e[s]);if(z)for(var s of z(e))P.call(e,s)&&g(t,s,e[s]);return t},l=(t,e)=>T(t,E(e));var i=(t,e,s)=>new Promise((f,p)=>{var c=o=>{try{d(s.next(o))}catch(w){p(w)}},h=o=>{try{d(s.throw(o))}catch(w){p(w)}},d=o=>o.done?f(o.value):Promise.resolve(o.value).then(c,h);d((s=s.apply(t,e)).next())});import{A as S}from"./app-auth-C1CjYxGY.js";const b=t=>{const e=t.message.toLowerCase();return e.includes("failed to fetch")?"Nie można połączyć się z serwerem. Sprawdź swoje połączenie internetowe.":e.includes("network request failed")?"Błąd połączenia sieciowego. Sprawdź swoje połączenie internetowe.":e.includes("timeout")?"Przekroczono czas oczekiwania na odpowiedź serwera.":e.includes("abort")?"Żądanie zostało przerwane.":e.includes("cors")?"Błąd polityki CORS. Brak dostępu do zasobu z tej domeny.":t.message},C={request(s){return i(this,arguments,function*(t,e={}){var f,p;try{const{method:c="GET",headers:h={},body:d}=e,w=`${S.replace("www.","")}${t}`,q=u({"Content-Type":"application/json",Accept:"application/json"},h),y={method:c,headers:q,mode:"cors",credentials:"include"};d&&(y.body=JSON.stringify(d));const a=yield fetch(w,y);let r=null,n=null;try{const m=yield a.text();m&&(r=JSON.parse(m))}catch(m){n="Nieprawidłowa odpowiedź serwera"}if(r){if(r.status==="success")return{data:r.data,error:null,status:r.code||a.status};if(r.status==="fail"||r.status==="error")return n=r.message||((p=(f=r.errors)==null?void 0:f[0])==null?void 0:p.message)||`Błąd: ${r.status}`,{data:null,error:n,status:r.code||a.status};if(r.error)return{data:null,error:r.error,status:a.status};if(r.token)return{data:r,error:null,status:a.status}}return!a.ok&&!n?(n=`Błąd ${a.status}: ${a.statusText}`,{data:null,error:n,status:a.status}):{data:(r==null?void 0:r.data)||r,error:n,status:a.status}}catch(c){return{data:null,error:c instanceof Error?b(c):"Nieznany błąd zapytania",status:0}}})},get(s){return i(this,arguments,function*(t,e={}){return this.request(t,l(u({},e),{method:"GET"}))})},post(f,p){return i(this,arguments,function*(t,e,s={}){return this.request(t,l(u({},s),{method:"POST",body:e}))})},put(f,p){return i(this,arguments,function*(t,e,s={}){return this.request(t,l(u({},s),{method:"PUT",body:e}))})},delete(s){return i(this,arguments,function*(t,e={}){return this.request(t,l(u({},e),{method:"DELETE"}))})},patch(f,p){return i(this,arguments,function*(t,e,s={}){return this.request(t,l(u({},s),{method:"PATCH",body:e}))})}};export{C as h};
