+++ server/public/assets/index-DTbAtaTr.js
@@ -1,0 +1,2 @@
+const __vite__mapDeps=(i,m=__vite__mapDeps,d=(m.f||(m.f=["assets/app-dashboard-D-SLglbn.js","assets/vendor-react-DZE_uCEP.js","assets/vendor-other-DT6kVT6s.js","assets/vendor-monaco-C9aCaKYu.js","assets/app-components-L1wo8hYs.js","assets/vendor-framer-motion-D6A4MpK3.js","assets/app-auth-C1CjYxGY.js","assets/app-services-vJYhj3YV.js","assets/vendor-charts-BQN0K2I5.js","assets/page-home.tsx-DKl2Yuyi.js","assets/page-auth.tsx-hSk5PX-M.js","assets/page-privacypolicy.tsx-Bcz1y0px.js"])))=>i.map(i=>d[i]);
+import{r,aX as _,j as t,aY as p,aZ as h,a_ as m,a$ as f,b0 as E,b1 as x}from"./vendor-react-DZE_uCEP.js";import{an as y}from"./vendor-other-DT6kVT6s.js";import{_ as o}from"./app-dashboard-D-SLglbn.js";import{L as j}from"./app-components-L1wo8hYs.js";import"./vendor-monaco-C9aCaKYu.js";import"./app-auth-C1CjYxGY.js";import"./vendor-framer-motion-D6A4MpK3.js";import"./app-services-vJYhj3YV.js";import"./vendor-charts-BQN0K2I5.js";(function(){const a=document.createElement("link").relList;if(a&&a.supports&&a.supports("modulepreload"))return;for(const s of document.querySelectorAll('link[rel="modulepreload"]'))l(s);new MutationObserver(s=>{for(const i of s)if(i.type==="childList")for(const d of i.addedNodes)d.tagName==="LINK"&&d.rel==="modulepreload"&&l(d)}).observe(document,{childList:!0,subtree:!0});function u(s){const i={};return s.integrity&&(i.integrity=s.integrity),s.referrerPolicy&&(i.referrerPolicy=s.referrerPolicy),s.crossOrigin==="use-credentials"?i.credentials="include":s.crossOrigin==="anonymous"?i.credentials="omit":i.credentials="same-origin",i}function l(s){if(s.ep)return;s.ep=!0;const i=u(s);fetch(s.href,i)}})();const P=r.lazy(()=>o(()=>import("./app-dashboard-D-SLglbn.js").then(e=>e.C),__vite__mapDeps([0,1,2,3,4,5,6,7,8])).then(e=>({default:e.CommunitySection}))),c=r.lazy(()=>o(()=>import("./app-components-L1wo8hYs.js").then(e=>e.q),__vite__mapDeps([4,1,2,3,5,6,7])).then(e=>({default:e.ErrorPage}))),g=r.lazy(()=>o(()=>import("./app-dashboard-D-SLglbn.js").then(e=>e.D),__vite__mapDeps([0,1,2,3,4,5,6,7,8]))),L=r.lazy(()=>o(()=>import("./app-components-L1wo8hYs.js").then(e=>e.r),__vite__mapDeps([4,1,2,3,5,6,7])).then(e=>({default:e.ProtectedRoute}))),R=r.lazy(()=>o(()=>import("./app-dashboard-D-SLglbn.js").then(e=>e.a),__vite__mapDeps([0,1,2,3,4,5,6,7,8])).then(e=>({default:e.DashboardContent}))),O=r.lazy(()=>o(()=>import("./app-dashboard-D-SLglbn.js").then(e=>e.S),__vite__mapDeps([0,1,2,3,4,5,6,7,8])).then(e=>({default:e.StatsSection}))),D=r.lazy(()=>o(()=>import("./app-dashboard-D-SLglbn.js").then(e=>e.L),__vite__mapDeps([0,1,2,3,4,5,6,7,8])).then(e=>({default:e.LearnSection}))),S=r.lazy(()=>o(()=>import("./app-dashboard-D-SLglbn.js").then(e=>e.b),__vite__mapDeps([0,1,2,3,4,5,6,7,8])).then(e=>({default:e.LessonPage}))),I=r.lazy(()=>o(()=>import("./app-dashboard-D-SLglbn.js").then(e=>e.c),__vite__mapDeps([0,1,2,3,4,5,6,7,8])).then(e=>({default:e.SettingsSection}))),A=r.lazy(()=>o(()=>import("./app-dashboard-D-SLglbn.js").then(e=>e.G),__vite__mapDeps([0,1,2,3,4,5,6,7,8])).then(e=>({default:e.GamesSection}))),T=r.lazy(()=>o(()=>import("./app-dashboard-D-SLglbn.js").then(e=>e.d),__vite__mapDeps([0,1,2,3,4,5,6,7,8])).then(e=>({default:e.GameplayRouter}))),z=r.lazy(()=>o(()=>import("./app-dashboard-D-SLglbn.js").then(e=>e.e),__vite__mapDeps([0,1,2,3,4,5,6,7,8])).then(e=>({default:e.CodeEditor}))),b=r.lazy(()=>o(()=>import("./app-auth-C1CjYxGY.js").then(e=>e.i),__vite__mapDeps([6,1,2,3,4,5,7])).then(e=>({default:e.ResetPasswordPage}))),V=r.lazy(()=>o(()=>import("./app-dashboard-D-SLglbn.js").then(e=>e.R),__vite__mapDeps([0,1,2,3,4,5,6,7,8])).then(e=>({default:e.ReportBugPage}))),C=r.lazy(()=>o(()=>import("./page-home.tsx-DKl2Yuyi.js"),__vite__mapDeps([9,1,2,3,4,5,6,7]))),w=r.lazy(()=>o(()=>import("./page-auth.tsx-hSk5PX-M.js"),__vite__mapDeps([10,1,2,3,4,5,6,7]))),G=r.lazy(()=>o(()=>import("./page-privacypolicy.tsx-Bcz1y0px.js"),__vite__mapDeps([11,1,2,3,4,5,6,7]))),n=({children:e})=>t.jsx(r.Suspense,{fallback:t.jsx(j,{fullScreen:!0,text:"Ładowanie strony..."}),children:e}),v=_([{path:"/",element:t.jsx(n,{children:t.jsx(C,{})}),errorElement:t.jsx(n,{children:t.jsx(c,{})})},{path:"/logowanie",element:t.jsx(n,{children:t.jsx(w,{})})},{path:"/reset-password/:token",element:t.jsx(n,{children:t.jsx(b,{})})},{path:"/polityka-prywatnosci",element:t.jsx(n,{children:t.jsx(G,{})})},{path:"/dashboard",element:t.jsx(n,{children:t.jsx(L,{children:t.jsx(g,{})})}),children:[{index:!0,element:t.jsx(n,{children:t.jsx(R,{})})},{path:"stats",element:t.jsx(n,{children:t.jsx(O,{})})},{path:"community/*",element:t.jsx(n,{children:t.jsx(P,{})})},{path:"learn",element:t.jsx(n,{children:t.jsx(D,{})})},{path:"learn/lesson/:lessonSlug",element:t.jsx(n,{children:t.jsx(S,{})})},{path:"settings",element:t.jsx(n,{children:t.jsx(I,{})})},{path:"play",element:t.jsx(n,{children:t.jsx(A,{})})},{path:"play/:slug",element:t.jsx(n,{children:t.jsx(T,{})})},{path:"code",element:t.jsx(n,{children:t.jsx(z,{})})},{path:"report",element:t.jsx(n,{children:t.jsx(V,{})})}]},{path:"*",element:t.jsx(n,{children:t.jsx(c,{})})}]),q="466348895611-kdgecrm842jsilesaj0s0c5bnqn7maiq.apps.googleusercontent.com",B=new y;function N(){return t.jsxs(p,{client:B,children:[t.jsx(h,{position:"top-right",toastOptions:{duration:3e3,style:{background:"#1E1E1E",color:"#fff",border:"1px solid rgba(247, 223, 30, 0.1)"},success:{iconTheme:{primary:"#F7DF1E",secondary:"#1E1E1E"}},error:{iconTheme:{primary:"#ff4b4b",secondary:"#1E1E1E"}}}}),t.jsx(m,{clientId:q,onScriptLoadError:()=>console.error("Google Script failed to load"),children:t.jsx(f,{router:v})})]})}E.createRoot(document.getElementById("root")).render(t.jsx(r.StrictMode,{children:t.jsx(x,{children:t.jsx(N,{})})}));
