--- server/public/assets/page-auth.tsx-CvX4pXmn.js
@@ -1,1 +1,0 @@
-import{j as r,r as t}from"./vendor-react-DZE_uCEP.js";import{e as o}from"./app-components-DCNmHqgN.js";import{b as s}from"./app-auth-BOk5gCNS.js";import"./vendor-other-DT6kVT6s.js";import"./vendor-monaco-C9aCaKYu.js";import"./vendor-framer-motion-D6A4MpK3.js";import"./app-services-MwNJNg2u.js";const c=()=>r.jsx(t.Suspense,{fallback:r.jsx(o,{}),children:r.jsx(s,{})});export{c as default};
