+++ server/public/assets/page-auth.tsx-hSk5PX-M.js
@@ -1,0 +1,1 @@
+import{j as r,r as t}from"./vendor-react-DZE_uCEP.js";import{e as o}from"./app-components-L1wo8hYs.js";import{b as s}from"./app-auth-C1CjYxGY.js";import"./vendor-other-DT6kVT6s.js";import"./vendor-monaco-C9aCaKYu.js";import"./vendor-framer-motion-D6A4MpK3.js";import"./app-services-vJYhj3YV.js";const c=()=>r.jsx(t.Suspense,{fallback:r.jsx(o,{}),children:r.jsx(s,{})});export{c as default};
