--- server/public/assets/page-home.tsx-BrPfLk_a.js
@@ -1,1 +1,0 @@
-import{j as s}from"./vendor-react-DZE_uCEP.js";import{H as t,h as o,i as e,j as a,T as i,k as r,W as m,l as j,C as x,m as n,n as c}from"./app-components-DCNmHqgN.js";import"./vendor-other-DT6kVT6s.js";import"./vendor-monaco-C9aCaKYu.js";import"./vendor-framer-motion-D6A4MpK3.js";import"./app-auth-BOk5gCNS.js";import"./app-services-MwNJNg2u.js";const H=()=>s.jsxs("div",{className:"bg-gray-900",children:[s.jsx(t,{}),s.jsx(o,{}),s.jsx(e,{}),s.jsx(a,{}),s.jsx(i,{}),s.jsx(r,{}),s.jsx(m,{}),s.jsx(j,{}),s.jsx(x,{}),s.jsx(n,{}),s.jsx(c,{})]});export{H as default};
