--- server/public/assets/page-privacypolicy.tsx-DKGB4Ooq.js
@@ -1,1 +1,0 @@
-import{j as e,a2 as a,L as i,i as o}from"./vendor-react-DZE_uCEP.js";import{o as n}from"./app-components-DCNmHqgN.js";import{m as s}from"./vendor-framer-motion-D6A4MpK3.js";import"./vendor-other-DT6kVT6s.js";import"./vendor-monaco-C9aCaKYu.js";import"./app-auth-BOk5gCNS.js";import"./app-services-MwNJNg2u.js";const y=()=>e.jsxs(e.Fragment,{children:[e.jsxs(a,{children:[e.jsx("title",{children:"Polityka Prywatności | CodeLinesJS"}),e.jsx("meta",{name:"description",content:"Polityka prywatności CodeLinesJS - dowiedz się, jak chronimy Twoje dane osobowe."})]}),e.jsx("div",{className:"bg-gradient-to-b from-dark via-dark-medium to-dark min-h-screen py-16",children:e.jsxs(n,{children:[e.jsx(i,{to:"/",children:e.jsxs("span",{className:"text-js flex items-center gap-2",children:[e.jsx(o,{}),e.jsx("span",{children:"Wróć do strony głównej"})]})}),e.jsxs(s.div,{initial:{opacity:0,y:20},animate:{opacity:1,y:0},transition:{duration:.5},className:"max-w-4xl mx-auto bg-dark-light/30 p-8 rounded-lg shadow-lg backdrop-blur-sm border border-js/10 mt-6",children:[e.jsx("h1",{className:"text-3xl md:text-4xl font-bold text-white mb-8 border-b border-js pb-4",children:"Polityka prywatności"}),e.jsxs("div",{className:"space-y-8 text-gray-200",children:[e.jsxs("section",{children:[e.jsx("h2",{className:"text-2xl font-semibold text-js mb-4",children:"1. Informacje ogólne"}),e.jsx("p",{className:"mb-3",children:"Niniejsza polityka dotyczy Serwisu www, funkcjonującego pod adresem url: codelinesjs.pl"}),e.jsx("p",{className:"mb-3",children:"Operatorem serwisu oraz Administratorem danych osobowych jest: Oliwier Markieiwcz 98-200, Sieradz"}),e.jsx("p",{className:"mb-3",children:"Adres kontaktowy poczty elektronicznej operatora: kontakt@codelinesjs.pl"}),e.jsx("p",{className:"mb-3",children:"Operator jest Administratorem Twoich danych osobowych w odniesieniu do danych podanych dobrowolnie w Serwisie."}),e.jsx("p",{className:"mb-3",children:"Serwis wykorzystuje dane osobowe w następujących celach:"}),e.jsxs("ul",{className:"list-disc pl-6 space-y-2 mb-3",children:[e.jsx("li",{children:"Prowadzenie newslettera"}),e.jsx("li",{children:"Prowadzenie systemu komentarzy"}),e.jsx("li",{children:"Prowadzenie rozmów typu chat online"}),e.jsx("li",{children:"Prezentacja profil użytkownika innym użytkownikom"}),e.jsx("li",{children:"Obsługa zapytań przez formularz"}),e.jsx("li",{children:"Prezentacja oferty lub informacji"})]}),e.jsx("p",{className:"mb-3",children:"Serwis realizuje funkcje pozyskiwania informacji o użytkownikach i ich zachowaniu w następujący sposób:"}),e.jsxs("ul",{className:"list-disc pl-6 space-y-2",children:[e.jsx("li",{children:"Poprzez dobrowolnie wprowadzone w formularzach dane, które zostają wprowadzone do systemów Operatora."}),e.jsx("li",{children:'Poprzez zapisywanie w urządzeniach końcowych plików cookie (tzw. "ciasteczka").'})]})]}),e.jsxs("section",{children:[e.jsx("h2",{className:"text-2xl font-semibold text-js mb-4",children:"2. Wybrane metody ochrony danych stosowane przez Operatora"}),e.jsxs("ul",{className:"list-disc pl-6 space-y-2",children:[e.jsx("li",{children:"Miejsca logowania i wprowadzania danych osobowych są chronione w warstwie transmisji (certyfikat SSL). Dzięki temu dane osobowe i dane logowania, wprowadzone na stronie, zostają zaszyfrowane w komputerze użytkownika i mogą być odczytane jedynie na docelowym serwerze."}),e.jsx("li",{children:"Dane osobowe przechowywane w bazie danych są zaszyfrowane w taki sposób, że jedynie posiadający Operator klucz może je odczytać. Dzięki temu dane są chronione na wypadek wykradzenia bazy danych z serwera."}),e.jsx("li",{children:"Hasła użytkowników są przechowywane w postaci hashowanej. Funkcja hashująca działa jednokierunkowo - nie jest możliwe odwrócenie jej działania, co stanowi obecnie współczesny standard w zakresie przechowywania haseł użytkowników."}),e.jsx("li",{children:"Operator okresowo zmienia swoje hasła administracyjne."}),e.jsx("li",{children:"W celu ochrony danych Operator regularnie wykonuje kopie bezpieczeństwa."}),e.jsx("li",{children:"Istotnym elementem ochrony danych jest regularna aktualizacja wszelkiego oprogramowania, wykorzystywanego przez Operatora do przetwarzania danych osobowych, co w szczególności oznacza regularne aktualizacje komponentów programistycznych."})]})]}),e.jsxs("section",{children:[e.jsx("h2",{className:"text-2xl font-semibold text-js mb-4",children:"3. Hosting"}),e.jsx("p",{className:"mb-3",children:"Serwis jest hostowany (technicznie utrzymywany) na serwerach operatora: Vercel"}),e.jsx("p",{className:"mb-3",children:"Firma hostingowa w celu zapewnienia niezawodności technicznej prowadzi logi na poziomie serwera. Zapisowi mogą podlegać:"}),e.jsxs("ul",{className:"list-disc pl-6 space-y-2",children:[e.jsx("li",{children:"zasoby określone identyfikatorem URL (adresy żądanych zasobów – stron, plików),"}),e.jsx("li",{children:"czas nadejścia zapytania,"}),e.jsx("li",{children:"czas wysłania odpowiedzi,"}),e.jsx("li",{children:"nazwę stacji klienta – identyfikacja realizowana przez protokół HTTP,"}),e.jsx("li",{children:"informacje o błędach jakie nastąpiły przy realizacji transakcji HTTP,"}),e.jsx("li",{children:"adres URL strony poprzednio odwiedzanej przez użytkownika (referer link) – w przypadku gdy przejście do Serwisu nastąpiło przez odnośnik,"}),e.jsx("li",{children:"informacje o przeglądarce użytkownika,"}),e.jsx("li",{children:"informacje o adresie IP,"}),e.jsx("li",{children:"informacje diagnostyczne związane z procesem samodzielnego zamawiania usług poprzez rejestratory na stronie,"}),e.jsx("li",{children:"informacje związane z obsługą poczty elektronicznej kierowanej do Operatora oraz wysyłanej przez Operatora."})]})]}),e.jsxs("section",{children:[e.jsx("h2",{className:"text-2xl font-semibold text-js mb-4",children:"4. Twoje prawa i dodatkowe informacje o sposobie wykorzystania danych"}),e.jsx("p",{className:"mb-3",children:"W niektórych sytuacjach Administrator ma prawo przekazywać Twoje dane osobowe innym odbiorcom, jeśli będzie to niezbędne do wykonania zawartej z Tobą umowy lub do zrealizowania obowiązków ciążących na Administratorze. Dotyczy to takich grup odbiorców:"}),e.jsxs("ul",{className:"list-disc pl-6 space-y-2 mb-3",children:[e.jsx("li",{children:"firma hostingowa na zasadzie powierzenia"}),e.jsx("li",{children:"operatorzy rozwiązania typu chat online"}),e.jsx("li",{children:"upoważnieni pracownicy i współpracownicy, którzy korzystają z danych w celu realizacji celu działania strony"})]}),e.jsx("p",{className:"mb-3",children:"Twoje dane osobowe przetwarzane przez Administratora nie dłużej, niż jest to konieczne do wykonania związanych z nimi czynności określonych osobnymi przepisami (np. o prowadzeniu rachunkowości). W odniesieniu do danych marketingowych dane nie będą przetwarzane dłużej niż przez 3 lata."}),e.jsx("p",{className:"mb-3",children:"Przysługuje Ci prawo żądania od Administratora:"}),e.jsxs("ul",{className:"list-disc pl-6 space-y-2 mb-3",children:[e.jsx("li",{children:"dostępu do danych osobowych Ciebie dotyczących,"}),e.jsx("li",{children:"ich sprostowania,"}),e.jsx("li",{children:"usunięcia,"}),e.jsx("li",{children:"ograniczenia przetwarzania,"}),e.jsx("li",{children:"oraz przenoszenia danych."})]}),e.jsx("p",{className:"mb-3",children:"Przysługuje Ci prawo do złożenia sprzeciwu w zakresie przetwarzania wskazanego w pkt 3.2 wobec przetwarzania danych osobowych w celu wykonania prawnie uzasadnionych interesów realizowanych przez Administratora, w tym profilowania, przy czym prawo sprzeciwu nie będzie mogło być wykonane w przypadku istnienia ważnych prawnie uzasadnionych podstaw do przetwarzania, nadrzędnych wobec Ciebie interesów, praw i wolności, w szczególności ustalenia, dochodzenia lub obrony roszczeń."}),e.jsx("p",{className:"mb-3",children:"Na działania Administratora przysługuje skarga do Prezesa Urzędu Ochrony Danych Osobowych, ul. Stawki 2, 00-193 Warszawa."}),e.jsx("p",{className:"mb-3",children:"Podanie danych osobowych jest dobrowolne, lecz niezbędne do obsługi Serwisu."}),e.jsx("p",{className:"mb-3",children:"W stosunku do Ciebie mogą być podejmowane czynności polegające na zautomatyzowanym podejmowaniu decyzji, w tym profilowaniu w celu świadczenia usług w ramach zawartej umowy oraz w celu prowadzenia przez Administratora marketingu bezpośredniego."}),e.jsx("p",{className:"mb-3",children:"Dane osobowe nie są przekazywane od krajów trzecich w rozumieniu przepisów o ochronie danych osobowych. Oznacza to, że nie przesyłamy ich poza teren Unii Europejskiej."})]}),e.jsxs("section",{children:[e.jsx("h2",{className:"text-2xl font-semibold text-js mb-4",children:"5. Informacje w formularzach"}),e.jsxs("ul",{className:"list-disc pl-6 space-y-2",children:[e.jsx("li",{children:"Serwis zbiera informacje podane dobrowolnie przez użytkownika, w tym dane osobowe, o ile zostaną one podane."}),e.jsx("li",{children:"Serwis może zapisać informacje o parametrach połączenia (oznaczenie czasu, adres IP)."}),e.jsx("li",{children:"Serwis, w niektórych wypadkach, może zapisać informację ułatwiającą powiązanie danych w formularzu z adresem e-mail użytkownika wypełniającego formularz. W takim wypadku adres e-mail użytkownika pojawia się wewnątrz adresu url strony zawierającej formularz."}),e.jsx("li",{children:"Dane podane w formularzu są przetwarzane w celu wynikającym z funkcji konkretnego formularza, np. w celu dokonania procesu obsługi zgłoszenia serwisowego lub kontaktu handlowego, rejestracji usług itp. Każdorazowo kontekst i opis formularza w czytelny sposób informuje, do czego on służy."})]})]}),e.jsxs("section",{children:[e.jsx("h2",{className:"text-2xl font-semibold text-js mb-4",children:"6. Logi Administratora"}),e.jsx("p",{className:"mb-3",children:"Informacje zachowaniu użytkowników w serwisie mogą podlegać logowaniu. Dane te są wykorzystywane w celu administrowania serwisem."})]}),e.jsxs("section",{children:[e.jsx("h2",{className:"text-2xl font-semibold text-js mb-4",children:"7. Istotne techniki marketingowe"}),e.jsxs("ul",{className:"list-disc pl-6 space-y-2",children:[e.jsx("li",{children:"Operator stosuje techniki remarketingowe, pozwalające na dopasowanie przekazów reklamowych do zachowania użytkownika na stronie, co może dawać złudzenie, że dane osobowe użytkownika są wykorzystywane do jego śledzenia, jednak w praktyce nie dochodzi do przekazania żadnych danych osobowych od Operatora do operatorom reklam. Technologicznym warunkiem takich działań jest włączona obsługa plików cookie."}),e.jsx("li",{children:"Operator stosuje rozwiązanie badające zachowanie użytkowników poprzez tworzenie map ciepła oraz nagrywanie zachowania na stronie. Te informacje są anonimizowane zanim zostaną przesłane do operatora usługi tak, że nie wie on jakiej osoby fizycznej one dotyczą. W szczególności nagrywaniu nie podlegają wpisywane hasła oraz inne dane osobowe."}),e.jsx("li",{children:"Operator stosuje rozwiązanie automatyzujące działanie Serwisu w odniesieniu do użytkowników, np. mogące przesłać maila do użytkownika po odwiedzeniu konkretnej podstrony, o ile wyraził on zgodę na otrzymywanie korespondencji handlowej od Operatora."})]})]}),e.jsxs("section",{children:[e.jsx("h2",{className:"text-2xl font-semibold text-js mb-4",children:"8. Informacja o plikach cookies"}),e.jsxs("ul",{className:"list-disc pl-6 space-y-2",children:[e.jsx("li",{children:"Serwis korzysta z plików cookies."}),e.jsx("li",{children:'Pliki cookies (tzw. "ciasteczka") stanowią dane informatyczne, w szczególności pliki tekstowe, które przechowywane są w urządzeniu końcowym Użytkownika Serwisu i przeznaczone są do korzystania ze stron internetowych Serwisu. Cookies zazwyczaj zawierają nazwę strony internetowej, z której pochodzą, czas przechowywania ich na urządzeniu końcowym oraz unikalny numer.'}),e.jsx("li",{children:"Podmiotem zamieszczającym na urządzeniu końcowym Użytkownika Serwisu pliki cookies oraz uzyskującym do nich dostęp jest operator Serwisu."}),e.jsxs("li",{children:["Pliki cookies wykorzystywane są w następujących celach:",e.jsxs("ul",{className:"list-disc pl-6 space-y-1 mt-2",children:[e.jsx("li",{children:"utrzymanie sesji użytkownika Serwisu (po zalogowaniu), dzięki której użytkownik nie musi na każdej podstronie Serwisu ponownie wpisywać loginu i hasła;"}),e.jsx("li",{children:'realizacji celów określonych powyżej w części "Istotne techniki marketingowe";'})]})]}),e.jsx("li",{children:'W ramach Serwisu stosowane są dwa zasadnicze rodzaje plików cookies: "sesyjne" (session cookies) oraz "stałe" (persistent cookies). Cookies "sesyjne" są plikami tymczasowymi, które przechowywane są w urządzeniu końcowym Użytkownika do czasu wylogowania, opuszczenia strony internetowej lub wyłączenia oprogramowania (przeglądarki internetowej). "Stałe" pliki cookies przechowywane są w urządzeniu końcowym Użytkownika przez czas określony w parametrach plików cookies lub do czasu ich usunięcia przez Użytkownika.'}),e.jsx("li",{children:"Oprogramowanie do przeglądania stron internetowych (przeglądarka internetowa) zazwyczaj domyślnie dopuszcza przechowywanie plików cookies w urządzeniu końcowym Użytkownika. Użytkownicy Serwisu mogą dokonać zmiany ustawień w tym zakresie. Przeglądarka internetowa umożliwia usunięcie plików cookies. Możliwe jest także automatyczne blokowanie plików cookies Szczegółowe informacje na ten temat zawiera pomoc lub dokumentacja przeglądarki internetowej."}),e.jsx("li",{children:"Ograniczenia stosowania plików cookies mogą wpłynąć na niektóre funkcjonalności dostępne na stronach internetowych Serwisu."}),e.jsx("li",{children:"Pliki cookies zamieszczane w urządzeniu końcowym Użytkownika Serwisu wykorzystywane mogą być również przez współpracujące z operatorem Serwisu podmioty, w szczególności dotyczy to firm: Google (Google Inc. z siedzibą w USA), Facebook (Facebook Inc. z siedzibą w USA), Twitter (Twitter Inc. z siedzibą w USA)."})]})]}),e.jsxs("section",{children:[e.jsx("h2",{className:"text-2xl font-semibold text-js mb-4",children:"9. Zarządzanie plikami cookies – jak w praktyce wyrażać i cofać zgodę?"}),e.jsx("p",{className:"mb-3",children:"Jeśli użytkownik nie chce otrzymywać plików cookies, może zmienić ustawienia przeglądarki. Zastrzegamy, że wyłączenie obsługi plików cookies niezbędnych dla procesów uwierzytelniania, bezpieczeństwa, utrzymania preferencji użytkownika może utrudnić, a w skrajnych przypadkach może uniemożliwić korzystanie ze stron www"}),e.jsx("p",{className:"mb-3",children:"W celu zarządzania ustawienia cookies wybierz z listy poniżej przeglądarkę internetową, której używasz i postępuj zgodnie z instrukcjami:"}),e.jsxs("ul",{className:"list-disc pl-6 space-y-1",children:[e.jsx("li",{children:e.jsx("a",{href:"#",className:"text-js hover:underline",children:"Edge"})}),e.jsx("li",{children:e.jsx("a",{href:"#",className:"text-js hover:underline",children:"Internet Explorer"})}),e.jsx("li",{children:e.jsx("a",{href:"#",className:"text-js hover:underline",children:"Chrome"})}),e.jsx("li",{children:e.jsx("a",{href:"#",className:"text-js hover:underline",children:"Safari"})}),e.jsx("li",{children:e.jsx("a",{href:"#",className:"text-js hover:underline",children:"Firefox"})}),e.jsx("li",{children:e.jsx("a",{href:"#",className:"text-js hover:underline",children:"Opera"})})]}),e.jsx("p",{className:"mt-3 mb-2",children:"Urządzenia mobilne:"}),e.jsxs("ul",{className:"list-disc pl-6 space-y-1",children:[e.jsx("li",{children:e.jsx("a",{href:"#",className:"text-js hover:underline",children:"Android"})}),e.jsx("li",{children:e.jsx("a",{href:"#",className:"text-js hover:underline",children:"Safari (iOS)"})}),e.jsx("li",{children:e.jsx("a",{href:"#",className:"text-js hover:underline",children:"Windows Phone"})})]})]}),e.jsx("div",{className:"pt-6 border-t border-js/10 text-sm text-gray-400",children:e.jsx("p",{children:"Ostatnia aktualizacja: 18.03.2025"})})]})]})]})})]});export{y as default};
