+++ server/public/index.html
@@ -9,17 +9,17 @@
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Space+Grotesk:wght@400;500;600;700&display=swap" rel="stylesheet">
     <link href="https://fonts.googleapis.com/css2?family=JetBrains+Mono:wght@400;500&display=swap" rel="stylesheet">
-    <script type="module" crossorigin src="/assets/index-DGy8klhn.js"></script>
+    <script type="module" crossorigin src="/assets/index-DTbAtaTr.js"></script>
     <link rel="modulepreload" crossorigin href="/assets/vendor-other-DT6kVT6s.js">
     <link rel="modulepreload" crossorigin href="/assets/vendor-monaco-C9aCaKYu.js">
     <link rel="modulepreload" crossorigin href="/assets/vendor-react-DZE_uCEP.js">
     <link rel="modulepreload" crossorigin href="/assets/vendor-framer-motion-D6A4MpK3.js">
-    <link rel="modulepreload" crossorigin href="/assets/app-services-MwNJNg2u.js">
-    <link rel="modulepreload" crossorigin href="/assets/app-auth-BOk5gCNS.js">
-    <link rel="modulepreload" crossorigin href="/assets/app-components-DCNmHqgN.js">
+    <link rel="modulepreload" crossorigin href="/assets/app-services-vJYhj3YV.js">
+    <link rel="modulepreload" crossorigin href="/assets/app-auth-C1CjYxGY.js">
+    <link rel="modulepreload" crossorigin href="/assets/app-components-L1wo8hYs.js">
     <link rel="modulepreload" crossorigin href="/assets/vendor-charts-BQN0K2I5.js">
-    <link rel="modulepreload" crossorigin href="/assets/app-dashboard-8I3Vw7Bc.js">
-    <link rel="stylesheet" crossorigin href="/assets/index-BW_MJYvl.css">
+    <link rel="modulepreload" crossorigin href="/assets/app-dashboard-D-SLglbn.js">
+    <link rel="stylesheet" crossorigin href="/assets/index-C0gYc-k0.css">
   </head>
   <body>
     <div id="root"></div>
